### 2025-07-01 #8529 @editorialbot
[PRE REVIEW]: patientflow: a Python package for real-time predictions of hospital bed demand from current and incoming patients
**Submitting author:** @zmek (<a href="http://orcid.org/0000-0001-7389-1527">Zella King</a>)
**Repository:** https://github.com/ucl-coru/patientflow
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
