### 2025-07-02 #8529 @zmek
Here are suggested reviewers, from the list [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review. I have suggested people with some background in working with Electronic Health Record Data or healthcare informatics

[HLasse](https://github.com/HLasse), Aarhus University
[KennethEnevoldsen](https://github.com/KennethEnevoldsen), Aarhus University
[limengbinggz](https://github.com/limengbinggz), University of Michigan
[hantswilliams](https://github.com/hantswilliams), Stony Brook University
[arianesasso](https://github.com/arianesasso), University of Manchester
