### 2025-07-03 93b0809
fix notebook errors after refactor
