+++ src/patientflow/calculate/admission_in_prediction_window.py
@@ -169,7 +169,8 @@ def calculate_probability(
     y2: float,
 ):
     """
-    Calculates the probability of an admission occurring within a specified prediction window after the moment of prediction, based on the patient's elapsed time in the ED prior to the moment of prediction and the length of the window
+    Calculates the probability of an admission occurring within a specified prediction window after the moment of prediction, 
+    based on the patient's elapsed time in the ED prior to the moment of prediction and the length of the window
 
     Parameters
     ----------
@@ -291,7 +292,9 @@ def get_survival_probability(survival_df, time_hours):
       interpolate between the nearest time points
     - If time_hours is less than the minimum time in the data, returns 1.0
     - If time_hours is greater than the maximum time in the data, returns the last
-      known survival probability
+      known survival probability and issues a warning. This behavior may need to be
+      reconsidered depending on the specific use case and whether the survival curve
+      has reached its asymptotic value.
 
     Examples
     --------
@@ -306,6 +309,19 @@ def get_survival_probability(survival_df, time_hours):
         return 1.0
 
     if time_hours > survival_df["time_hours"].max():
+        # For times beyond the data, we have a few options:
+        # 1. Return the last known value (current behavior)
+        # 2. Extrapolate using the last few points
+        # 3. Return 0 (assuming all patients are eventually admitted/discharged)
+        # 4. Use exponential decay extrapolation
+        
+        # For now, we'll use the last known value, but warn the user
+        warnings.warn(
+            f"Time {time_hours} hours exceeds maximum time in survival data "
+            f"({survival_df['time_hours'].max()} hours). Using last known survival probability "
+            f"({survival_df['survival_probability'].iloc[-1]:.3f}). Consider extending the survival data "
+            "or using a different extrapolation method."
+        )
         return survival_df["survival_probability"].iloc[-1]
 
     # Find the closest time points for interpolation
@@ -326,3 +342,122 @@ def get_survival_probability(survival_df, time_hours):
 
     # Linear interpolation
     return p1 + (p2 - p1) * (time_hours - t1) / (t2 - t1)
+
+
+def calculate_admission_probability_from_survival_curve(
+    elapsed_los: timedelta,
+    prediction_window: timedelta,
+    survival_df,
+):
+    """
+    Calculates the probability of an admission occurring within a specified prediction window after the moment of prediction, 
+    based on the patient's elapsed time in the ED prior to the moment of prediction and the length of the window.
+    Uses empirical survival curve data instead of an aspirational curve.
+
+    Parameters
+    ----------
+    elapsed_los : timedelta
+        The elapsed time since the patient arrived at the ED.
+    prediction_window : timedelta
+        The duration of the prediction window after the point of prediction, for which the probability is calculated.
+    survival_df : pandas.DataFrame
+        DataFrame containing survival curve data with columns:
+        - time_hours: Time points in hours
+        - survival_probability: Probability of still being in ED at each time point
+
+    Returns
+    -------
+    float
+        The probability of admission occurring within the given prediction window.
+
+    Notes
+    -----
+    This function calculates the conditional probability of admission within the prediction window,
+    given that the patient hasn't been admitted yet. It uses the formula:
+    
+    P(admission in window | not admitted yet) = 
+        (P(not admitted at elapsed) - P(not admitted at elapsed + window)) / P(not admitted at elapsed)
+    
+    Since survival probability = P(not admitted), this becomes:
+    
+    P(admission in window | not admitted yet) = 
[…]

[…]
