+++ src/patientflow/calculate/admission_in_prediction_window.py
@@ -169,7 +169,7 @@ def calculate_probability(
     y2: float,
 ):
     """
-    Calculates the probability of an admission occurring within a specified prediction window after the moment of prediction, 
+    Calculates the probability of an admission occurring within a specified prediction window after the moment of prediction,
     based on the patient's elapsed time in the ED prior to the moment of prediction and the length of the window
 
     Parameters
@@ -291,10 +291,8 @@ def get_survival_probability(survival_df, time_hours):
     - If the exact time_hours is not in the survival curve data, the function will
       interpolate between the nearest time points
     - If time_hours is less than the minimum time in the data, returns 1.0
-    - If time_hours is greater than the maximum time in the data, returns the last
-      known survival probability and issues a warning. This behavior may need to be
-      reconsidered depending on the specific use case and whether the survival curve
-      has reached its asymptotic value.
+    - If time_hours is greater than the maximum time in the data, uses exponential
+      decay extrapolation based on the last few data points
 
     Examples
     --------
@@ -309,20 +307,59 @@ def get_survival_probability(survival_df, time_hours):
         return 1.0
 
     if time_hours > survival_df["time_hours"].max():
-        # For times beyond the data, we have a few options:
-        # 1. Return the last known value (current behavior)
-        # 2. Extrapolate using the last few points
-        # 3. Return 0 (assuming all patients are eventually admitted/discharged)
-        # 4. Use exponential decay extrapolation
-        
-        # For now, we'll use the last known value, but warn the user
-        warnings.warn(
-            f"Time {time_hours} hours exceeds maximum time in survival data "
-            f"({survival_df['time_hours'].max()} hours). Using last known survival probability "
-            f"({survival_df['survival_probability'].iloc[-1]:.3f}). Consider extending the survival data "
-            "or using a different extrapolation method."
-        )
-        return survival_df["survival_probability"].iloc[-1]
+        # Use exponential decay extrapolation for times beyond the data
+        max_time = survival_df["time_hours"].max()
+        max_prob = survival_df["survival_probability"].iloc[-1]
+
+        # Track whether we attempted exponential fitting and whether it failed
+        attempted_exponential_fitting = False
+        exponential_fitting_failed = False
+
+        # Use the last 3 points to fit exponential decay if available
+        if len(survival_df) >= 3:
+            attempted_exponential_fitting = True
+
+            # Get the last 3 points for fitting
+            last_times = survival_df["time_hours"].iloc[-3:].values
+            last_probs = survival_df["survival_probability"].iloc[-3:].values
+
+            # Fit exponential decay: P(t) = P0 * exp(-lambda * (t - t0))
+            # where P0 is the probability at t0, and lambda is the decay rate
+            t0 = last_times[0]
+            P0 = last_probs[0]
+
+            # Calculate decay rate lambda using the last two points
+            t1, t2 = last_times[-2], last_times[-1]
+            P1, P2 = last_probs[-2], last_probs[-1]
+
+            # For exponential fitting, replace P2 = 0 with a small positive value
+            # to ensure the decay can be calculated while maintaining asymptotic behavior
+            if P2 == 0:
+                P2_fit = 1e-6  # Small positive value for fitting
+            else:
+                P2_fit = P2
+
+            # lambda = -ln(P2/P1) / (t2 - t1)
+            if P2_fit > 0 and P1 > 0 and t2 > t1:
+                lambda_decay = -np.log(P2_fit / P1) / (t2 - t1)
+
+                # Extrapolate using exponential decay
+                extrapolated_prob = P0 * np.exp(-lambda_decay * (time_hours - t0))
+
+                # Ensure the probability doesn't go below 0
+                return max(0.0, extrapolated_prob)
+            else:
[…]
