+++ tests/test_incoming_admission_predictors.py
@@ -7,6 +7,7 @@ from datetime import datetime, timedelta
 from patientflow.predictors.incoming_admission_predictors import (
     ParametricIncomingAdmissionPredictor,
     EmpiricalIncomingAdmissionPredictor,
+    DirectAdmissionPredictor,
     weighted_poisson_binomial,
     aggregate_probabilities,
     convolute_distributions,
@@ -582,6 +583,309 @@ class TestIncomingAdmissionPredictors(unittest.TestCase):
         # Predictions should still be generated successfully
         self.assertIn("medical", empirical_predictions)
         self.assertIn("surgical", empirical_predictions)
+
+    def test_direct_predictor_initialization(self):
+        """Test DirectAdmissionPredictor initialization."""
+        predictor = DirectAdmissionPredictor(filters=self.filters)
+        self.assertEqual(predictor.filters, self.filters)
+        self.assertFalse(predictor.verbose)
+
+        # Test with verbose mode
+        verbose_predictor = DirectAdmissionPredictor(filters=self.filters, verbose=True)
+        self.assertTrue(verbose_predictor.verbose)
+
+        # Test with no filters
+        no_filter_predictor = DirectAdmissionPredictor(filters=None)
+        self.assertEqual(no_filter_predictor.filters, {})
+
+    def test_direct_predictor_fit(self):
+        """Test DirectAdmissionPredictor fit method."""
+        predictor = DirectAdmissionPredictor(filters=self.filters)
+
+        # Test successful fit
+        predictor.fit(
+            self.test_df,
+            self.prediction_window,
+            self.yta_time_interval,
+            self.prediction_times,
+            self.num_days,
+        )
+
+        # Check that weights were calculated
+        self.assertIsInstance(predictor.weights, dict)
+        self.assertTrue(len(predictor.weights) > 0)
+
+        # Check that all filter keys are present
+        for key in self.filters.keys():
+            self.assertIn(key, predictor.weights)
+
+        # Check that prediction times are stored correctly
+        self.assertEqual(predictor.prediction_times, self.prediction_times)
+
+        # Check that arrival_rates are stored for each filter and prediction time
+        for filter_key in self.filters.keys():
+            self.assertIn(filter_key, predictor.weights)
+            for prediction_time in self.prediction_times:
+                self.assertIn(prediction_time, predictor.weights[filter_key])
+                self.assertIn("arrival_rates", predictor.weights[filter_key][prediction_time])
+                arrival_rates = predictor.weights[filter_key][prediction_time]["arrival_rates"]
+                self.assertIsInstance(arrival_rates, list)
+                self.assertTrue(len(arrival_rates) > 0)
+
+    def test_direct_predictor_predict(self):
+        """Test DirectAdmissionPredictor predict method."""
+        predictor = DirectAdmissionPredictor(filters=self.filters)
+        predictor.fit(
+            self.test_df,
+            self.prediction_window,
+            self.yta_time_interval,
+            self.prediction_times,
+            self.num_days,
+        )
+
+        # Test prediction
+        prediction_context = {
+            "medical": {"prediction_time": (8, 0)},
+            "surgical": {"prediction_time": (12, 0)},
+        }
+
+        predictions = predictor.predict(prediction_context, max_value=30)
+
+        # Check output format
+        self.assertIsInstance(predictions, dict)
+        self.assertIn("medical", predictions)
+        self.assertIn("surgical", predictions)
+
+        # Check that predictions are DataFrames with expected columns
+        for key, pred_df in predictions.items():
+            self.assertIsInstance(pred_df, pd.DataFrame)
+            self.assertIn("agg_proba", pred_df.columns)
+            self.assertTrue(np.allclose(pred_df["agg_proba"].sum(), 1.0, atol=1e-10))
+            
[…]
