### 2025-09-24 88d4705
handle database fragmentation in spec model
