+++ src/patientflow/predictors/sequence_to_outcome_predictor.py
@@ -223,6 +223,7 @@ class SequenceToOutcomePredictor(BaseEstimator, TransformerMixin):
 
         # Reweight probabilities of ending with each observed specialty
         # by the likelihood of each grouping sequence occurring
+        # This transforms conditional probabilities into overall probabilities
         reweighted_proportions = proportions.copy()
         for col in proportions.columns:  
             reweighted_proportions[col] *= probability_of_grouping_sequence
@@ -235,8 +236,9 @@ class SequenceToOutcomePredictor(BaseEstimator, TransformerMixin):
         # Row-wise function to return, for each input sequence,
         # the proportion that end up in each final sequence and thereby
         # the probability of it ending in any observed category
-        prob_input_var_ends_in_observed_specialty = pd.Series(grouping_sequence_to_string).apply(
-            lambda x: self._string_match_input_var(x, reweighted_proportions, prop_keys)
+        grouping_sequence_series = pd.Series(grouping_sequence_to_string, index=proportions.index)
+        prob_input_var_ends_in_observed_specialty = grouping_sequence_series.apply(
+            lambda x: self._string_match_input_var(x, reweighted_proportions, prop_keys, grouping_sequence_to_string)
         )
 
         # Combine all new columns at once to avoid DataFrame fragmentation
@@ -274,7 +276,7 @@ class SequenceToOutcomePredictor(BaseEstimator, TransformerMixin):
 
         return self
 
-    def _string_match_input_var(self, input_var_string, proportions, prop_keys):
+    def _string_match_input_var(self, input_var_string, proportions, prop_keys, grouping_sequence_to_string):
         """
         Matches a given input sequence string with grouped sequences (expressed as strings) in the dataset and aggregates
         their probabilities for each outcome category. This function filters the data to
@@ -289,10 +291,12 @@ class SequenceToOutcomePredictor(BaseEstimator, TransformerMixin):
         input_var_string : str
             The sequence of inputs represented as a string, used to match against sequences in the proportions DataFrame.
         proportions : pd.DataFrame
-            DataFrame containing proportions data with an additional column 'grouping_sequence_to_string'
-            which includes string representations of sequences.
+            DataFrame containing reweighted proportions data for each outcome category. Must have the same index as grouping_sequence_to_string.
         prop_keys : np.array
-            Array of unique outcomes to consider in calculations.
+            Array of unique outcome category names to consider in calculations.
+        grouping_sequence_to_string : pd.Index
+            Index containing string representations of the grouping sequences, with the same index as proportions.
+            Used to create boolean masks for filtering proportions rows.
 
         Returns
         -------
@@ -302,9 +306,8 @@ class SequenceToOutcomePredictor(BaseEstimator, TransformerMixin):
 
         """
         # Filter rows where the grouped sequence string starts with the input sequence string
-        props = proportions[
-            proportions["grouping_sequence_to_string"].str.match("^" + input_var_string)
-        ][prop_keys].sum()
+        mask = grouping_sequence_to_string.str.match("^" + input_var_string)
+        props = proportions[mask][prop_keys].sum()
 
         # Sum of all probabilities to normalize them
         props_total = props.sum()
