### 2025-09-25 c728d7f
implement generating function in aggregate.py
