### 2025-09-25 2c98128
Create system for handling subgroups
- flag create_special_category_objects as legacy
- use create_subgroup_functions() handler for identifying subgroups
- create MultiSubgroupPredictor to manage multiple SequenceToOutcomePredictor models, one per subgroup
