+++ src/patientflow/group.py
@@ -1,0 +1,281 @@
+"""
+Simple subgroup system for managing multiple patient prediction models.
+
+This module provides an approach to training and using separate
+SequenceToOutcomePredictor models for different patient subgroups.
+
+Functions
+---------
+create_subgroup_system
+    Main entry point - creates subgroup functions and multi-model predictor
+create_special_category_objects  
+    Backward compatibility function (deprecated)
+"""
+
+import warnings
+from typing import Dict, Any, Union, List, Type, Callable, Optional, Tuple
+import pandas as pd
+from patientflow.predictors.sequence_to_outcome_predictor import SequenceToOutcomePredictor
+
+
+def get_age(row: Union[pd.Series, dict]) -> int:
+    """Get patient age, defaulting to representative values for age groups."""
+    if 'age_on_arrival' in row and pd.notna(row['age_on_arrival']):
+        return int(row['age_on_arrival'])
+    
+    age_group = str(row.get('age_group', '')).lower()
+    if '0-17' in age_group:
+        return 10
+    elif '65' in age_group:
+        return 70
+    else:
+        return 40
+
+
+def create_subgroup_functions() -> Dict[str, Callable[[Union[pd.Series, dict]], bool]]:
+    """Create the 5 standard subgroup identification functions."""
+    
+    def is_pediatric(row):
+        return get_age(row) < 18
+    
+    def is_adult_male_young(row):
+        age = get_age(row)
+        return 18 <= age < 65 and row.get('sex') == 'M'
+    
+    def is_adult_female_young(row):
+        age = get_age(row)
+        return 18 <= age < 65 and row.get('sex') == 'F'
+    
+    def is_adult_male_senior(row):
+        age = get_age(row)
+        return age >= 65 and row.get('sex') == 'M'
+    
+    def is_adult_female_senior(row):
+        age = get_age(row)
+        return age >= 65 and row.get('sex') == 'F'
+    
+    return {
+        'pediatric': is_pediatric,
+        'adult_male_young': is_adult_male_young,
+        'adult_female_young': is_adult_female_young,
+        'adult_male_senior': is_adult_male_senior,
+        'adult_female_senior': is_adult_female_senior
+    }
+
+
+class MultiSubgroupPredictor:
+    """Manages multiple SequenceToOutcomePredictor models, one per subgroup.
+    
+    Parameters
+    ----------
+    subgroup_functions : Dict[str, Callable]
+        Dictionary mapping subgroup names to functions that identify patients in each subgroup
+    base_predictor_class : Type[SequenceToOutcomePredictor], optional
+        The SequenceToOutcomePredictor class to instantiate for each subgroup
+    input_var : str, default='consultation_sequence'
+        Name of the input sequence column
+    grouping_var : str, default='final_sequence'
+        Name of the grouping sequence column  
+    outcome_var : str, default='observed_specialty'
+        Name of the outcome variable column
+    min_samples : int, default=50
+        Minimum number of samples required to train a model for a subgroup
+    """
+    
+    def __init__(self, 
+                 subgroup_functions: Dict[str, Callable[[Union[pd.Series, dict]], bool]], 
+                 base_predictor_class: Optional[Type[SequenceToOutcomePredictor]], 
+                 input_var: str = 'consultation_sequence',
+                 grouping_var: str = 'final_sequence',
+                 outcome_var: str = 'observed_specialty',
+                 min_samples: int = 50):
+        self.subgroup_functions: Dict[str, Callable[[Union[pd.Series, dict]], bool]] = subgroup_functions
+        self.base_predictor_class: Optional[Type[SequenceToOutcomePredictor]] = base_predictor_class
+        self.input_var: str = input_var
[…]
