### 2025-09-25 0eec805
add note about legacy code
