+++ src/patientflow/predictors/subgroup_predictor.py
@@ -16,20 +16,7 @@ import warnings
 from typing import Dict, Any, Union, List, Type, Callable, Optional, Tuple
 import pandas as pd
 from patientflow.predictors.sequence_to_outcome_predictor import SequenceToOutcomePredictor
-
-
-def get_age(row: Union[pd.Series, dict]) -> int:
-    """Get patient age, defaulting to representative values for age groups."""
-    if 'age_on_arrival' in row and pd.notna(row['age_on_arrival']):
-        return int(row['age_on_arrival'])
-    
-    age_group = str(row.get('age_group', '')).lower()
-    if '0-17' in age_group:
-        return 10
-    elif '65' in age_group:
-        return 70
-    else:
-        return 40
+from patientflow.predictors.legacy_compatibility import get_age, create_special_category_objects, create_yta_filters
 
 
 def create_subgroup_functions() -> Dict[str, Callable[[Union[pd.Series, dict]], bool]]:
@@ -193,7 +180,7 @@ def create_subgroup_system(columns: Union[List[str], pd.Index],
     """
     subgroup_functions = create_subgroup_functions()
     
-    if                    :
+    if base_predictor_class is not None:
         predictor = MultiSubgroupPredictor(
             subgroup_functions, 
             base_predictor_class,
@@ -230,52 +217,3 @@ def create_subgroup_system(columns: Union[List[str], pd.Index],
         'special_func_map': legacy_params['special_func_map']
     }
 
-
-def create_special_category_objects(columns: Union[List[str], pd.Index]) -> Dict[str, Any]:
-    """Legacy function - returns ONLY original pediatric/adult logic."""
-    warnings.warn(
-        "create_special_category_objects() is deprecated. Use create_subgroup_system() instead.",
-        DeprecationWarning
-    )
-    
-    def is_pediatric(row):
-        age = get_age(row)
-        return age < 18
-    
-    def is_not_pediatric(row):
-        age = get_age(row) 
-        return age >= 18
-    
-    # Return ONLY the original legacy structure
-    return {
-        'special_category_func': is_pediatric,
-        'special_category_dict': {
-            'medical': 0.0, 
-            'surgical': 0.0, 
-            'haem/onc': 0.0, 
-            'paediatric': 1.0
-        },
-        'special_func_map': {
-            'paediatric': is_pediatric,
-            'default': is_not_pediatric
-            # NO adult subgroup functions here!
-        }
-    }
-    
-def create_yta_filters(df: pd.DataFrame) -> Dict[str, Dict[str, Any]]:
-    """Create filters for yet-to-arrive predictions (backward compatibility).
-
-    Includes all specialties present in the legacy special category dict, e.g.
-    'medical', 'surgical', 'haem/onc', and 'paediatric'.
-    """
-    special_params = create_special_category_objects(df.columns)
-    special_category_dict = special_params["special_category_dict"]
-
-    filters: Dict[str, Dict[str, Any]] = {}
-    for specialty, is_paediatric_flag in special_category_dict.items():
-        if is_paediatric_flag == 1.0:
-            filters[specialty] = {"is_child": True}
-        else:
-            filters[specialty] = {"specialty": specialty, "is_child": False}
-
-    return filters
