### 2025-09-29 d9b5fd7
add predict method to MultiSubgroupPredictor
