### 2025-09-29 1020297
save mapping from subgroup to subspecialty with MultiSubgroupPredictor
