### 2025-09-30 93d07f9
modify create_predictions for MultiSubgroupPredictor
