+++ src/patientflow/predict/emergency_demand.py
@@ -22,7 +22,7 @@ create_predictions : function
 
 """
 
-from typing import List, Dict, Tuple, Union
+from typing import List, Dict, Tuple, Union, Optional, Any
 from datetime import timedelta
 import pandas as pd
 from sklearn.preprocessing import OneHotEncoder, OrdinalEncoder, StandardScaler
@@ -45,6 +45,7 @@ from patientflow.predictors.sequence_to_outcome_predictor import (
     SequenceToOutcomePredictor,
 )
 from patientflow.predictors.value_to_outcome_predictor import ValueToOutcomePredictor
+from patientflow.predictors.subgroup_predictor import MultiSubgroupPredictor
 from patientflow.predictors.incoming_admission_predictors import (
     ParametricIncomingAdmissionPredictor,
     EmpiricalIncomingAdmissionPredictor,
@@ -234,7 +235,9 @@ def get_specialty_probs(
 def create_predictions(
     models: Tuple[
         TrainedClassifier,
-        Union[SequenceToOutcomePredictor, ValueToOutcomePredictor],
+        Union[
+            SequenceToOutcomePredictor, ValueToOutcomePredictor, MultiSubgroupPredictor
+        ],
         Union[
             ParametricIncomingAdmissionPredictor, EmpiricalIncomingAdmissionPredictor
         ],
@@ -247,17 +250,17 @@ def create_predictions(
     y1: float,
     x2: float,
     y2: float,
-    cdf_cut_points: List[float],
+    cdf_cut_points: Optional[List[float]] = None,
     use_admission_in_window_prob: bool = True,
-) -> Dict[str, Dict[str, List[int]]]:
+) -> Dict[str, Dict[str, Any]]:
     """Create predictions for emergency demand for a single prediction moment.
 
     Parameters
     ----------
-    models : Tuple[TrainedClassifier, Union[SequenceToOutcomePredictor, ValueToOutcomePredictor], Union[ParametricIncomingAdmissionPredictor, EmpiricalIncomingAdmissionPredictor]]
+    models : Tuple[TrainedClassifier, Union[SequenceToOutcomePredictor, ValueToOutcomePredictor, MultiSubgroupPredictor], Union[ParametricIncomingAdmissionPredictor, EmpiricalIncomingAdmissionPredictor]]
         Tuple containing:
         - classifier: TrainedClassifier containing admission predictions
-        - spec_model: SequenceToOutcomePredictor or ValueToOutcomePredictor for specialty predictions
+        - spec_model: SequenceToOutcomePredictor or ValueToOutcomePredictor or MultiSubgroupPredictor for specialty predictions
         - yet_to_arrive_model: ParametricIncomingAdmissionPredictor or EmpiricalIncomingAdmissionPredictor for yet-to-arrive predictions
     prediction_time : Tuple
         Hour and minute of time for model inference
@@ -275,8 +278,10 @@ def create_predictions(
         X-coordinate of second point for probability curve
     y2 : float
         Y-coordinate of second point for probability curve
-    cdf_cut_points : List[float]
-        List of cumulative distribution function cut points (e.g., [0.9, 0.7])
+    cdf_cut_points : Optional[List[float]], optional
+        If provided, return per-specialty lists of threshold indices for each cut point
+        (e.g., [0.9, 0.7]). If omitted/None, return the raw PMFs (agg_proba) for
+        both in-ED and yet-to-arrive predictions for backward compatibility.
     use_admission_in_window_prob : bool, optional
         Whether to use probability calculation for admission within prediction window for patients
         already in the ED. If False, probability is set to 1.0 for all current ED patients.
@@ -284,14 +289,21 @@ def create_predictions(
 
     Returns
     -------
[…]
