### 2025-09-30 2490d61
add generating functions to incoming admission classes
