+++ src/patientflow/predictors/incoming_admission_predictors.py
@@ -9,21 +9,28 @@ accommodate different data filters for tailored predictions across various hospi
 
 Classes
 -------
+AdmissionGeneratingFunction
+    Unified generating function class for computation of admission prediction
+    distributions using probability generating functions.
+
 IncomingAdmissionPredictor : BaseEstimator, TransformerMixin
     Base class for admission predictors that handles filtering and arrival rate calculation.
+    Now supports both generating function and traditional approaches via use_generating_functions flag.
 
 DirectAdmissionPredictor : IncomingAdmissionPredictor
     Simplest predictor that assumes every arrival is admitted immediately. Uses direct
     Poisson distribution based on total arrival rates without any admission probability
-    adjustments.
+    adjustments. Now supports both generating function and traditional approaches.
 
 ParametricIncomingAdmissionPredictor : IncomingAdmissionPredictor
     Predicts the number of admissions within a given prediction window based on historical
     data and Poisson-binomial distribution using parametric aspirational curves.
+    Now uses efficient generating functions for better performance.
 
 EmpiricalIncomingAdmissionPredictor : IncomingAdmissionPredictor
     Predicts the number of admissions using empirical survival curves and convolution
-    of Poisson distributions instead of parametric curves.
+    of Poisson distributions instead of parametric curves. Now supports both generating
+    function and traditional approaches.
 
 Notes
 -----
@@ -285,6 +292,121 @@ def find_nearest_previous_prediction_time(requested_time, prediction_times):
     return closest_prediction_time
 
 
+class AdmissionGeneratingFunction:
+    """Unified generating function for admission prediction distributions.
+    
+    This class provides efficient computation of probability distributions for
+    different admission prediction approaches using simple convolution methods.
+    
+    For typical use cases (32 intervals, max ~50), simple convolution is more
+    appropriate than FFT-based methods which add unnecessary complexity.
+    
+    Parameters
+    ----------
+    arrival_rates : numpy.ndarray
+        Array of arrival rates for each time interval
+    admission_probs : numpy.ndarray
+        Array of admission probabilities for each time interval
+    method : str, default='empirical'
+        Method to use: 'direct', 'parametric', or 'empirical'
+    
+    Attributes
+    ----------
+    arrival_rates : numpy.ndarray
+        Arrival rates for each time interval
+    admission_probs : numpy.ndarray
+        Admission probabilities for each time interval
+    method : str
+        The prediction method being used
+    """
+    
+    def __init__(self, arrival_rates, admission_probs, method='empirical'):
+        self.arrival_rates = np.array(arrival_rates)
+        self.admission_probs = np.array(admission_probs)
+        self.method = method
+        
+        if len(self.arrival_rates) != len(self.admission_probs):
+            raise ValueError("arrival_rates and admission_probs must have the same length")
+    
+    def get_distribution(self, max_value=50):
+        """Get the probability distribution for the specified method.
+        
+        Parameters
+        ----------
+        max_value : int, default=50
+            Maximum value for the discrete distribution support
+            
+        Returns
+        -------
+        pandas.DataFrame
+            DataFrame with 'sum' and 'agg_proba' columns representing the distribution
+        """
+        if self.method == 'direct':
+            return self._direct_distribution(max_value)
+        elif self.method == 'empirical':
+            return self._empirical_distribution(max_value)
+        elif self.method == 'parametric':
+            return self._parametric_distribution(max_value)
+        else:
+            raise ValueError(f"Unknown method: {self.method}")
+    
+    def _direct_distribution(self, max_value=50):
[…]
