### 2025-09-30 39b8bbf
fix init method for EmpiricalIncomingAdmissionPredictor
