### 2025-09-30 96104a1
reduce duplication in incoming_admission_predictors.py
