+++ src/patientflow/predictors/incoming_admission_predictors.py
@@ -483,6 +483,85 @@ class IncomingAdmissionPredictor(BaseEstimator, TransformerMixin, ABC):
         # Apply filters
         self.filters = filters if filters else {}
 
+    # -------------------------
+    # Shared helper utilities
+    # -------------------------
+    def _validate_only_kwargs(self, kwargs: Dict, allowed: set, context: str) -> None:
+        unexpected = set(kwargs.keys()) - allowed
+        if unexpected:
+            raise ValueError(
+                f"{context} only accepts these parameters: {allowed}. Remove these unexpected parameters: {unexpected}"
+            )
+
+    def _ensure_required_kwargs(self, kwargs: Dict, required: set, context: str) -> Dict:
+        missing = [k for k in required if kwargs.get(k) is None]
+        if missing:
+            raise ValueError(
+                f"{context} requires these parameters: {missing}."
+            )
+        return {k: kwargs[k] for k in required}
+
+    def _iter_prediction_inputs(self, prediction_context: Dict):
+        """Yield (filter_key, resolved_prediction_time, arrival_rates_np).
+
+        Centralizes validation, time resolution, and array conversion.
+        """
+        for filter_key, filter_values in prediction_context.items():
+            if filter_key not in self.weights:
+                raise ValueError(
+                    f"Filter key '{filter_key}' is not recognized in the model weights."
+                )
+
+            prediction_time = filter_values.get("prediction_time")
+            if prediction_time is None:
+                raise ValueError(
+                    f"No 'prediction_time' provided for filter '{filter_key}'."
+                )
+
+            if prediction_time not in self.prediction_times:
+                prediction_time = find_nearest_previous_prediction_time(
+                    prediction_time, self.prediction_times
+                )
+
+            arrival_rates = self.weights[filter_key][prediction_time].get(
+                "arrival_rates"
+            )
+            if arrival_rates is None:
+                raise ValueError(
+                    f"No arrival_rates found for the time of day '{prediction_time}' under filter '{filter_key}'."
+                )
+
+            yield filter_key, prediction_time, np.array(arrival_rates)
+
+    def _get_window_and_interval_hours(self):
+        """Return (prediction_window_hours, interval_hours, NTimes)."""
+        return (
+            self.prediction_window_hours,
+            self.yta_time_interval_hours,
+            self.NTimes,
+        )
+
+    def _default_max_value(self, arrival_rates, floor: int = 20, cap: int = 200) -> int:
+        """Unified default max_value based on epsilon and max arrival rate.
+
+        Applies a floor and cap to ensure stable PMF support.
+        """
+        try:
+            mv = int(poisson.ppf(1 - self.epsilon, float(np.max(arrival_rates))))
+        except Exception:
+            mv = floor
+        mv = max(floor, mv)
+        mv = min(cap, mv)
+        return mv
+
+    def _pmf_to_dataframe(self, probabilities: np.ndarray) -> pd.DataFrame:
+        """Convert PMF array to standardized DataFrame indexed by 'sum' with 'agg_proba'."""
+        result_df = pd.DataFrame({"sum": range(len(probabilities)), "agg_proba": probabilities})
+        result_df = result_df[result_df["agg_proba"] > 1e-10]
+        if len(result_df) > 0:
+            result_df["agg_proba"] = result_df["agg_proba"] / result_df["agg_proba"].sum()
+        return result_df.set_index("sum")
+
     def filter_dataframe(self, df: pd.DataFrame, filters: Dict) -> pd.DataFrame:
         """Apply a set of filters to a dataframe.
 
@@ -687,6 +766,11 @@ class IncomingAdmissionPredictor(BaseEstimator, TransformerMixin, ABC):
[…]
