### 2025-09-30 4f25878
remove legacy code from incoming_admission_predictors.py
