+++ src/patientflow/predictors/incoming_admission_predictors.py
@@ -15,12 +15,12 @@ AdmissionGeneratingFunction
 
 IncomingAdmissionPredictor : BaseEstimator, TransformerMixin
     Base class for admission predictors that handles filtering and arrival rate calculation.
-    Now supports both generating function and traditional approaches via use_generating_functions flag.
+    Uses generating functions for all predictors; the earlier flag is deprecated and ignored.
 
 DirectAdmissionPredictor : IncomingAdmissionPredictor
     Simplest predictor that assumes every arrival is admitted immediately. Uses direct
     Poisson distribution based on total arrival rates without any admission probability
-    adjustments. Now supports both generating function and traditional approaches.
+    adjustments. Implemented via generating functions.
 
 ParametricIncomingAdmissionPredictor : IncomingAdmissionPredictor
     Predicts the number of admissions within a given prediction window based on historical
@@ -29,8 +29,7 @@ ParametricIncomingAdmissionPredictor : IncomingAdmissionPredictor
 
 EmpiricalIncomingAdmissionPredictor : IncomingAdmissionPredictor
     Predicts the number of admissions using empirical survival curves and convolution
-    of Poisson distributions instead of parametric curves. Now supports both generating
-    function and traditional approaches.
+    of Poisson distributions; implemented via generating functions.
 
 Notes
 -----
@@ -451,7 +450,15 @@ class IncomingAdmissionPredictor(BaseEstimator, TransformerMixin, ABC):
         """
         self.filters = filters if filters else {}
         self.verbose = verbose
-        self.use_generating_functions = use_generating_functions
+        # Always use generating-function path; keep flag only for backward compatibility
+        # If user explicitly sets False, warn and proceed with GF implementation
+        self.use_generating_functions = True
+        if use_generating_functions is False:
+            warnings.warn(
+                "use_generating_functions=False is deprecated and ignored; generating-function implementation is always used.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
         self.metrics = {}  # Add metrics dictionary to store metadata
 
         if verbose:
@@ -874,24 +881,15 @@ class DirectAdmissionPredictor(IncomingAdmissionPredictor):
 
         for filter_key, prediction_time, arrival_rates in self._iter_prediction_inputs(prediction_context):
 
-                if self.use_generating_functions:
-                    # Use the new generating function approach for consistency
-                    # For direct case, admission probabilities are all 1.0 (100% admission rate)
-                    admission_probs = np.ones(len(arrival_rates))
-                    
-                    gf = AdmissionGeneratingFunction(
-                        arrival_rates, admission_probs, method='direct'
-                    )
-                    mv = max_value if max_value is not None else self._default_max_value(arrival_rates)
-                    predictions[filter_key] = gf.get_distribution(max_value=mv)
-                else:
-                    # Use the original approach for backward compatibility
-                    total_arrival_rate = arrival_rates.sum()
-                    poisson_dist = stats.poisson(total_arrival_rate)
-                    mv = max_value if max_value is not None else self._default_max_value(arrival_rates)
-                    x_values = np.arange(mv)
-                    probabilities = poisson_dist.pmf(x_values)
-                    predictions[filter_key] = self._pmf_to_dataframe(probabilities)
+                # Generating-function approach (always)
+                # For direct case, admission probabilities are all 1.0 (100% admission rate)
+                admission_probs = np.ones(len(arrival_rates))
+                
+                gf = AdmissionGeneratingFunction(
+                    arrival_rates, admission_probs, method='direct'
+                )
+                mv = max_value if max_value is not None else self._default_max_value(arrival_rates)
[…]
