### 2025-09-30 c4e55a9
replace poisson-bin omial logic with exact Poisson, thinned
