+++ src/patientflow/predictors/value_to_outcome_predictor.py
@@ -67,7 +67,6 @@ class ValueToOutcomePredictor(BaseEstimator, TransformerMixin):
         self.metrics = {}
 
     def __repr__(self):
-        """Return a string representation of the estimator."""
         class_name = self.__class__.__name__
         return (
             f"{class_name}(\n"
