+++ src/patientflow/predictors/incoming_admission_predictors.py
@@ -31,6 +31,11 @@ EmpiricalIncomingAdmissionPredictor : IncomingAdmissionPredictor
     Predicts the number of admissions using empirical survival curves and convolution
     of Poisson distributions; implemented via generating functions.
 
+Functions
+---------
+find_nearest_previous_prediction_time
+    Find the nearest previous prediction time for a requested time
+
 Notes
 -----
 The DirectAdmissionPredictor is the simplest approach, summing all arrival rates across time
@@ -311,13 +316,13 @@ def find_nearest_previous_prediction_time(requested_time, prediction_times):
 
 class AdmissionGeneratingFunction:
     """Unified generating function for admission prediction distributions.
-    
+
     This class provides efficient computation of probability distributions for
     different admission prediction approaches using simple convolution methods.
-    
+
     For typical use cases (32 intervals, max ~50), simple convolution is more
     appropriate than FFT-based methods which add unnecessary complexity.
-    
+
     Parameters
     ----------
     arrival_rates : numpy.ndarray
@@ -326,7 +331,7 @@ class AdmissionGeneratingFunction:
         Array of admission probabilities for each time interval
     method : str, default='empirical'
         Method to use: 'direct', 'parametric', or 'empirical'
-    
+
     Attributes
     ----------
     arrival_rates : numpy.ndarray
@@ -336,84 +341,90 @@ class AdmissionGeneratingFunction:
     method : str
         The prediction method being used
     """
-    
-    def __init__(self, arrival_rates, admission_probs, method='empirical'):
+
+    def __init__(self, arrival_rates, admission_probs, method="empirical"):
         self.arrival_rates = np.array(arrival_rates)
         self.admission_probs = np.array(admission_probs)
         self.method = method
-        
+
         if len(self.arrival_rates) != len(self.admission_probs):
-            raise ValueError("arrival_rates and admission_probs must have the same length")
-    
+            raise ValueError(
+                "arrival_rates and admission_probs must have the same length"
+            )
+
     def get_distribution(self, max_value=50):
         """Get the probability distribution for the specified method.
-        
+
         Parameters
         ----------
         max_value : int, default=50
             Maximum value for the discrete distribution support
-            
+
         Returns
         -------
         pandas.DataFrame
             DataFrame with 'sum' and 'agg_proba' columns representing the distribution
         """
-        if self.method == 'direct':
+        if self.method == "direct":
             return self._direct_distribution(max_value)
-        elif self.method == 'empirical':
+        elif self.method == "empirical":
             return self._empirical_distribution(max_value)
-        elif self.method == 'parametric':
+        elif self.method == "parametric":
             return self._parametric_distribution(max_value)
         else:
             raise ValueError(f"Unknown method: {self.method}")
-    
+
     def _direct_distribution(self, max_value=50):
         """Simple Poisson distribution for direct method."""
         total_rate = np.sum(self.arrival_rates)
         x_values = np.arange(max_value)
         probabilities = poisson.pmf(x_values, total_rate)
-        
-        result_df = pd.DataFrame({
-            'sum': range(len(probabilities)), 
-            'agg_proba': probabilities
-        })
-        
+
+        result_df = pd.DataFrame(
+            {"sum": range(len(probabilities)), "agg_proba": probabilities}
+        )
+
         # Filter and normalize
-        result_df = result_df[result_df['agg_proba'] > 1e-10]
+        result_df = result_df[result_df["agg_proba"] > 1e-10]
         if len(result_df) > 0:
[…]
