### 2025-10-01 8d13aca
remove redistribution of pmf
