+++ src/patientflow/predictors/incoming_admission_predictors.py
@@ -56,21 +56,27 @@ specific hospital settings or specialties.
 
 Assumptions
 -----------
-Parametric incoming-admissions use a simple per-slice Poisson arrivals model with independent
-filtering. The following assumptions are required for the simplified parametric route to be exact:
+Parametric incoming-admissions use a simple per-slice Poisson arrivals model with
+independent filtering. The following assumptions are required for the simplified
+parametric route to be exact:
+
 - Symbols and units:
-  - λ_t: expected arrivals within time-slice t (the value produced by `time_varying_arrival_rates`
-    for that slice; units match the slice length).
-  - θ_t: probability an arrival in slice t is admitted within the prediction window
-    (computed via `get_y_from_aspirational_curve`).
+    - λ_t: expected arrivals within time-slice t (the value produced by
+      `time_varying_arrival_rates` for that slice; units match the slice length).
+    - θ_t: probability an arrival in slice t is admitted within the prediction
+      window (computed via `get_y_from_aspirational_curve`).
+
 - Arrivals within each time-slice t follow a Poisson process with rate λ_t.
-- Each arrival is independently admitted within the prediction window with probability θ_t,
-  which is constant across that slice (given the model inputs). This independent filtering is
-  sometimes called “Poisson thinning”.
+
+- Each arrival is independently admitted within the prediction window with
+  probability θ_t, which is constant across that slice (given the model inputs).
+  This independent filtering is sometimes called "Poisson thinning".
+
 - Slices are independent for this filtering.
-Under these assumptions, admitted arrivals in slice t are Poisson(λ_t θ_t) and the total admitted
-count is Poisson(Σ_t λ_t θ_t). Intuition: filtering arrivals with probability θ_t reduces the
-effective rate from λ_t to λ_t θ_t.
+
+Under these assumptions, admitted arrivals in slice t are Poisson(λ_t θ_t) and
+the total admitted count is Poisson(Σ_t λ_t θ_t). Intuition: filtering arrivals
+with probability θ_t reduces the effective rate from λ_t to λ_t θ_t.
 
 """
 
@@ -103,155 +109,7 @@ from patientflow.calculate.survival_curve import (
 # Import sklearn base classes for custom transformer creation
 from sklearn.base import BaseEstimator, TransformerMixin
 
-from scipy.stats import binom, poisson
-
-
-def weighted_poisson_binomial(i, lam, theta):
-    """Calculate weighted probabilities using Poisson and Binomial distributions.
-
-    Parameters
-    ----------
-    i : int
-        The upper bound of the range for the binomial distribution.
-    lam : float
-        The lambda parameter for the Poisson distribution.
-    theta : float
-        The probability of success for the binomial distribution.
-
-    Returns
-    -------
-    numpy.ndarray
-        An array of weighted probabilities.
-
-    Raises
-    ------
-    ValueError
-        If i < 0, lam < 0, or theta is not between 0 and 1.
-    """
-    if i < 0 or lam < 0 or not 0 <= theta <= 1:
-        raise ValueError("Ensure i >= 0, lam >= 0, and 0 <= theta <= 1.")
-
-    arr_seq = np.arange(i + 1)
-    probabilities = binom.pmf(arr_seq, i, theta)
-    return poisson.pmf(i, lam) * probabilities
-
-
-def aggregate_probabilities(lam, kmax, theta, time_index):
-    """Aggregate probabilities for a range of values using the weighted Poisson-Binomial distribution.
-
-    Parameters
-    ----------
-    lam : numpy.ndarray
-        An array of lambda values for each time interval.
-    kmax : int
-        The maximum number of events to consider.
-    theta : numpy.ndarray
[…]
