### 2025-10-01 c32b196
method to return poisson mean
