+++ src/patientflow/predictors/incoming_admission_predictors.py
@@ -688,6 +688,97 @@ class IncomingAdmissionPredictor(BaseEstimator, TransformerMixin, ABC):
         """
         pass
 
+    @abstractmethod
+    def _get_admission_probabilities(self, **kwargs) -> np.ndarray:
+        """Get admission probabilities for each time interval.
+
+        This is an abstract method that must be implemented by subclasses.
+        Each subclass implements its own logic for calculating admission probabilities
+        based on their specific approach (direct, parametric, or empirical).
+
+        Parameters
+        ----------
+        **kwargs
+            Additional keyword arguments specific to the prediction method.
+
+        Returns
+        -------
+        numpy.ndarray
+            Array of admission probabilities for each time interval.
+        """
+        pass
+
+    def predict_mean(self, prediction_context: Dict, **kwargs) -> Dict:
+        """Return just the Poisson mean (expected value) for each context.
+
+        This method extracts the underlying Poisson parameters without computing
+        the full probability distribution, making it suitable for later generation
+        or when only the expected value is needed.
+
+        Parameters
+        ----------
+        prediction_context : dict
+            A dictionary defining the context for which predictions are to be made.
+            It should specify either a general context or one based on the applied filters.
+        **kwargs
+            Additional keyword arguments specific to the prediction method.
+            These are passed to the subclass-specific _get_admission_probabilities method.
+
+        Returns
+        -------
+        dict
+            Dictionary with Poisson means and parameters for each context.
+            Structure::
+
+                {
+                    filter_key: {
+                        "mean": float,           # Poisson mean (expected value)
+                        "method": str,           # Prediction method used
+                        "arrival_rates": list,   # Original arrival rates
+                        "admission_probs": list, # Calculated admission probabilities
+                        "prediction_time": tuple # Time of prediction
+                    }
+                }
+
+        Raises
+        ------
+        ValueError
+            If filter key is not recognized or prediction_time is not provided.
+        KeyError
+            If required keys are missing from the prediction context.
+        """
+        # Get admission probabilities using subclass-specific method
+        admission_probs = self._get_admission_probabilities(**kwargs)
+
+        predictions = {}
+
+        for filter_key, prediction_time, arrival_rates in self._iter_prediction_inputs(
+            prediction_context
+        ):
+            # Calculate Poisson mean: sum of arrival_rates * admission_probs
+            poisson_mean = float(
+                np.sum(np.array(arrival_rates) * np.array(admission_probs))
+            )
+
+            predictions[filter_key] = {
+                "mean": poisson_mean,
+                "method": self.__class__.__name__.replace(
+                    "IncomingAdmissionPredictor", ""
+                ).lower()
+                or "direct",
+                "arrival_rates": arrival_rates.tolist(),
+                "admission_probs": admission_probs.tolist(),
+                "prediction_time": prediction_time,
+            }
+
+            if self.verbose:
+                self.logger.info(
+                    f"Mean prediction for {filter_key} at {prediction_time}: "
+                    f"Expected admissions = {poisson_mean:.2f}"
+                )
+
+        return predictions
+
 
 class DirectAdmissionPredictor(IncomingAdmissionPredictor):
     """A predictor that assumes every arrival is admitted immediately.
@@ -774,6 +865,26 @@ class DirectAdmissionPredictor(IncomingAdmissionPredictor):
                 )
 
         return predictions
+
+    def _get_admission_probabilities(self, **kwargs) -> np.ndarray:
+        """Get admission probabilities for direct method (always 1.0).
+
+        For the direct predictor, all arrivals are assumed to be admitted immediately,
[…]
