### 2025-10-06 a0eb334
simplify predict mean function
