### 2025-10-06 be9a4b6
return single value from predict mean function
