### 2025-10-06 2536dcb
create hierarchy | prepare subspec preds for consolidation
