+++ src/patientflow/predict/hierarchy.py
@@ -1,0 +1,1012 @@
+"""Hierarchical demand prediction for hospital bed capacity management.
+
+This module provides classes and functions for predicting hospital bed demand at multiple
+hierarchical levels (subspecialty, reporting unit, division, board, hospital) using
+probability distributions and convolution operations.
+
+The main components are:
+
+- DemandPrediction: A dataclass representing prediction results with probabilities,
+  expected values, and percentiles
+- DemandPredictor: Core prediction engine using convolution of probability distributions
+- HospitalHierarchy: Represents organizational structure of a hospital
+- HierarchicalPredictor: High-level interface for making predictions across all levels
+- Utility functions for populating hierarchy from DataFrames
+
+Functions
+---------
+populate_hierarchy_from_dataframe
+    Create HospitalHierarchy from pandas DataFrame with organizational structure
+populate_hierarchical_predictor_from_dataframe
+    Create complete HierarchicalPredictor from DataFrame and parameters
+"""
+
+import numpy as np
+import pandas as pd
+from typing import List, Dict, Optional
+from dataclasses import dataclass
+from scipy.stats import poisson
+
+
+@dataclass
+class DemandPrediction:
+    """Result of a demand prediction at any hierarchical level.
+    
+    This dataclass encapsulates the complete prediction results for a single
+    organizational entity (subspecialty, reporting unit, division, board, or hospital).
+    It contains the probability mass function, expected value, and key percentiles.
+    
+    Parameters
+    ----------
+    entity_id : str
+        Unique identifier for the organizational entity (e.g., subspecialty name)
+    entity_type : str
+        Type of entity ('subspecialty', 'reporting_unit', 'division', 'board', 'hospital')
+    probabilities : numpy.ndarray
+        Probability mass function for bed demand (probability of 0, 1, 2, ... beds)
+    expected_value : float
+        Expected number of beds needed (mean of the distribution)
+    percentiles : dict[int, int]
+        Dictionary mapping percentile values (50, 75, 90, 95, 99) to bed counts
+        
+    Attributes
+    ----------
+    entity_id : str
+        Unique identifier for the organizational entity
+    entity_type : str
+        Type of entity in the hierarchy
+    probabilities : numpy.ndarray
+        Probability mass function for bed demand
+    expected_value : float
+        Expected number of beds needed
+    percentiles : dict[int, int]
+        Percentile values for bed demand
+    max_beds : int
+        Maximum number of beds in the probability distribution
+        
+    Notes
+    -----
+    The probabilities array represents P(X=k) where X is the number of beds needed
+    and k ranges from 0 to max_beds. The sum of all probabilities should equal 1.0.
+    """
+    entity_id: str
+    entity_type: str
+    probabilities: np.ndarray
+    expected_value: float
+    percentiles: Dict[int, int]
+    
+    @property
+    def max_beds(self) -> int:
+        """Maximum number of beds in the probability distribution.
+        
+        Returns
+        -------
+        int
+            Maximum bed count (length of probabilities array minus 1)
+        """
+        return len(self.probabilities) - 1
+
+
+class DemandPredictor:
+    """Hierarchical demand prediction for hospital bed capacity.
+    
+    This class provides the core prediction engine for computing bed demand at
+    different hierarchical levels. It uses convolution of probability distributions
+    to combine predictions from lower levels into higher levels.
+    
+    The prediction process involves:
+    1. Generating Poisson distributions for yet-to-arrive patients
+    2. Combining with current patient distributions using convolution
+    3. Aggregating across organizational levels using multiple convolutions
+    4. Computing statistics (expected value, percentiles) for each level
+    
+    Parameters
+    ----------
+    epsilon : float, default=1e-7
[…]
