+++ tests/test_subspecialty.py
@@ -1,0 +1,441 @@
+import unittest
+from datetime import timedelta
+
+import numpy as np
+import pandas as pd
+
+from patientflow.predict.subspecialty import build_subspecialty_data
+from patientflow.model_artifacts import TrainedClassifier, TrainingResults
+from patientflow.load import get_model_key
+
+from sklearn.compose import ColumnTransformer
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import OneHotEncoder
+from xgboost import XGBClassifier
+
+from patientflow.predictors.sequence_to_outcome_predictor import (
+    SequenceToOutcomePredictor,
+)
+from patientflow.predictors.value_to_outcome_predictor import (
+    ValueToOutcomePredictor,
+)
+from patientflow.predictors.subgroup_predictor import (
+    create_subgroup_functions,
+    MultiSubgroupPredictor,
+)
+from patientflow.predictors.incoming_admission_predictors import (
+    ParametricIncomingAdmissionPredictor,
+    EmpiricalIncomingAdmissionPredictor,
+    DirectAdmissionPredictor,
+)
+from patientflow.prepare import create_yta_filters
+
+
+def _create_random_df(n=1000, include_consults=False):
+    np.random.seed(0)
+    snapshot_date = pd.Timestamp("2023-01-01") + pd.Timedelta(
+        minutes=np.random.randint(0, 60 * 24 * 7)
+    )
+    snapshot_date = [snapshot_date] * n
+    age_on_arrival = np.random.randint(1, 100, size=n)
+    elapsed_los = np.random.randint(0, 3 * 24 * 3600, size=n)
+    arrival_method = np.random.choice(
+        ["ambulance", "public_transport", "walk-in"], size=n
+    )
+    sex = np.random.choice(["M", "F"], size=n)
+    is_admitted = np.random.choice([0, 1], size=n)
+
+    df = pd.DataFrame(
+        {
+            "snapshot_date": snapshot_date,
+            "age_on_arrival": age_on_arrival,
+            "elapsed_los": elapsed_los,
+            "arrival_method": arrival_method,
+            "sex": sex,
+            "is_admitted": is_admitted,
+        }
+    )
+
+    if include_consults:
+        consultations = ["medical", "surgical", "haem/onc", "paediatric"]
+        df["final_sequence"] = [
+            [
+                str(x)
+                for x in np.random.choice(consultations, size=np.random.randint(1, 4))
+            ]
+            for _ in range(n)
+        ]
+        df["consultation_sequence"] = [
+            seq[: -np.random.randint(0, len(seq))] if len(seq) > 0 else []
+            for seq in df["final_sequence"]
+        ]
+        df["specialty"] = [
+            str(np.random.choice(seq)) if len(seq) > 0 else consultations[0]
+            for seq in df["final_sequence"]
+        ]
+
+    return df
+
+
+def _create_random_arrivals(n=1000):
+    np.random.seed(0)
+    base_date = pd.Timestamp("2023-01-01")
+    arrival_datetime = [
+        base_date + pd.Timedelta(minutes=np.random.randint(0, 60 * 24 * 7))
+        for _ in range(n)
+    ]
+    specialties = ["medical", "surgical", "haem/onc", "paediatric"]
+    specialty = np.random.choice(specialties, size=n)
+    is_child = np.random.choice([True, False], size=n)
+    df = pd.DataFrame(
+        {
+            "arrival_datetime": arrival_datetime,
+            "specialty": specialty,
+            "is_child": is_child,
+        }
+    )
+    return df
+
+
+def _create_random_arrivals_with_departures(n=1000):
+    np.random.seed(0)
[…]
