### 2025-10-06 6400d31
create hierarchy | prepare subspec preds for consolidation
