+++ src/patientflow/predict/subspecialty.py
@@ -1,0 +1,367 @@
+"""Demand preparation utilities for later consolidation.
+
+This module prepares per-subspecialty inputs that can be consumed by any
+downstream roll-up or forecasting workflow. It converts trained model
+outputs and current patient snapshots into:
+
+- A probability mass function (PMF) for admissions from current ED patients
+- Poisson means for yet-to-arrive admissions (ED, non-ED emergency, elective)
+
+The outputs are independent per subspecialty and do not presuppose any
+particular consolidation hierarchy. They can be used directly for single-
+specialty analyses or fed into any combination scheme (including hierarchical
+schemes) implemented elsewhere.
+
+Functions
+---------
+build_subspecialty_data
+    Prepare subspecialty-level prediction inputs from trained models and snapshots
+
+Notes
+-----
+This module integrates with the broader patientflow ecosystem by:
+1. Using trained classifiers and specialty models from the train module
+2. Processing patient snapshots from the prepare module  
+3. Computing admission probabilities using the calculate module
+
+"""
+
+from typing import Dict, List, Optional, Tuple, Union, Any
+
+import numpy as np
+import pandas as pd
+
+from patientflow.predict.emergency_demand import (
+    add_missing_columns,
+    get_specialty_probs,
+)
+from patientflow.predictors.incoming_admission_predictors import (
+    ParametricIncomingAdmissionPredictor,
+    EmpiricalIncomingAdmissionPredictor,
+    DirectAdmissionPredictor,
+)
+from patientflow.predictors.sequence_to_outcome_predictor import (
+    SequenceToOutcomePredictor,
+)
+from patientflow.predictors.value_to_outcome_predictor import (
+    ValueToOutcomePredictor,
+)
+from patientflow.predictors.subgroup_predictor import (
+    MultiSubgroupPredictor,
+)
+from patientflow.aggregate import (
+    model_input_to_pred_proba,
+    pred_proba_to_agg_predicted,
+)
+from patientflow.calculate.admission_in_prediction_window import (
+    calculate_probability,
+    calculate_admission_probability_from_survival_curve,
+)
+from patientflow.model_artifacts import TrainedClassifier
+
+
+def build_subspecialty_data(
+    models: Tuple[
+        TrainedClassifier,
+        Union[SequenceToOutcomePredictor, ValueToOutcomePredictor, MultiSubgroupPredictor],
+        Union[
+            ParametricIncomingAdmissionPredictor,
+            EmpiricalIncomingAdmissionPredictor,
+        ],
+        DirectAdmissionPredictor,
+        DirectAdmissionPredictor,
+    ],
+    prediction_time: Tuple[int, int],
+    prediction_snapshots: pd.DataFrame,
+    specialties: List[str],
+    prediction_window,
+    x1: float,
+    y1: float,
+    x2: float,
+    y2: float,
+    cdf_cut_points: Optional[List[float]] = None,
+    use_admission_in_window_prob: bool = True,
+) -> Dict[str, Dict[str, Any]]:
+    """Build per-subspecialty inputs for downstream roll-up.
+
+    This function processes current patient snapshots through trained models and
+    computes, for each subspecialty, the probability distribution of admissions
+    from current ED patients and the expected means of yet-to-arrive admissions.
+    
+    The function combines three sources of demand:
+    1. Current ED patients (converted to probability mass function)
+    2. Yet-to-arrive ED patients (converted to Poisson parameters)
+    3. Yet-to-arrive non-ED emergency patients (converted to Poisson parameters)
+    4. Yet-to-arrive elective patients (converted to Poisson parameters)
+    
+    Parameters
+    ----------
+    models : tuple
+        Tuple of five trained models:
+        - classifier: TrainedClassifier for admission probability prediction
[…]
