### 2025-10-06 71acbcd
create hierarchy | prepare subspec preds for consolidation
