### 2025-10-06 6a5fc31
reinstate special_params logic
