### 2025-10-07 9dce2d2
make naming of subspecialty_data clearer
