+++ src/patientflow/predict/subspecialty.py
@@ -21,7 +21,7 @@ Notes
 -----
 This module integrates with the broader patientflow ecosystem by:
 1. Using trained classifiers and specialty models from the train module
-2. Processing patient snapshots from the prepare module  
+2. Processing patient snapshots from the prepare module
 3. Computing admission probabilities using the calculate module
 
 """
@@ -63,7 +63,9 @@ from patientflow.model_artifacts import TrainedClassifier
 def build_subspecialty_data(
     models: Tuple[
         TrainedClassifier,
-        Union[SequenceToOutcomePredictor, ValueToOutcomePredictor, MultiSubgroupPredictor],
+        Union[
+            SequenceToOutcomePredictor, ValueToOutcomePredictor, MultiSubgroupPredictor
+        ],
         Union[
             ParametricIncomingAdmissionPredictor,
             EmpiricalIncomingAdmissionPredictor,
@@ -87,13 +89,13 @@ def build_subspecialty_data(
     This function processes current patient snapshots through trained models and
     computes, for each subspecialty, the probability distribution of admissions
     from current ED patients and the expected means of yet-to-arrive admissions.
-    
+
     The function combines three sources of demand:
     1. Current ED patients (converted to probability mass function)
     2. Yet-to-arrive ED patients (converted to Poisson parameters)
     3. Yet-to-arrive non-ED emergency patients (converted to Poisson parameters)
     4. Yet-to-arrive elective patients (converted to Poisson parameters)
-    
+
     Parameters
     ----------
     models : tuple
@@ -128,14 +130,14 @@ def build_subspecialty_data(
     -------
     dict[str, dict[str, Any]]
         Dictionary mapping subspecialty_id to prediction parameters:
-        - 'prob_admission_pats_in_ed': numpy.ndarray
-          Probability mass function for current ED admissions within window
-        - 'lambda_ed_yta': float
-          Poisson parameter for ED yet-to-arrive admissions
-        - 'lambda_non_ed_yta': float
-          Poisson parameter for non-ED emergency admissions
-        - 'lambda_elective_yta': float
-          Poisson parameter for elective admissions
+        - 'pmf_ed_current_within_window': numpy.ndarray
+          Probability mass function for current ED admissions within the prediction window
+        - 'lambda_ed_yta_within_window': float
+          Poisson parameter for ED yet-to-arrive admissions within the prediction window
+        - 'lambda_non_ed_yta_within_window': float
+          Poisson parameter for non-ED emergency admissions within the prediction window
+        - 'lambda_elective_yta_within_window': float
+          Poisson parameter for elective admissions within the prediction window
 
     Raises
     ------
@@ -146,7 +148,13 @@ def build_subspecialty_data(
         don't match between models and requested parameters
 
     """
-    classifier, spec_model, yet_to_arrive_model, non_ed_yta_model, elective_yta_model = models
+    (
+        classifier,
+        spec_model,
+        yet_to_arrive_model,
+        non_ed_yta_model,
+        elective_yta_model,
+    ) = models
 
     # Validate model types
     if not isinstance(classifier, TrainedClassifier):
@@ -174,9 +182,13 @@ def build_subspecialty_data(
 
     # Validate that non-ED and elective models are DirectAdmissionPredictor
     if not isinstance(non_ed_yta_model, DirectAdmissionPredictor):
-        raise TypeError("Fourth model must be of type DirectAdmissionPredictor (non-ED emergency)")
+        raise TypeError(
+            "Fourth model must be of type DirectAdmissionPredictor (non-ED emergency)"
+        )
[…]
