+++ src/patientflow/predict/hierarchy.py
@@ -31,11 +31,11 @@ from scipy.stats import poisson
 @dataclass
 class DemandPrediction:
     """Result of a demand prediction at any hierarchical level.
-    
+
     This dataclass encapsulates the complete prediction results for a single
     organizational entity (subspecialty, reporting unit, division, board, or hospital).
     It contains the probability mass function, expected value, and key percentiles.
-    
+
     Parameters
     ----------
     entity_id : str
@@ -48,7 +48,7 @@ class DemandPrediction:
         Expected number of beds needed (mean of the distribution)
     percentiles : dict[int, int]
         Dictionary mapping percentile values (50, 75, 90, 95, 99) to bed counts
-        
+
     Attributes
     ----------
     entity_id : str
@@ -63,22 +63,23 @@ class DemandPrediction:
         Percentile values for bed demand
     max_beds : int
         Maximum number of beds in the probability distribution
-        
+
     Notes
     -----
     The probabilities array represents P(X=k) where X is the number of beds needed
     and k ranges from 0 to max_beds. The sum of all probabilities should equal 1.0.
     """
+
     entity_id: str
     entity_type: str
     probabilities: np.ndarray
     expected_value: float
     percentiles: Dict[int, int]
-    
+
     @property
     def max_beds(self) -> int:
         """Maximum number of beds in the probability distribution.
-        
+
         Returns
         -------
         int
@@ -89,87 +90,86 @@ class DemandPrediction:
 
 class DemandPredictor:
     """Hierarchical demand prediction for hospital bed capacity.
-    
+
     This class provides the core prediction engine for computing bed demand at
     different hierarchical levels. It uses convolution of probability distributions
     to combine predictions from lower levels into higher levels.
-    
+
     The prediction process involves:
     1. Generating Poisson distributions for yet-to-arrive patients
     2. Combining with current patient distributions using convolution
     3. Aggregating across organizational levels using multiple convolutions
     4. Computing statistics (expected value, percentiles) for each level
-    
+
     Parameters
     ----------
     epsilon : float, default=1e-7
         Truncation threshold for probability distribution tails. Distributions
         are truncated when the cumulative probability exceeds (1 - epsilon).
-        
+
     Attributes
     ----------
     epsilon : float
         Truncation threshold for probability distributions
     cache : dict
         Cache for storing computed predictions (currently unused)
-        
+
     Notes
     -----
     The class uses discrete convolution to combine probability distributions.
     For computational efficiency, distributions are periodically truncated during
     multiple convolutions to prevent exponential growth in array sizes.
     """
-    
+
     def __init__(self, epsilon: float = 1e-7):
         """Initialize the demand predictor.
-        
+
         Parameters
         ----------
         epsilon : float, default=1e-7
             Truncation threshold for probability distribution tails
         """
         self.epsilon = epsilon
-        self.cache = {}
-    
+        self.cache: Dict[str, DemandPrediction] = {}
+
     def predict_subspecialty(
         self,
         subspecialty_id: str,
-        prob_admission_pats_in_ed: np.ndarray,
-        lambda_ed_yta: float,
-        lambda_non_ed_yta: float,
-        lambda_elective_yta: float
+        pmf_ed_current_within_window: np.ndarray,
+        lambda_ed_yta_within_window: float,
+        lambda_non_ed_yta_within_window: float,
+        lambda_elective_yta_within_window: float,
     ) -> DemandPrediction:
         """Predict demand for a single subspecialty.
-        
+
         This method combines three sources of patient demand:
-        1. Current ED patients (using pre-computed probability distribution)
-        2. Yet-to-arrive ED patients (Poisson distribution)
[…]

[…]
