### 2025-10-07 4ad430d
make naming of subspecialty_data clearer
