### 2025-10-07 d57a346
remove duplicate function
