+++ src/patientflow/predict/hierarchy.py
@@ -903,68 +903,10 @@ def populate_hierarchy_from_dataframe(
     return hierarchy
 
 
-def populate_hierarchical_predictor_from_dataframe(
-    df: pd.DataFrame,
-    hospital_id: str,
-    subspecialty_data: Dict[str, Dict],
-    epsilon: float = 1e-7,
-) -> HierarchicalPredictor:
-    """Create and populate a complete HierarchicalPredictor from DataFrame.
-
-    This convenience function creates a fully configured HierarchicalPredictor
-    by combining a hospital hierarchy (derived from the DataFrame) with a
-    DemandPredictor configured with the specified epsilon value.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        DataFrame containing hospital organizational structure. Must have columns:
-        - 'board': Board identifier
-        - 'division': Division identifier
-        - 'reporting_unit': Reporting unit identifier
-        - 'sub_specialty': Subspecialty identifier
-    hospital_id : str
-        Hospital identifier to link all boards to the hospital
-    subspecialty_data : dict[str, dict]
-        Dictionary mapping subspecialty_id to prediction parameters.
-        Not used in this function but included for signature compatibility.
-    epsilon : float, default=1e-7
-        Truncation threshold for probability distributions in the DemandPredictor
-
-    Returns
-    -------
-    HierarchicalPredictor
-        Fully configured predictor ready for hierarchical predictions
-
-    Notes
-    -----
-    This function is a convenience wrapper that:
-    1. Creates a HospitalHierarchy from the DataFrame
-    2. Links all boards to the specified hospital
-    3. Creates a DemandPredictor with the specified epsilon
-    4. Returns a configured HierarchicalPredictor
-
-    The subspecialty_data parameter is included for API compatibility but
-    is not used in this function.
-    """
-    # Create hierarchy from DataFrame
-    hierarchy = populate_hierarchy_from_dataframe(df)
-
-    # Add all boards to hospital
-    unique_boards = df["board"].dropna().unique()
-    for board in unique_boards:
-        hierarchy.add_board(board, hospital_id)
-
-    # Create predictor with specified epsilon
-    predictor = DemandPredictor(epsilon=epsilon)
-    hierarchical_predictor = HierarchicalPredictor(hierarchy, predictor)
-
-    return hierarchical_predictor
-
-
 def create_hierarchical_predictor(
     specs_df: pd.DataFrame,
     hospital_id: str,
+    subspecialty_data: Dict[str, Dict],
     epsilon: float = 1e-7,
 ) -> HierarchicalPredictor:
     """Create a HierarchicalPredictor from a hospital structure DataFrame.
@@ -985,6 +927,10 @@ def create_hierarchical_predictor(
         Additional columns are ignored.
     hospital_id : str
         Hospital identifier to link all boards to a single hospital
+    subspecialty_data : dict[str, dict]
+        Dictionary mapping subspecialty_id to prediction parameters prepared by
+        build_subspecialty_data. Used for validation here and passed to downstream
+        prediction calls via HierarchicalPredictor.predict_all_levels().
     epsilon : float, default=1e-7
         Truncation threshold for probability distribution tails during
         convolution operations. Smaller values provide higher accuracy but
@@ -996,7 +942,7 @@ def create_hierarchical_predictor(
         Fully configured predictor with:
         - HospitalHierarchy populated from specs_df
         - DemandPredictor configured with specified epsilon
-        - Ready to accept subspecialty_data for predictions
+        - Ready to use with the provided subspecialty_data for predictions
 
     Notes
     -----
@@ -1004,12 +950,23 @@ def create_hierarchical_predictor(
 
[…]
