### 2025-10-07 6087be5
clarity instantiation of HierarchicalPredictor
