### 2025-10-07 5fbdccc
prettify DemandPrediction output
