### 2025-10-07 afbf8fc
add discharge prediction to subspecialty_data
