+++ src/patientflow/predict/subspecialty.py
@@ -63,6 +63,7 @@ from patientflow.model_artifacts import TrainedClassifier
 def build_subspecialty_data(
     models: Tuple[
         TrainedClassifier,
+        TrainedClassifier,
         Union[
             SequenceToOutcomePredictor, ValueToOutcomePredictor, MultiSubgroupPredictor
         ],
@@ -74,7 +75,8 @@ def build_subspecialty_data(
         DirectAdmissionPredictor,
     ],
     prediction_time: Tuple[int, int],
-    prediction_snapshots: pd.DataFrame,
+    ed_snapshots: pd.DataFrame,
+    inpatient_snapshots: pd.DataFrame,
     specialties: List[str],
     prediction_window,
     x1: float,
@@ -88,19 +90,22 @@ def build_subspecialty_data(
 
     This function processes current patient snapshots through trained models and
     computes, for each subspecialty, the probability distribution of admissions
-    from current ED patients and the expected means of yet-to-arrive admissions.
+    from current ED patients, departures from current inpatients, and the expected
+    means of yet-to-arrive admissions.
 
-    The function combines three sources of demand:
+    The function combines five sources of demand:
     1. Current ED patients (converted to probability mass function)
-    2. Yet-to-arrive ED patients (converted to Poisson parameters)
-    3. Yet-to-arrive non-ED emergency patients (converted to Poisson parameters)
-    4. Yet-to-arrive elective patients (converted to Poisson parameters)
+    2. Current inpatients (converted to probability mass function for departures)
+    3. Yet-to-arrive ED patients (converted to Poisson parameters)
+    4. Yet-to-arrive non-ED emergency patients (converted to Poisson parameters)
+    5. Yet-to-arrive elective patients (converted to Poisson parameters)
 
     Parameters
     ----------
     models : tuple
-        Tuple of five trained models:
-        - classifier: TrainedClassifier for admission probability prediction
+        Tuple of six trained models:
+        - ed_classifier: TrainedClassifier for ED admission probability prediction
+        - inpatient_classifier: TrainedClassifier for inpatient departure probability prediction
         - spec_model: SequenceToOutcomePredictor | ValueToOutcomePredictor | MultiSubgroupPredictor
           for specialty assignment probabilities
         - ed_yta_model: ParametricIncomingAdmissionPredictor | EmpiricalIncomingAdmissionPredictor
@@ -109,9 +114,12 @@ def build_subspecialty_data(
         - elective_yta_model: DirectAdmissionPredictor for elective predictions
     prediction_time : tuple[int, int]
         Hour and minute for inference time
-    prediction_snapshots : pandas.DataFrame
+    ed_snapshots : pandas.DataFrame
         DataFrame of current ED patients. Must include 'elapsed_los' column as timedelta.
         Each row represents a patient currently in the ED.
+    inpatient_snapshots : pandas.DataFrame
+        DataFrame of current inpatients. Must include 'elapsed_los' column as timedelta.
+        Each row represents a patient currently in a subspecialty ward.
     specialties : list[str]
         List of subspecialties to prepare inputs for
     prediction_window : datetime.timedelta
@@ -132,6 +140,8 @@ def build_subspecialty_data(
         Dictionary mapping subspecialty_id to prediction parameters:
         - 'pmf_ed_current_within_window': numpy.ndarray
           Probability mass function for current ED admissions within the prediction window
+        - 'pmf_inpatient_departures_within_window': numpy.ndarray
+          Probability mass function for current inpatient departures within the prediction window
         - 'lambda_ed_yta_within_window': float
[…]
