### 2025-10-07 c7c9c09
add discharge prediction to subspecialty_data
