+++ src/patientflow/predict/__init__.py
@@ -3,3 +3,17 @@
 This module provides functions for making predictions about future patient flow,
 including emergency demand forecasting and other predictive analytics.
 """
+
+from patientflow.predict.subspecialty import (
+    build_subspecialty_data,
+    compute_transfer_arrivals,
+    scale_pmf_by_probability,
+    convolve_pmfs,
+)
+
+__all__ = [
+    "build_subspecialty_data",
+    "compute_transfer_arrivals",
+    "scale_pmf_by_probability",
+    "convolve_pmfs",
+]
