+++ src/patientflow/predictors/transfer_predictor.py
@@ -1,0 +1,434 @@
+"""
+Transfer probability predictor for patient movements between subspecialties.
+
+This module provides a predictor for modelling patient transfers between hospital
+subspecialties based on historical movement patterns. It computes the probability
+that a departing patient will transfer to another subspecialty versus being
+discharged, and the distribution of destination subspecialties for transfers.
+
+The predictor follows sklearn conventions with fit() and predict() methods, allowing
+it to be saved and loaded as a model artifact alongside other predictors in the
+patientflow ecosystem.
+
+Classes
+-------
+TransferProbabilityPredictor : BaseEstimator, TransformerMixin
+    Predictor for patient transfer probabilities between subspecialties.
+    Stores static transfer probabilities derived from historical patient movement data.
+
+Notes
+-----
+This predictor is designed to work in conjunction with inpatient departure classifiers
+to provide a complete picture of subspecialty bed demand by accounting for both
+external discharges and internal transfers.
+
+The predictor stores transition probabilities between subspecialties, which can be
+used at prediction time to model internal patient flows and compute arrival
+distributions for each subspecialty from transfers.
+"""
+
+import pandas as pd
+import numpy as np
+from typing import Dict, Set, List
+from sklearn.base import BaseEstimator, TransformerMixin
+
+
+class TransferProbabilityPredictor(BaseEstimator, TransformerMixin):
+    """Predict patient transfer destinations from subspecialty movement patterns.
+
+    This predictor computes and stores transfer probabilities between subspecialties
+    based on historical patient movement data. For each source subspecialty, it
+    calculates:
+    1. The probability that a departure results in a transfer (vs discharge)
+    2. The distribution of destinations given that a transfer occurs
+
+    The predictor follows sklearn conventions with fit() and predict() methods,
+    allowing it to be saved and loaded as a model artifact.
+
+    Parameters
+    ----------
+    source_col : str, default='current_subspecialty'
+        Name of the column containing the source subspecialty (where patient is leaving from)
+    destination_col : str, default='next_subspecialty'
+        Name of the column containing the destination subspecialty (where patient is going to).
+        None/NaN values indicate discharge rather than transfer.
+
+    Attributes
+    ----------
+    source_col : str
+        Name of the source subspecialty column
+    destination_col : str
+        Name of the destination subspecialty column
+    transfer_probabilities : dict
+        Nested dictionary containing transfer statistics for each subspecialty:
+        {
+            'source_subspecialty': {
+                'prob_transfer': float,
+                    Probability that a departure from this subspecialty results
+                    in a transfer to another subspecialty (vs discharge)
+                'destination_distribution': dict,
+                    Distribution of destinations given a transfer occurs.
+                    Maps destination subspecialty names to probabilities that
+                    sum to 1.0
+            }
+        }
+    subspecialties : set
+        Set of all subspecialties in the system
+    is_fitted_ : bool
+        Whether the predictor has been fitted
+
+    Examples
+    --------
+    >>> # Prepare training data with subspecialty movements
+    >>> X = pd.DataFrame({
+    ...     'current_subspecialty': ['cardiology', 'cardiology', 'surgery'],
+    ...     'next_subspecialty': ['surgery', None, None]  # None = discharge
+    ... })
+    >>> subspecialties = {'cardiology', 'surgery', 'medicine'}
+    >>> 
+    >>> # Train the predictor with default column names
+    >>> predictor = TransferProbabilityPredictor()
+    >>> predictor.fit(X, subspecialties)
+    >>> 
+    >>> # Or with custom column names
+    >>> predictor = TransferProbabilityPredictor(
+    ...     source_col='from_ward',
[…]
