+++ src/patientflow/predict/subspecialty.py
@@ -5,7 +5,9 @@ downstream roll-up or forecasting workflow. It converts trained model
 outputs and current patient snapshots into:
 
 - A probability mass function (PMF) for admissions from current ED patients
+- A probability mass function (PMF) for departures from current inpatients
 - Poisson means for yet-to-arrive admissions (ED, non-ED emergency, elective)
+- Transfer arrival distributions from internal subspecialty movements
 
 The outputs are independent per subspecialty and do not presuppose any
 particular consolidation hierarchy. They can be used directly for single-
@@ -16,6 +18,12 @@ Functions
 ---------
 build_subspecialty_data
     Prepare subspecialty-level prediction inputs from trained models and snapshots
+compute_transfer_arrivals
+    Calculate arrival PMFs from internal transfers between subspecialties
+scale_pmf_by_probability
+    Scale a departure PMF by compound transfer probability
+convolve_pmfs
+    Convolve two probability mass functions for aggregation
 
 Notes
 -----
@@ -23,6 +31,7 @@ This module integrates with the broader patientflow ecosystem by:
 1. Using trained classifiers and specialty models from the train module
 2. Processing patient snapshots from the prepare module
 3. Computing admission probabilities using the calculate module
+4. Modeling internal patient transfers using the transfer predictor
 
 """
 
@@ -488,3 +497,250 @@ def build_subspecialty_data(
         subspecialty_data[spec]["lambda_elective_yta_within_window"] = lambda_elective
 
     return subspecialty_data
+
+
+def scale_pmf_by_probability(pmf: np.ndarray, compound_prob: float) -> np.ndarray:
+    """Scale a departure PMF by compound probability for transfer destination.
+
+    This function creates a mixture distribution representing: "given these
+    departures, how many actually transfer to our destination?" It accounts
+    for the fact that most departures don't go to any specific destination.
+
+    The scaling works as follows:
+    - P(0 transfers) = P(0 departed) + P(n>0 departed but none went to destination)
+    - P(k transfers, k>0) = compound_prob × P(k departed)
+
+    This is effectively a Binomial thinning of each possible departure count.
+
+    Parameters
+    ----------
+    pmf : numpy.ndarray
+        Probability mass function for number of departures. pmf[k] = P(k departures)
+    compound_prob : float
+        Combined probability: prob_transfer × prob_destination.
+        Represents P(departure goes to this specific destination)
+
+    Returns
+    -------
+    numpy.ndarray
+        Scaled PMF representing the distribution of transfers to the destination
+
+    Examples
+    --------
+    >>> # 50% chance of 0 departures, 50% chance of 2 departures
+    >>> pmf = np.array([0.5, 0.0, 0.5])
+    >>> # 20% of departures go to our destination
+    >>> scaled = scale_pmf_by_probability(pmf, 0.2)
+    >>> # Most weight on 0 (no departures OR departures went elsewhere)
+    >>> # Some weight on 1 or 2 transfers
+
+    Notes
+    -----
+    For compound_prob = 0.0, returns [1.0, 0.0, ...] (no transfers possible).
+    For compound_prob = 1.0, returns the original pmf (all departures transfer here).
+    """
+    if compound_prob == 0.0:
+        # No transfers to this destination
+        return np.array([1.0, 0.0])
+
+    if compound_prob == 1.0:
+        # All departures go to this destination
+        return pmf.copy()
+
+    n_max = len(pmf) - 1
+    scaled_pmf = np.zeros(n_max + 1)
+
+    # P(0 transfers to destination) includes:
+    # 1. P(0 departures) - these definitely don't transfer
+    # 2. P(k>0 departures) * P(none go to destination)
+    #    = P(k departures) * (1 - compound_prob)^k
[…]
