### 2025-10-09 98fbb75
improve documentation for TransferProbabilityEstimator
