+++ tests/test_transfer_predictor.py
@@ -1,0 +1,159 @@
+import unittest
+import numpy as np
+import pandas as pd
+from patientflow.predictors.transfer_predictor import TransferProbabilityEstimator
+
+
+class TestTransferProbabilityEstimator(unittest.TestCase):
+    """Test suite for TransferProbabilityEstimator class."""
+
+    def setUp(self):
+        """Set up test fixtures."""
+        self.subspecialties = {"cardiology", "surgery", "medicine", "oncology"}
+
+    def test_fit_and_calculate_transfer_probabilities(self):
+        """Test basic fitting and calculation of transfer probabilities."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": [
+                    "cardiology",
+                    "cardiology",
+                    "surgery",
+                    "surgery",
+                    "medicine",
+                ],
+                "next_subspecialty": [
+                    "surgery",
+                    None,
+                    "medicine",
+                    None,
+                    None,
+                ],  # None = discharge
+            }
+        )
+        predictor = TransferProbabilityEstimator()
+        predictor.fit(X, self.subspecialties)
+
+        # Check cardiology: 1 transfer out of 2 departures = 50%
+        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology"), 0.5)
+        # Check surgery: 1 transfer out of 2 departures = 50%
+        self.assertAlmostEqual(predictor.get_transfer_prob("surgery"), 0.5)
+        # Check medicine: 0 transfers out of 1 departure = 0%
+        self.assertAlmostEqual(predictor.get_transfer_prob("medicine"), 0.0)
+        # Check oncology: no data = 0%
+        self.assertAlmostEqual(predictor.get_transfer_prob("oncology"), 0.0)
+
+    def test_destination_distribution_calculation(self):
+        """Test calculation of destination distributions."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology"] * 10,
+                "next_subspecialty": ["surgery"] * 5 + ["medicine"] * 3 + ["oncology"] * 2,
+            }
+        )
+        predictor = TransferProbabilityEstimator()
+        predictor.fit(X, self.subspecialties)
+
+        dests = predictor.get_destination_distribution("cardiology")
+        self.assertAlmostEqual(dests["surgery"], 0.5)
+        self.assertAlmostEqual(dests["medicine"], 0.3)
+        self.assertAlmostEqual(dests["oncology"], 0.2)
+        self.assertAlmostEqual(sum(dests.values()), 1.0)
+
+    def test_all_transfers_calculation(self):
+        """Test when all departures are transfers."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology", "cardiology"],
+                "next_subspecialty": ["surgery", "medicine"],
+            }
+        )
+        predictor = TransferProbabilityEstimator()
+        predictor.fit(X, self.subspecialties)
+
+        self.assertEqual(predictor.get_transfer_prob("cardiology"), 1.0)
+        dests = predictor.get_destination_distribution("cardiology")
+        self.assertAlmostEqual(dests["surgery"], 0.5)
+        self.assertAlmostEqual(dests["medicine"], 0.5)
+
+    def test_no_transfers_calculation(self):
+        """Test when all departures are discharges."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology", "surgery"],
+                "next_subspecialty": [None, None],
+            }
+        )
+        predictor = TransferProbabilityEstimator()
+        predictor.fit(X, self.subspecialties)
+
+        self.assertEqual(predictor.get_transfer_prob("cardiology"), 0.0)
[…]
