+++ src/patientflow/predictors/transfer_predictor.py
@@ -10,12 +10,6 @@ The predictor follows sklearn conventions with fit() and predict() methods, allo
 it to be saved and loaded as a model artifact alongside other predictors in the
 patientflow ecosystem.
 
-Classes
--------
-TransferProbabilityPredictor : BaseEstimator, TransformerMixin
-    Predictor for patient transfer probabilities between subspecialties.
-    Stores static transfer probabilities derived from historical patient movement data.
-
 Notes
 -----
 This predictor is designed to work in conjunction with inpatient departure classifiers
@@ -28,21 +22,20 @@ distributions for each subspecialty from transfers.
 """
 
 import pandas as pd
-import numpy as np
-from typing import Dict, Set, List
+from typing import Dict, Set, Optional
 from sklearn.base import BaseEstimator, TransformerMixin
 
 
-class TransferProbabilityPredictor(BaseEstimator, TransformerMixin):
-    """Predict patient transfer destinations from subspecialty movement patterns.
+class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
+    """Estimate patient transfer probabilities from subspecialty movement patterns.
 
-    This predictor computes and stores transfer probabilities between subspecialties
+    This estimator computes and stores transfer probabilities between subspecialties
     based on historical patient movement data. For each source subspecialty, it
     calculates:
     1. The probability that a departure results in a transfer (vs discharge)
     2. The distribution of destinations given that a transfer occurs
 
-    The predictor follows sklearn conventions with fit() and predict() methods,
+    The estimator follows sklearn conventions with fit() and predict() methods,
     allowing it to be saved and loaded as a model artifact.
 
     Parameters
@@ -85,20 +78,20 @@ class TransferProbabilityPredictor(BaseEstimator, TransformerMixin):
     ...     'next_subspecialty': ['surgery', None, None]  # None = discharge
     ... })
     >>> subspecialties = {'cardiology', 'surgery', 'medicine'}
-    >>> 
-    >>> # Train the predictor with default column names
-    >>> predictor = TransferProbabilityPredictor()
-    >>> predictor.fit(X, subspecialties)
-    >>> 
+    >>>
+    >>> # Train the estimator with default column names
+    >>> estimator = TransferProbabilityEstimator()
+    >>> estimator.fit(X, subspecialties)
+    >>>
     >>> # Or with custom column names
-    >>> predictor = TransferProbabilityPredictor(
+    >>> estimator = TransferProbabilityEstimator(
     ...     source_col='from_ward',
     ...     destination_col='to_ward'
     ... )
-    >>> 
+    >>>
     >>> # Get transfer probabilities
-    >>> prob_transfer = predictor.get_transfer_prob('cardiology')
-    >>> destinations = predictor.get_destination_distribution('cardiology')
+    >>> prob_transfer = estimator.get_transfer_prob('cardiology')
+    >>> destinations = estimator.get_destination_distribution('cardiology')
 
     Notes
     -----
@@ -111,9 +104,13 @@ class TransferProbabilityPredictor(BaseEstimator, TransformerMixin):
     probability distributions, even for subspecialties with no observed transfers.
     """
 
-    def __init__(self, source_col: str = "current_subspecialty", destination_col: str = "next_subspecialty"):
-        """Initialize the transfer probability predictor.
-        
+    def __init__(
+        self,
+        source_col: str = "current_subspecialty",
+        destination_col: str = "next_subspecialty",
+    ):
+        """Initialize the transfer probability estimator.
+
         Parameters
         ----------
         source_col : str, default='current_subspecialty'
@@ -123,12 +120,11 @@ class TransferProbabilityPredictor(BaseEstimator, TransformerMixin):
         """
         self.source_col = source_col
         self.destination_col = destination_col
[…]

[…]
