### 2025-10-09 b2c0c58
improve documentation for TransferProbabilityEstimator
