### 2025-10-09 62e9300
make SubspecialtyPredictionInputs dataclass
