+++ src/patientflow/predict/subspecialty.py
@@ -21,10 +21,11 @@ This module integrates with the broader patientflow ecosystem by:
 1. Using trained classifiers and specialty models from the train module
 2. Processing patient snapshots from the prepare module
 3. Computing admission probabilities using the calculate module
-4. Modeling internal patient transfers using the transfer predictor
+4. Modelling internal patient transfers using the transfer predictor
 
 """
 
+from dataclasses import dataclass
 from typing import Dict, List, Optional, Tuple, Union, Any
 
 import numpy as np
@@ -59,6 +60,45 @@ from patientflow.calculate.admission_in_prediction_window import (
 from patientflow.model_artifacts import TrainedClassifier
 
 
+@dataclass(frozen=True)
+class SubspecialtyPredictionInputs:
+    """Input parameters for subspecialty demand prediction.
+
+    These inputs represent the probability distributions and parameters
+    needed to predict demand for a single subspecialty. This dataclass packages
+    the outputs from build_subspecialty_data for use in hierarchical prediction.
+
+    Attributes
+    ----------
+    pmf_ed_current_within_window : np.ndarray
+        Probability mass function for current ED admissions within the prediction window.
+        Array where pmf[k] represents P(k ED patients admitted to this subspecialty).
+    pmf_inpatient_departures_within_window : np.ndarray
+        Probability mass function for current inpatient departures within the prediction window.
+        Array where pmf[k] represents P(k current inpatients depart from this subspecialty).
+    lambda_ed_yta_within_window : float
+        Poisson parameter (rate) for ED yet-to-arrive admissions within the prediction window.
+        Represents the expected number of new ED arrivals who will be admitted to this subspecialty.
+    lambda_non_ed_yta_within_window : float
+        Poisson parameter (rate) for non-ED emergency admissions within the prediction window.
+        Represents the expected number of direct emergency admissions to this subspecialty.
+    lambda_elective_yta_within_window : float
+        Poisson parameter (rate) for elective admissions within the prediction window.
+        Represents the expected number of planned/elective admissions to this subspecialty.
+
+    Notes
+    -----
+    This dataclass is immutable (frozen=True) to prevent accidental modification after creation.
+    All arrays and parameters should represent distributions/rates for the same prediction window.
+    """
+
+    pmf_ed_current_within_window: np.ndarray
+    pmf_inpatient_departures_within_window: np.ndarray
+    lambda_ed_yta_within_window: float
+    lambda_non_ed_yta_within_window: float
+    lambda_elective_yta_within_window: float
+
+
 def build_subspecialty_data(
     models: Tuple[
         TrainedClassifier,
@@ -84,7 +124,7 @@ def build_subspecialty_data(
     y2: float,
     cdf_cut_points: Optional[List[float]] = None,
     use_admission_in_window_prob: bool = True,
-) -> Dict[str, Dict[str, Any]]:
+) -> Dict[str, SubspecialtyPredictionInputs]:
     """Build per-subspecialty inputs for downstream roll-up.
 
     This function processes current patient snapshots through trained models and
@@ -129,19 +169,9 @@ def build_subspecialty_data(
 
     Returns
     -------
-    dict of str to dict
-        Dictionary mapping subspecialty_id to prediction parameters:
-
-        - 'pmf_ed_current_within_window': numpy.ndarray
-          Probability mass function for current ED admissions within the prediction window
-        - 'pmf_inpatient_departures_within_window': numpy.ndarray
-          Probability mass function for current inpatient departures within the prediction window
-        - 'lambda_ed_yta_within_window': float
[…]
