### 2025-10-09 a78d14f
make SubspecialtyPredictionInputs dataclass
