+++ src/patientflow/predict/__init__.py
@@ -6,6 +6,7 @@ including emergency demand forecasting and other predictive analytics.
 
 from patientflow.predict.subspecialty import (
     build_subspecialty_data,
+    SubspecialtyPredictionInputs,
     compute_transfer_arrivals,
     scale_pmf_by_probability,
     convolve_pmfs,
@@ -13,6 +14,7 @@ from patientflow.predict.subspecialty import (
 
 __all__ = [
     "build_subspecialty_data",
+    "SubspecialtyPredictionInputs",
     "compute_transfer_arrivals",
     "scale_pmf_by_probability",
     "convolve_pmfs",
