### 2025-10-09 84ba224
make SubspecialtyPredictionInputs dataclass
