### 2025-10-09 df918b0
simplify set up of hospital hierarchy
