### 2025-10-09 ea5184b
prettify prediction outputs
