### 2025-10-09 d22a101
prettify prediction outputs
