### 2025-10-09 5e5c44a
choose pmf output range dynamically
