### 2025-10-09 2e2e615
index pmf in pretty output
