### 2025-10-09 5289991
add transfer predictor to build_subspecialty_data
