### 2025-10-09 977c4c2
add transfer predictor to build_subspecialty_data
