+++ src/patientflow/predict/subspecialty.py
@@ -7,7 +7,7 @@ outputs and current patient snapshots into:
 - A probability mass function (PMF) for admissions from current ED patients
 - A probability mass function (PMF) for departures from current inpatients
 - Poisson means for yet-to-arrive admissions (ED, non-ED emergency, elective)
-- Transfer arrival distributions from internal subspecialty movements
+- A probability mass function (PMF) for transfer arrivals from other subspecialties
 
 The outputs are independent per subspecialty and do not presuppose any
 particular consolidation hierarchy. They can be used directly for single-
@@ -49,6 +49,9 @@ from patientflow.predictors.value_to_outcome_predictor import (
 from patientflow.predictors.subgroup_predictor import (
     MultiSubgroupPredictor,
 )
+from patientflow.predictors.transfer_predictor import (
+    TransferProbabilityEstimator,
+)
 from patientflow.aggregate import (
     model_input_to_pred_proba,
     pred_proba_to_agg_predicted,
@@ -73,9 +76,6 @@ class SubspecialtyPredictionInputs:
     pmf_ed_current_within_window : np.ndarray
         Probability mass function for current ED admissions within the prediction window.
         Array where pmf[k] represents P(k ED patients admitted to this subspecialty).
-    pmf_inpatient_departures_within_window : np.ndarray
-        Probability mass function for current inpatient departures within the prediction window.
-        Array where pmf[k] represents P(k current inpatients depart from this subspecialty).
     lambda_ed_yta_within_window : float
         Poisson parameter (rate) for ED yet-to-arrive admissions within the prediction window.
         Represents the expected number of new ED arrivals who will be admitted to this subspecialty.
@@ -85,6 +85,12 @@ class SubspecialtyPredictionInputs:
     lambda_elective_yta_within_window : float
         Poisson parameter (rate) for elective admissions within the prediction window.
         Represents the expected number of planned/elective admissions to this subspecialty.
+    pmf_transfer_arrivals_within_window : np.ndarray
+        Probability mass function for transfer arrivals from other subspecialties within the prediction window.
+        Array where pmf[k] represents P(k patients transfer into this subspecialty from other subspecialties).
+    pmf_inpatient_departures_within_window : np.ndarray
+        Probability mass function for current inpatient departures within the prediction window.
+        Array where pmf[k] represents P(k current inpatients depart from this subspecialty).
 
     Notes
     -----
@@ -93,10 +99,11 @@ class SubspecialtyPredictionInputs:
     """
 
     pmf_ed_current_within_window: np.ndarray
-    pmf_inpatient_departures_within_window: np.ndarray
     lambda_ed_yta_within_window: float
     lambda_non_ed_yta_within_window: float
     lambda_elective_yta_within_window: float
+    pmf_transfer_arrivals_within_window: np.ndarray
+    pmf_inpatient_departures_within_window: np.ndarray
 
     def __repr__(self) -> str:
         """Return a clean, readable representation showing PMF values and lambdas."""
@@ -133,15 +140,17 @@ class SubspecialtyPredictionInputs:
                 return f"PMF[{start_idx}:{end_idx}]: [{display_values}]{suffix} (mode@{mode_idx})"
 
         ed_pmf_str = format_pmf(self.pmf_ed_current_within_window)
+        transfer_pmf_str = format_pmf(self.pmf_transfer_arrivals_within_window)
         inpt_pmf_str = format_pmf(self.pmf_inpatient_departures_within_window)
 
         return (
             f"SubspecialtyPredictionInputs(\n"
-            f"  ED current        {ed_pmf_str}\n"
-            f"  Inpatient depart  {inpt_pmf_str}\n"
[…]
