### 2025-10-10 3e48ed9
option to return transition matrix
