### 2025-10-10 e5eb982
option to return transition matrix
