### 2025-10-11 bc5023a
show expectation with all pmfs
