### 2025-10-13 d91ed66
show length of all pmfs
