### 2025-10-13 91cb94a
validate subspecialties when fitting transfer predictor
