### 2025-10-13 6a98372
validate subspecialties when fitting transfer predictor
