+++ src/patientflow/predict/subspecialty.py
@@ -1,27 +1,15 @@
 """Demand preparation utilities for later consolidation.
 
-This module prepares per-subspecialty inputs that can be consumed by any
-downstream roll-up or forecasting workflow. It converts trained model
-outputs and current patient snapshots into:
-
-- A probability mass function (PMF) for admissions from current ED patients
-- A probability mass function (PMF) for departures from current inpatients
-- Poisson means for yet-to-arrive admissions (ED, non-ED emergency, elective)
-- A probability mass function (PMF) for transfer arrivals from other subspecialties
+This module prepares per-subspecialty demand inputs using a flexible 
+architecture. It converts trained model outputs and current patient snapshots
+into structured representations organised by direction (inflows/outflows).
+The flow-based structure allows flexible selection of which flows to include
+in predictions and easy extension to new flow types in the future.
 
 The outputs are independent per subspecialty and do not presuppose any
 particular consolidation hierarchy. They can be used directly for single-
 specialty analyses or fed into any combination scheme (including hierarchical
 schemes) implemented elsewhere.
-
-Notes
------
-This module integrates with the broader patientflow ecosystem by:
-
-1. Using trained classifiers and specialty models from the train module
-2. Processing patient snapshots from the prepare module
-3. Computing admission probabilities using the calculate module
-4. Modelling internal patient transfers using the transfer predictor
 
 """
 
@@ -64,6 +52,63 @@ from patientflow.model_artifacts import TrainedClassifier
 
 
 @dataclass(frozen=True)
+class FlowInputs:
+    """Represents a single source of patient flow.
+
+    This class encapsulates a flow of patients (either arriving or departing)
+    with its distribution type and parameters. It provides a uniform interface
+    for both probability mass functions and Poisson-distributed flows.
+
+    Attributes
+    ----------
+    flow_id : str
+        Unique identifier for this flow (e.g., "ed_current", "transfers_in")
+    flow_type : str
+        Type of distribution: "pmf" for probability mass function or "poisson" for Poisson
+    distribution : np.ndarray or float
+        For "pmf": numpy array where distribution[k] = P(k patients)
+        For "poisson": float representing the Poisson rate parameter (lambda)
+    display_name : str, optional
+        Human-readable name for display purposes. If not provided, flow_id will be
+        formatted automatically (underscores replaced with spaces, title cased).
+
+    Examples
+    --------
+    >>> # PMF flow (e.g., current ED patients)
+    >>> ed_flow = FlowInputs(
+    ...     flow_id="ed_current",
+    ...     flow_type="pmf",
+    ...     distribution=np.array([0.5, 0.3, 0.2]),
+    ...     display_name="Admissions from current ED"
+    ... )
+
+    >>> # Poisson flow (e.g., yet-to-arrive patients)
+    >>> yta_flow = FlowInputs(
+    ...     flow_id="ed_yta",
+    ...     flow_type="poisson",
+    ...     distribution=2.5,
+    ...     display_name="ED yet-to-arrive admissions"
+    ... )
+    """
+    flow_id: str
+    flow_type: str
+    distribution: Union[np.ndarray, float]
+    display_name: Optional[str] = None
+
+    def get_display_name(self) -> str:
+        """Get human-readable display name.
+
+        Returns
+        -------
+        str
+            Display name if provided, otherwise formatted flow_id.
+        """
+        if self.display_name:
+            return self.display_name
+        return self.flow_id.replace("_", " ").title()
+
+
+@dataclass(frozen=True)
 class SubspecialtyPredictionInputs:
     """Input parameters for subspecialty demand prediction.
 
@@ -71,43 +116,49 @@ class SubspecialtyPredictionInputs:
[…]
