### 2025-10-13 9494492
implement flow-based architecture for subspecialty demand
replace flat attribute structure in SubspecialtyPredictionInputs with
flexible approach using inflows/outflows dictionaries

Key changes:
- Add FlowInputs dataclass to represent individual patient flows
- Organize flows by direction (inflows: arrivals, outflows: departures)
- Update build_subspecialty_data() to construct new structure
- Add FlowSelection for customisable flow inclusion
- Add PredictionBundle to track arrivals, departures, and net flow
- Update DemandPredictor.predict_subspecialty() to return bundle
- Add predict_flow_to
[…]
