+++ src/patientflow/predict/hierarchy.py
@@ -23,11 +23,11 @@ create_hierarchical_predictor
 
 import numpy as np
 import pandas as pd
-from typing import List, Dict, Optional
+from typing import List, Dict, Optional, Any
 from dataclasses import dataclass
 from scipy.stats import poisson
 
-from patientflow.predict.subspecialty import SubspecialtyPredictionInputs
+from patientflow.predict.subspecialty import SubspecialtyPredictionInputs, FlowInputs
 
 
 @dataclass
@@ -139,6 +139,161 @@ class DemandPrediction:
         return self.to_pretty()
 
 
+@dataclass
+class FlowSelection:
+    """Configuration for which flows to include in predictions.
+    
+    This class controls which patient flows (arrivals and departures) are included
+    when making predictions. It allows users to customize predictions based on their
+    specific use cases, such as excluding transfers or focusing only on certain types
+    of admissions.
+    
+    Attributes
+    ----------
+    inflow_keys : List[str]
+        List of inflow identifiers to include in predictions.
+        Standard values: ["ed_current", "ed_yta", "non_ed_yta", "elective_yta", "transfers_in"]
+    outflow_keys : List[str]
+        List of outflow identifiers to include in predictions.
+        Standard values: ["departures"]
+        Future extensions may include: ["transfers_out", "deaths"]
+    
+    Examples
+    --------
+    >>> # Include all flows
+    >>> selection = FlowSelection.default()
+    
+    >>> # Exclude transfers
+    >>> selection = FlowSelection.admissions_only()
+    
+    >>> # Custom selection
+    >>> selection = FlowSelection.custom(
+    ...     include_inflows=["ed_current", "ed_yta"],
+    ...     include_outflows=["departures"]
+    ... )
+    """
+    inflow_keys: List[str]
+    outflow_keys: List[str]
+    
+    @classmethod
+    def default(cls) -> 'FlowSelection':
+        """All flows included (default behavior).
+        
+        Returns
+        -------
+        FlowSelection
+            Configuration including all standard inflows and outflows
+        """
+        return cls(
+            inflow_keys=["ed_current", "ed_yta", "non_ed_yta", "elective_yta", "transfers_in"],
+            outflow_keys=["departures"]
+        )
+    
+    @classmethod
+    def admissions_only(cls) -> 'FlowSelection':
+        """Only admission flows, excluding transfers.
+        
+        This is useful for analyzing demand independent of internal transfers,
+        focusing only on new patients entering the system.
+        
+        Returns
+        -------
+        FlowSelection
+            Configuration excluding transfer-related flows
+        """
+        return cls(
+            inflow_keys=["ed_current", "ed_yta", "non_ed_yta", "elective_yta"],
+            outflow_keys=["departures"]
+        )
+    
+    @classmethod
+    def custom(cls, include_inflows: List[str], include_outflows: List[str]) -> 'FlowSelection':
+        """Custom flow selection.
+        
+        Parameters
+        ----------
+        include_inflows : List[str]
+            List of inflow identifiers to include
+        include_outflows : List[str]
+            List of outflow identifiers to include
+        
+        Returns
+        -------
+        FlowSelection
+            Custom configuration with specified flows
+        """
+        return cls(inflow_keys=include_inflows, outflow_keys=include_outflows)
+
+
+@dataclass
+class PredictionBundle:
+    """Complete prediction results for arrivals, departures, and net flow.
+    
+    This dataclass bundles together predictions for patient arrivals, departures,
+    and the net change in bed occupancy for a single organizational entity.
+    It provides a comprehensive view of demand dynamics.
+    
+    Attributes
+    ----------
+    entity_id : str
[…]
