### 2025-10-13 6e437d2
update HierarchicalPredictor to track arrivals and departures separately
- Change cache and return types from DemandPrediction to PredictionBundle
- Add flow_selection parameter to predict_all_levels()
- Aggregate arrivals and departures separately at each hierarchical level
- Compute net flow as difference of expectations at each level
- Update docstrings to reflect bundle-based predictions
