+++ src/patientflow/predict/hierarchy.py
@@ -985,8 +985,9 @@ class HierarchicalPredictor:
     levels of the hospital hierarchy. It orchestrates the bottom-up prediction
     process, starting from subspecialties and aggregating up to the hospital level.
 
-    The class maintains a cache of predictions for efficient retrieval and
-    provides methods to compute predictions for all levels at once.
+    The class maintains a cache of prediction bundles for efficient retrieval and
+    provides methods to compute predictions for all levels at once. Each bundle
+    tracks arrivals, departures, and net flow separately.
 
     Parameters
     ----------
@@ -1002,30 +1003,33 @@ class HierarchicalPredictor:
     predictor : DemandPredictor
         Core prediction engine
     cache : dict
-        Cache for storing computed predictions
+        Cache for storing computed prediction bundles (PredictionBundle objects)
 
     Notes
     -----
     Predictions are computed bottom-up: subspecialties -> reporting units ->
-    divisions -> boards -> hospital. Each level aggregates predictions from
-    its children using convolution.
+    divisions -> boards -> hospital. At each level, arrivals and departures
+    are aggregated separately using convolution, and net flows are computed.
     """
 
     def __init__(self, hierarchy: HospitalHierarchy, predictor: DemandPredictor):
         self.hierarchy = hierarchy
         self.predictor = predictor
-        self.cache: Dict[str, DemandPrediction] = {}
+        self.cache: Dict[str, PredictionBundle] = {}
 
     def predict_all_levels(
         self,
         subspecialty_data: Dict[str, SubspecialtyPredictionInputs],
         hospital_id: Optional[str] = None,
-    ) -> Dict[str, DemandPrediction]:
+        flow_selection: Optional[FlowSelection] = None,
+    ) -> Dict[str, PredictionBundle]:
         """Compute predictions for all levels bottom-up.
 
         This method orchestrates the complete hierarchical prediction process,
         starting from subspecialties and aggregating predictions up through
         reporting units, divisions, boards, and finally to the hospital level.
+        
+        For each level, tracks arrivals, departures, and net flow separately.
 
         Parameters
         ----------
@@ -1036,12 +1040,16 @@ class HierarchicalPredictor:
             Unique identifier for the hospital. If not provided and the hierarchy
             contains exactly one hospital, that hospital will be used automatically.
             Required if the hierarchy contains multiple hospitals.
+        flow_selection : FlowSelection, optional
+            Configuration for which flows to include. If None, uses FlowSelection.default()
+            which includes all flows.
 
         Returns
         -------
-        dict[str, DemandPrediction]
-            Dictionary mapping entity_id to DemandPrediction for all levels:
-            subspecialties, reporting units, divisions, boards, and hospital
+        dict[str, PredictionBundle]
+            Dictionary mapping entity_id to PredictionBundle for all levels:
+            subspecialties, reporting units, divisions, boards, and hospital.
+            Each bundle contains arrivals, departures, and net flow predictions.
 
         Raises
         ------
@@ -1051,11 +1059,11 @@ class HierarchicalPredictor:
         Notes
         -----
         The prediction process follows this sequence:
-        1. Predict subspecialties using provided parameters
-        2. Aggregate subspecialties into reporting units
-        3. Aggregate reporting units into divisions
-        4. Aggregate divisions into boards
-        5. Aggregate boards into hospital
+        1. Predict subspecialties using provided parameters (returns bundles)
+        2. Aggregate subspecialty arrivals/departures into reporting unit bundles
+        3. Aggregate reporting unit arrivals/departures into division bundles
+        4. Aggregate division arrivals/departures into board bundles
+        5. Aggregate board arrivals/departures into hospital bundle
 
[…]
