### 2025-10-14 7a0a13e
add flow_selection to PredictionBundle
