+++ FLOW_STRUCTURE_DOCUMENTATION.md
@@ -1,0 +1,263 @@
+# Patient Flow Structure Documentation
+
+## Overview
+
+This document describes the refactored flow-based architecture for subspecialty demand prediction in the patientflow package. The new structure provides a flexible, extensible way to model patient arrivals and departures.
+
+## Key Components
+
+### 1. `FlowInputs` Dataclass
+
+Represents a single source of patient flow (either arriving or departing).
+
+**Attributes:**
+- `flow_id` (str): Unique identifier (e.g., "ed_current", "transfers_in", "departures")
+- `flow_type` (str): Either "pmf" (probability mass function) or "poisson" (Poisson distribution)
+- `distribution` (np.ndarray | float): PMF array or Poisson lambda parameter
+- `display_name` (str, optional): Human-readable name for display
+
+**Standard Flow IDs:**
+
+**Inflows (arrivals):**
+- `"ed_current"`: Current ED patients who will be admitted (PMF)
+- `"ed_yta"`: Yet-to-arrive ED patients (Poisson)
+- `"non_ed_yta"`: Non-ED emergency admissions (Poisson)
+- `"elective_yta"`: Elective admissions (Poisson)
+- `"transfers_in"`: Patients transferring from other subspecialties (PMF)
+
+**Outflows (departures):**
+- `"departures"`: Current inpatients who will depart (PMF)
+- Future: `"transfers_out"`, `"deaths"`, etc.
+
+### 2. `SubspecialtyPredictionInputs` Dataclass
+
+Organizes all flows for a single subspecialty.
+
+**Attributes:**
+- `subspecialty_id` (str): Unique identifier
+- `prediction_window` (timedelta): Time window for predictions
+- `inflows` (Dict[str, FlowInputs]): Dictionary of arrival flows
+- `outflows` (Dict[str, FlowInputs]): Dictionary of departure flows
+
+### 3. `FlowSelection` Configuration
+
+Controls which flows to include in predictions.
+
+**Methods:**
+- `FlowSelection.default()`: All flows included
+- `FlowSelection.admissions_only()`: Excludes transfers
+- `FlowSelection.custom(inflow_keys, outflow_keys)`: Custom selection
+
+### 4. `PredictionBundle` Results
+
+Contains complete prediction results.
+
+**Attributes:**
+- `entity_id` (str): Subspecialty identifier
+- `entity_type` (str): Type ("subspecialty", "reporting_unit", etc.)
+- `arrivals` (DemandPrediction): Total arrivals distribution
+- `departures` (DemandPrediction): Total departures distribution
+- `net_flow_expected` (float): Expected net change (arrivals - departures)
+
+## Usage Examples
+
+### Example 1: Basic Usage with All Flows
+
+```python
+from patientflow.predict.subspecialty import build_subspecialty_data
+from patientflow.predict.hierarchy import DemandPredictor, FlowSelection
+
+# Build subspecialty inputs (automatically includes all flows)
+subspecialty_data = build_subspecialty_data(
+    models=models,
+    prediction_time=(7, 0),
+    ed_snapshots=ed_snapshots,
+    inpatient_snapshots=inpatient_snapshots,
+    specialties=["cardiology", "surgery", "medicine"],
+    prediction_window=timedelta(hours=8),
+    x1=0.0, y1=0.0, x2=8.0, y2=0.9
+)
+
+# Make prediction with all flows included
+predictor = DemandPredictor()
+bundle = predictor.predict_subspecialty(
+    subspecialty_id="cardiology",
+    inputs=subspecialty_data["cardiology"],
+    flow_selection=FlowSelection.default()  # Optional, this is the default
+)
+
+# Access results
[…]
