+++ src/patientflow/predict/hierarchy.py
@@ -50,6 +50,9 @@ class DemandPrediction:
         Expected number of beds needed (mean of the distribution)
     percentiles : dict[int, int]
         Dictionary mapping percentile values (50, 75, 90, 95, 99) to bed counts
+    offset : int, optional
+        Offset for the support of the distribution. Index i corresponds to value (i + offset).
+        Default is 0 for non-negative distributions. For net flow, offset is typically negative.
 
     Attributes
     ----------
@@ -63,13 +66,17 @@ class DemandPrediction:
         Expected number of beds needed
     percentiles : dict[int, int]
         Percentile values for bed demand
+    offset : int
+        Offset for the support (index 0 corresponds to value = offset)
     max_beds : int
         Maximum number of beds in the probability distribution
 
     Notes
     -----
     The probabilities array represents P(X=k) where X is the number of beds needed
-    and k ranges from 0 to max_beds. The sum of all probabilities should equal 1.0.
+    and k ranges from offset to (offset + max_beds). The sum of all probabilities should equal 1.0.
+    For non-negative distributions (arrivals, departures), offset=0.
+    For net flow distributions, offset can be negative (e.g., -10 means support starts at -10).
     """
 
     entity_id: str
@@ -77,6 +84,7 @@ class DemandPrediction:
     probabilities: np.ndarray
     expected_value: float
     percentiles: Dict[int, int]
+    offset: int = 0
 
     @property
     def max_beds(self) -> int:
@@ -160,11 +168,14 @@ class FlowSelection:
     
     Examples
     --------
-    >>> # Include all flows
+    >>> # Include all flows (arrivals and departures)
     >>> selection = FlowSelection.default()
     
-    >>> # Exclude transfers
-    >>> selection = FlowSelection.admissions_only()
+    >>> # Only incoming flows (all arrivals, no departures)
+    >>> selection = FlowSelection.incoming_flow()
+    
+    >>> # Only outgoing flows (all departures, no arrivals)
+    >>> selection = FlowSelection.outgoing_flow()
     
     >>> # Custom selection
     >>> selection = FlowSelection.custom(
@@ -190,19 +201,37 @@ class FlowSelection:
         )
     
     @classmethod
-    def admissions_only(cls) -> 'FlowSelection':
-        """Only admission flows, excluding transfers.
+    def incoming_flow(cls) -> 'FlowSelection':
+        """Select incoming patient flows only (all arrivals, no departures).
         
-        This is useful for analyzing demand independent of internal transfers,
-        focusing only on new patients entering the system.
+        Includes all inflow types (ed_current, ed_yta, non_ed_yta, 
+        elective_yta, transfers_in) but excludes all outflows.
+        Useful for analyzing arrival patterns independently of departures.
         
         Returns
         -------
         FlowSelection
-            Configuration excluding transfer-related flows
+            Configuration with all inflows and no outflows
         """
         return cls(
-            inflow_keys=["ed_current", "ed_yta", "non_ed_yta", "elective_yta"],
+            inflow_keys=["ed_current", "ed_yta", "non_ed_yta", "elective_yta", "transfers_in"],
+            outflow_keys=[]
+        )
+    
+    @classmethod
+    def outgoing_flow(cls) -> 'FlowSelection':
+        """Select outgoing patient flows only (all departures, no arrivals).
+        
+        Includes all outflow types but excludes all inflows.
+        Useful for analysing discharge patterns independently of arrivals.
+        
+        Returns
+        -------
+        FlowSelection
+            Configuration with no inflows and all outflows
+        """
+        return cls(
+            inflow_keys=[],
             outflow_keys=["departures"]
         )
     
@@ -243,8 +272,9 @@ class PredictionBundle:
         Prediction for total patient arrivals
     departures : DemandPrediction
[…]
