### 2025-10-14 97ef04e
implement full PMF for net flow and refactor FlowSelection API
- Add offset tracking to DemandPrediction for negative support
- Compute net flow as full probability distribution (not just expected value)
- Update _expected_value() and _percentiles() to handle offset
- Change PredictionBundle.net_flow from float to DemandPrediction
- Apply net flow PMF computation at all hierarchy levels
- Replace FlowSelection.admissions_only() with incoming_flow() and outgoing_flow()
- Add test_hierarchy.py
