### 2025-10-14 06eba73
change summary of PredictionBundle
