### 2025-10-14 1ebf477
change summary of PredictionBundle
