+++ tests/test_hierarchy.py
@@ -1,7 +1,6 @@
 """Tests for hierarchical demand prediction functionality."""
 
 import numpy as np
-import pytest
 
 from patientflow.predict.hierarchy import (
     DemandPredictor,
@@ -26,7 +25,7 @@ class TestDemandPrediction:
             percentiles={50: 0, 75: 1, 90: 2, 95: 2, 99: 2},
             offset=-2,  # Support starts at -2
         )
-        
+
         assert pred.offset == -2
         assert pred.expected_value == -0.5
 
@@ -37,12 +36,12 @@ class TestDemandPredictor:
     def test_convolution(self):
         """Test basic convolution of two distributions."""
         predictor = DemandPredictor()
-        
+
         p1 = np.array([0.5, 0.5])  # 0 or 1 with equal probability
         p2 = np.array([0.3, 0.7])  # 0 or 1 with different probabilities
-        
+
         result = predictor._convolve(p1, p2)
-        
+
         # Result should have support [0, 2]
         assert len(result) == 3
         assert np.allclose(result.sum(), 1.0)
@@ -52,44 +51,44 @@ class TestDemandPredictor:
     def test_expected_value_without_offset(self):
         """Test expected value calculation without offset."""
         predictor = DemandPredictor()
-        
+
         p = np.array([0.1, 0.2, 0.4, 0.2, 0.1])  # Mean should be 2.0
         expected = predictor._expected_value(p, offset=0)
-        
+
         assert np.isclose(expected, 2.0)
 
     def test_expected_value_with_offset(self):
         """Test expected value calculation with negative offset."""
         predictor = DemandPredictor()
-        
+
         # Same distribution but now support is [-2, -1, 0, 1, 2]
         p = np.array([0.1, 0.2, 0.4, 0.2, 0.1])
         expected = predictor._expected_value(p, offset=-2)
-        
+
         # Mean should be 0.0 (symmetric around 0)
         assert np.isclose(expected, 0.0)
 
     def test_percentiles_with_offset(self):
         """Test percentile calculation with negative offset."""
         predictor = DemandPredictor()
-        
+
         # Support: [-2, -1, 0, 1, 2]
         p = np.array([0.1, 0.2, 0.4, 0.2, 0.1])
         percentiles = predictor._percentiles(p, [50], offset=-2)
-        
+
         # Median should account for offset
         assert percentiles[50] == 0  # Index 2, offset -2 -> value 0
 
     def test_net_flow_pmf_deterministic(self):
         """Test net flow PMF with deterministic arrivals and departures."""
         predictor = DemandPredictor()
-        
+
         # Always 3 arrivals, always 2 departures -> net flow always 1
         p_arrivals = np.array([0.0, 0.0, 0.0, 1.0])
         p_departures = np.array([0.0, 0.0, 1.0])
-        
+
         p_net, offset = predictor._compute_net_flow_pmf(p_arrivals, p_departures)
-        
+
         # Should have a single peak at net flow = 1
         assert np.isclose(p_net.sum(), 1.0, atol=1e-6)
         mode_idx = np.argmax(p_net)
@@ -100,13 +99,13 @@ class TestDemandPredictor:
     def test_net_flow_pmf_negative_result(self):
         """Test net flow PMF when result is negative."""
         predictor = DemandPredictor()
-        
+
         # Always 1 arrival, always 3 departures -> net flow always -2
         p_arrivals = np.array([0.0, 1.0])
         p_departures = np.array([0.0, 0.0, 0.0, 1.0])
-        
+
         p_net, offset = predictor._compute_net_flow_pmf(p_arrivals, p_departures)
-        
+
         assert np.isclose(p_net.sum(), 1.0, atol=1e-6)
         mode_idx = np.argmax(p_net)
         mode_value = mode_idx + offset
[…]
