### 2025-10-14 20642fd
change summary of PredictionBundle
