+++ src/patientflow/predict/hierarchy.py
@@ -113,7 +113,7 @@ class DemandPrediction:
             Formatted summary string
         """
         probs = self.probabilities
-        
+
         # Find where the probability mass is concentrated
         if len(probs) <= max_probs:
             start_idx = 0
@@ -125,7 +125,7 @@ class DemandPrediction:
             end_idx = min(len(probs), start_idx + max_probs)
             if end_idx - start_idx < max_probs:
                 start_idx = max(0, end_idx - max_probs)
-        
+
         head = ", ".join([f"{p:.{precision}g}" for p in probs[start_idx:end_idx]])
         remaining = len(probs) - end_idx
         tail_note = f" … +{remaining} more" if remaining > 0 else ""
@@ -150,12 +150,12 @@ class DemandPrediction:
 @dataclass
 class FlowSelection:
     """Configuration for which flows to include in predictions.
-    
+
     This class controls which patient flows (arrivals and departures) are included
     when making predictions. It allows users to customize predictions based on their
     specific use cases, such as excluding transfers or focusing only on certain types
     of admissions.
-    
+
     Attributes
     ----------
     inflow_keys : List[str]
@@ -165,87 +165,99 @@ class FlowSelection:
         List of outflow identifiers to include in predictions.
         Standard values: ["departures"]
         Future extensions may include: ["transfers_out", "deaths"]
-    
+
     Examples
     --------
     >>> # Include all flows (arrivals and departures)
     >>> selection = FlowSelection.default()
-    
+
     >>> # Only incoming flows (all arrivals, no departures)
     >>> selection = FlowSelection.incoming_flow()
-    
+
     >>> # Only outgoing flows (all departures, no arrivals)
     >>> selection = FlowSelection.outgoing_flow()
-    
+
     >>> # Custom selection
     >>> selection = FlowSelection.custom(
     ...     include_inflows=["ed_current", "ed_yta"],
     ...     include_outflows=["departures"]
     ... )
     """
+
     inflow_keys: List[str]
     outflow_keys: List[str]
-    
+
     @classmethod
-    def default(cls) -> 'FlowSelection':
+    def default(cls) -> "FlowSelection":
         """All flows included (default behavior).
-        
+
         Returns
         -------
         FlowSelection
             Configuration including all standard inflows and outflows
         """
         return cls(
-            inflow_keys=["ed_current", "ed_yta", "non_ed_yta", "elective_yta", "transfers_in"],
-            outflow_keys=["departures"]
+            inflow_keys=[
+                "ed_current",
+                "ed_yta",
+                "non_ed_yta",
+                "elective_yta",
+                "transfers_in",
+            ],
+            outflow_keys=["departures"],
         )
-    
+
     @classmethod
-    def incoming_flow(cls) -> 'FlowSelection':
+    def incoming_flow(cls) -> "FlowSelection":
         """Select incoming patient flows only (all arrivals, no departures).
-        
-        Includes all inflow types (ed_current, ed_yta, non_ed_yta, 
+
+        Includes all inflow types (ed_current, ed_yta, non_ed_yta,
         elective_yta, transfers_in) but excludes all outflows.
         Useful for analyzing arrival patterns independently of departures.
-        
+
         Returns
         -------
         FlowSelection
             Configuration with all inflows and no outflows
         """
         return cls(
-            inflow_keys=["ed_current", "ed_yta", "non_ed_yta", "elective_yta", "transfers_in"],
-            outflow_keys=[]
+            inflow_keys=[
+                "ed_current",
+                "ed_yta",
+                "non_ed_yta",
+                "elective_yta",
+                "transfers_in",
+            ],
+            outflow_keys=[],
         )
-    
+
     @classmethod
-    def outgoing_flow(cls) -> 'FlowSelection':
+    def outgoing_flow(cls) -> "FlowSelection":
[…]
