### 2025-10-15 4309f0b
commit to save debugdebug_subspecialty_convolution.py
