+++ debug_subspecialty_convolution.py
@@ -1,0 +1,587 @@
+#!/usr/bin/env python3
+"""Debug script for checking subspecialty convolution logic step by step.
+
+This script loads saved SubspecialtyPredictionInputs for a specific subspecialty
+and walks through the convolution logic, showing intermediate results at each step
+up to calculating the net flow.
+
+Usage:
+    python debug_subspecialty_convolution.py <subspecialty_data_path> <check_spec>
+
+Arguments:
+    subspecialty_data_path: Path to the saved subspecialty_data dictionary (joblib/pickle)
+    check_spec: Name of the subspecialty to debug (e.g., 'Cardiology')
+
+Example:
+    python debug_subspecialty_convolution.py subspecialty_data.joblib "Cardiology"
+"""
+
+import sys
+import numpy as np
+import joblib
+from typing import Dict, List
+from dataclasses import dataclass
+
+# Import the classes from your patientflow package
+from patientflow.predict.subspecialty import SubspecialtyPredictionInputs, FlowInputs
+from patientflow.predict.hierarchy import (
+    DemandPredictor,
+    DemandPrediction,
+    FlowSelection,
+    PredictionBundle,
+)
+
+
+def print_separator(title: str = "", char: str = "=", width: int = 80):
+    """Print a formatted separator line."""
+    if title:
+        title_str = f" {title} "
+        padding = (width - len(title_str)) // 2
+        print(f"\n{char * padding}{title_str}{char * padding}")
+    else:
+        print(f"\n{char * width}")
+
+
+def print_pmf_summary(pmf: np.ndarray, name: str, offset: int = 0, max_display: int = 15):
+    """Print a summary of a PMF with key statistics."""
+    expectation = np.sum((np.arange(len(pmf)) + offset) * pmf)
+    
+    # Find mode
+    mode_idx = np.argmax(pmf)
+    mode_value = mode_idx + offset
+    
+    # Calculate percentiles
+    cumsum = np.cumsum(pmf)
+    p50_idx = np.searchsorted(cumsum, 0.50)
+    p75_idx = np.searchsorted(cumsum, 0.75)
+    p95_idx = np.searchsorted(cumsum, 0.95)
+    
+    print(f"\n{name}:")
+    print(f"  Length: {len(pmf)} values (support: {offset} to {len(pmf) - 1 + offset})")
+    print(f"  Expectation (mean): {expectation:.3f}")
+    print(f"  Mode: {mode_value} (probability: {pmf[mode_idx]:.4f})")
+    print(f"  Percentiles: P50={p50_idx + offset}, P75={p75_idx + offset}, P95={p95_idx + offset}")
+    print(f"  Sum of probabilities: {np.sum(pmf):.6f}")
+    
+    # Display the PMF values (centered around expectation)
+    if len(pmf) <= max_display:
+        display_values = ", ".join(f"{v:.4f}" for v in pmf)
+        print(f"  PMF: [{display_values}]")
+    else:
+        center_idx = int(np.round(expectation - offset))
+        half_window = max_display // 2
+        start_idx = max(0, center_idx - half_window)
+        end_idx = min(len(pmf), start_idx + max_display)
+        
+        if end_idx - start_idx < max_display:
+            start_idx = max(0, end_idx - max_display)
+        
+        display_values = ", ".join(f"{v:.4f}" for v in pmf[start_idx:end_idx])
+        start_val = start_idx + offset
+        end_val = end_idx - 1 + offset
+        print(f"  PMF[{start_val}:{end_val}]: [{display_values}]")
+        
+        # Show where the mass is concentrated
+        top_5_indices = np.argsort(pmf)[-5:][::-1]
+        print(f"  Top 5 probabilities:")
+        for idx in top_5_indices:
+            val = idx + offset
[…]
