+++ DEBUG_CONVOLUTION_GUIDE.md
@@ -1,0 +1,326 @@
+# Subspecialty Convolution Debug Guide
+
+This guide explains how to debug the convolution logic for subspecialty predictions.
+
+## Quick Start
+
+### Method 1: Using the standalone script (for remote environments)
+
+```bash
+# Run the debug script on saved data
+python debug_subspecialty_convolution.py <path_to_saved_data> "<subspecialty_name>"
+
+# Example:
+python debug_subspecialty_convolution.py subspecialty_data.joblib "Cardiology"
+
+# With custom epsilon:
+python debug_subspecialty_convolution.py subspecialty_data.joblib "Cardiology" 1e-6
+```
+
+### Method 2: Using the inline debug function (for interactive sessions)
+
+```python
+from debug_subspecialty_inline import debug_subspecialty_convolution
+
+# Assuming you have subspecialty_data dict
+debug_subspecialty_convolution(subspecialty_data, "Cardiology")
+```
+
+## What the Debug Script Shows
+
+The debug script walks through the complete convolution pipeline for a subspecialty and **compares the DemandPredictor's internal methods with manual calculations** to verify correctness:
+
+### 1. **Flow Inputs Summary**
+   - Lists all inflow sources (ED current, ED YTA, non-ED YTA, elective, transfers)
+   - Lists all outflow sources (departures, etc.)
+   - Shows distribution type (PMF or Poisson) and parameters for each
+   - Displays expected values
+
+### 2. **Arrivals Convolution (Step-by-Step with Verification)**
+   For each inflow:
+   - **Poisson Generation** (if applicable):
+     - Compares `predictor._calculate_poisson_max()` with manual calculation
+     - Compares `predictor._poisson_pmf()` with manual Poisson PMF
+     - Shows ✓ or ✗ for agreement
+   - **Convolution**:
+     - Compares `predictor._convolve()` with manual `np.convolve()`
+     - Verifies expectation addition: E[A+B] = E[A] + E[B]
+     - Shows array comparison results
+   - **Truncation**:
+     - Compares `predictor._truncate()` with manual truncation
+     - Shows how many values were removed
+     - Verifies expectation is preserved
+   - **Running verification**: Each step confirms predictor matches manual calculation
+
+### 3. **Departures Convolution (Step-by-Step with Verification)**
+   For each outflow:
+   - Same detailed verification as arrivals
+   - Independent manual calculation tracks alongside predictor
+
+### 4. **Net Flow Calculation (with Verification)**
+   - Compares `predictor._compute_net_flow_pmf()` with manual nested-loop computation
+   - Verifies offset calculation for negative support
+   - Compares truncation from both ends
+   - Shows ✓ or ✗ for each comparison
+   - Verifies that E[Net Flow] = E[Arrivals] - E[Departures]
+
+### 5. **Final Summary**
+   - Expected values for arrivals, departures, and net flow
+   - Percentiles (P50, P75, P90, P95, P99) for each distribution
+   - Final verification that all expectations match
+
+## Understanding the Output
+
+### Verification Symbols
+
+The script uses these symbols to indicate agreement between methods:
+- **✓** Green checkmark: Predictor and manual calculations match (within tolerance)
+- **✗** Red X: Calculations differ (investigation needed)
+- **⚠️** Warning: Different array lengths or other structural differences
+
+### Comparison Output Format
+
+When comparing arrays:
+```
+  ✓ Convolution result: Arrays match (within tolerance 1e-10)
+```
+or
+```
[…]
