### 2025-10-15 8384b96
implement capped truncation
