### 2025-10-15 dd85ac0
implement capped truncation
