+++ src/patientflow/predict/hierarchy.py
@@ -384,14 +384,16 @@ class DemandPredictor:
 
     Parameters
     ----------
-    epsilon : float, default=1e-7
-        Truncation threshold for probability distribution tails. Distributions
-        are truncated when the cumulative probability exceeds (1 - epsilon).
+    k_sigma : float, default=4.0
+        Cap width measured in standard deviations. Final (and intermediate)
+        distributions are hard-clipped to mean + k_sigma * std for non-negative
+        support. Net-flow uses asymmetric caps around the mean with the same
+        k_sigma multiplier and physical bounds.
 
     Attributes
     ----------
-    epsilon : float
-        Truncation threshold for probability distributions
+    k_sigma : float
+        Number of standard deviations used to cap supports
     cache : dict
         Cache for storing computed predictions (currently unused)
 
@@ -405,12 +407,12 @@ class DemandPredictor:
     4. Computing statistics (expected value, percentiles) for each level
 
     The class uses discrete convolution to combine probability distributions.
-    For computational efficiency, distributions are periodically truncated during
-    multiple convolutions to prevent exponential growth in array sizes.
+    Supports are clamped deterministically using k-sigma caps to prevent
+    exponential growth in array sizes.
     """
 
-    def __init__(self, epsilon: float = 1e-7):
-        self.epsilon = epsilon
+    def __init__(self, k_sigma: float = 4.0):
+        self.k_sigma = k_sigma
         self.cache: Dict[str, DemandPrediction] = {}
 
     def predict_flow_total(
@@ -443,28 +445,49 @@ class DemandPredictor:
         Notes
         -----
         Flows are combined through convolution, which represents the distribution
-        of the sum of independent random variables. Periodic truncation is applied
-        to maintain computational efficiency.
+        of the sum of independent random variables. Supports are clamped using a
+        k-sigma cap to maintain computational efficiency.
         """
-        # Start with degenerate distribution at 0
-        p_total = np.array([1.0])
-
+        # First pass: compute per-flow means and variances only
+        means: List[float] = []
+        variances: List[float] = []
+        flow_specs: List[tuple[str, np.ndarray | float]] = []
         for flow in flow_inputs:
             if flow.flow_type == "poisson":
-                # Generate Poisson PMF
-                max_k = self._calculate_poisson_max(flow.distribution)
-                p_flow = self._poisson_pmf(flow.distribution, max_k)
+                lam = float(flow.distribution)
+                means.append(lam)
+                variances.append(lam)
+                flow_specs.append(("poisson", lam))
             elif flow.flow_type == "pmf":
-                # Use PMF directly
                 p_flow = flow.distribution
+                means.append(self._expected_value(p_flow))
+                variances.append(self._variance(p_flow))
+                flow_specs.append(("pmf", p_flow))
             else:
                 raise ValueError(
                     f"Unknown flow type: {flow.flow_type}. Expected 'pmf' or 'poisson'."
                 )
 
-            # Convolve this flow with running total
+        # Global cap for the total non-negative support from combined mean/std
+        total_mean = float(np.sum(means)) if means else 0.0
+        total_std = float(np.sqrt(np.sum(variances))) if variances else 0.0
+        cap_max = max(0, int(np.floor(total_mean + self.k_sigma * total_std)))
+
+        # Second pass: materialize flows, using the global cap for Poisson
+        prepared_flows: List[np.ndarray] = []
+        for flow_type, distribution_data in flow_specs:
+            if flow_type == "poisson":
+                lam = float(distribution_data)
[…]
