### 2025-10-20 3bec6e5
create a Distribution object
