+++ src/patientflow/predict/distribution.py
@@ -1,0 +1,222 @@
+"""Composable distribution algebra used for demand prediction.
+
+This module introduces a minimal, immutable Distribution abstraction that wraps
+discrete probability mass functions (PMFs) with an optional integer offset for
+their support. It provides algebraic operations used throughout demand
+prediction while mirroring the numerical behavior of the existing
+implementation (k-sigma clamping and Poisson truncation).
+
+Scope: Phase 1 only – no changes to existing call sites. This module is
+standalone and does not modify current behavior elsewhere.
+"""
+
+from dataclasses import dataclass
+from typing import Iterable, List, Dict
+
+import numpy as np
+from scipy.stats import poisson as _poisson
+from scipy.stats import binom as _binom
+
+
+def _validate_pmf(p: np.ndarray) -> None:
+    if p.ndim != 1:
+        raise ValueError("PMF must be a 1D array")
+    if p.size == 0:
+        raise ValueError("PMF must not be empty")
+    if np.any(p < 0):
+        raise ValueError("PMF must be non-negative")
+    if np.any(~np.isfinite(p)):
+        raise ValueError("PMF must be finite (no NaN or inf)")
+
+
+@dataclass(frozen=True)
+class Distribution:
+    """Immutable discrete distribution with optional integer offset.
+
+    The PMF is defined over integer support {offset, offset+1, ..., offset+N}.
+    """
+
+    probabilities: np.ndarray
+    offset: int = 0
+
+    # ---- Constructors ----
+    @classmethod
+    def from_pmf(cls, pmf: np.ndarray, offset: int = 0) -> "Distribution":
+        p = np.asarray(pmf, dtype=float).copy()
+        _validate_pmf(p)
+        return cls(probabilities=p, offset=int(offset))
+
+    @classmethod
+    def from_poisson_with_cap(
+        cls, lambda_param: float, max_k: int
+    ) -> "Distribution":
+        if lambda_param < 0:
+            raise ValueError("Poisson lambda must be non-negative")
+        if lambda_param == 0.0:
+            return cls.from_pmf(np.array([1.0]))
+        k = np.arange(int(max_k) + 1)
+        p = _poisson.pmf(k, float(lambda_param))
+        return cls.from_pmf(p)
+
+    @staticmethod
+    def poisson_cap_from_k_sigma(lambda_param: float, k_sigma: float) -> int:
+        if lambda_param <= 0:
+            return 0
+        return int(np.ceil(lambda_param + k_sigma * np.sqrt(lambda_param)))
+
+    # ---- Basic properties ----
+    def __len__(self) -> int:
+        return self.probabilities.shape[0]
+
+    # ---- Algebraic operations (pure) ----
+    def convolve(self, other: "Distribution") -> "Distribution":
+        """Convolve two distributions (sum of independent integer-valued RVs)."""
+        p = np.convolve(self.probabilities, other.probabilities)
+        new_offset = self.offset + other.offset
+        return Distribution.from_pmf(p, new_offset)
+
+    def clamp_nonnegative(self, cap_max: int) -> "Distribution":
+        """Clamp to non-negative support [offset, offset+cap_max] if offset>=0.
+
+        For the common case where offset==0, this mirrors the current
+        implementation's deterministic cap to limit array growth.
+        """
+        if cap_max < 0:
+            # retain only the first mass; ensures non-empty result
+            return Distribution.from_pmf(self.probabilities[:1], self.offset)
+        max_len = cap_max + 1
+        if len(self) <= max_len:
+            return self
+        return Distribution.from_pmf(self.probabilities[:max_len], self.offset)
+
+    def renorm(self) -> "Distribution":
[…]
