### 2025-10-20 484390d
create a Distribution object
