+++ src/patientflow/predict/hierarchy.py
@@ -25,9 +25,10 @@ import numpy as np
 import pandas as pd
 from typing import List, Dict, Optional, Any
 from dataclasses import dataclass
-from scipy.stats import poisson
+ 
 
 from patientflow.predict.subspecialty import SubspecialtyPredictionInputs, FlowInputs
+from patientflow.predict.distribution import Distribution
 
 
 @dataclass
@@ -473,23 +474,17 @@ class DemandPredictor:
         total_std = float(np.sqrt(np.sum(variances))) if variances else 0.0
         cap_max = max(0, int(np.floor(total_mean + self.k_sigma * total_std)))
 
-        # Second pass: materialize flows, using the global cap for Poisson
-        prepared_flows: List[np.ndarray] = []
+        # Second pass: materialize flows via Distribution using the global cap for Poisson
+        dist_total = Distribution.from_pmf(np.array([1.0]))
         for flow_type, distribution_data in flow_specs:
             if flow_type == "poisson":
                 lam = float(distribution_data)
-                p_flow = self._poisson_pmf(lam, cap_max)
+                flow_dist = Distribution.from_poisson_with_cap(lam, cap_max)
             else:  # pmf
-                p_flow = distribution_data  # type: ignore[assignment]
-            prepared_flows.append(p_flow)
+                flow_dist = Distribution.from_pmf(distribution_data)  # type: ignore[arg-type]
+            dist_total = dist_total.convolve(flow_dist).clamp_nonnegative(cap_max)
 
-        # Start with degenerate distribution at 0; convolve and clamp to global cap
-        p_total = np.array([1.0])
-        for p_flow in prepared_flows:
-            p_total = self._convolve(p_total, p_flow)
-            p_total = self._clamp_nonnegative(p_total, cap_max)
-
-        return self._create_prediction(entity_id, entity_type, p_total)
+        return self._create_prediction(entity_id, entity_type, dist_total.probabilities)
 
     def predict_subspecialty(
         self,
@@ -573,14 +568,14 @@ class DemandPredictor:
         arrivals_p = self._renormalize(arrivals.probabilities)
         departures_p = self._renormalize(departures.probabilities)
 
-        # Compute net flow distribution
-        p_net_flow, net_offset = self._compute_net_flow_pmf(
-            arrivals_p, departures_p
+        # Compute net flow distribution using Distribution algebra
+        net_dist = Distribution.from_pmf(arrivals_p).net(
+            Distribution.from_pmf(departures_p), self.k_sigma
         )
         # Ensure expected value matches arrivals - departures exactly for tests
         expected_diff = float(arrivals.expected_value - departures.expected_value)
         net_flow = self._create_prediction(
-            subspecialty_id, "net_flow", p_net_flow, net_offset, expected_override=expected_diff
+            subspecialty_id, "net_flow", net_dist.probabilities, net_dist.offset, expected_override=expected_diff
         )
 
         return PredictionBundle(
@@ -773,38 +768,7 @@ class DemandPredictor:
             return p
         return p[:max_len]
 
-    def _calculate_poisson_max(self, lambda_param: float) -> int:
-        """Compute k-sigma cap for a Poisson distribution."""
-        if lambda_param <= 0:
-            return 0
-        return int(np.ceil(lambda_param + self.k_sigma * np.sqrt(lambda_param)))
-
-    def _poisson_pmf(self, lambda_param: float, max_k: int) -> np.ndarray:
-        """Generate Poisson PMF from 0 to max_k.
-
-        This method generates the probability mass function for a Poisson
-        distribution truncated at max_k.
-
-        Parameters
-        ----------
-        lambda_param : float
-            Poisson parameter (mean and variance)
-        max_k : int
-            Maximum value to include in the PMF
-
[…]
