### 2025-10-20 136bb78
tidy docstrings
