### 2025-10-20 aba97d0
tidy docstrings
