+++ src/patientflow/predict/distribution.py
@@ -1,17 +1,13 @@
 """Composable distribution algebra used for demand prediction.
 
-This module introduces a minimal, immutable Distribution abstraction that wraps
-discrete probability mass functions (PMFs) with an optional integer offset for
-their support. It provides algebraic operations used throughout demand
-prediction while mirroring the numerical behavior of the existing
-implementation (k-sigma clamping and Poisson truncation).
-
-Scope: Phase 1 only – no changes to existing call sites. This module is
-standalone and does not modify current behavior elsewhere.
+Provides an immutable ``Distribution`` abstraction for discrete probability
+mass functions (PMFs) over integers, with an optional integer offset for the
+support. Includes common operations (convolution, thinning, clamping,
+renormalisation, difference) and summary statistics.
 """
 
 from dataclasses import dataclass
-from typing import Iterable, List, Dict
+from typing import Iterable, Dict
 
 import numpy as np
 from scipy.stats import poisson as _poisson
@@ -33,7 +29,14 @@ def _validate_pmf(p: np.ndarray) -> None:
 class Distribution:
     """Immutable discrete distribution with optional integer offset.
 
-    The PMF is defined over integer support {offset, offset+1, ..., offset+N}.
+    The PMF is defined over integer support ``{offset, offset+1, ..., offset+N}``.
+
+    Parameters
+    ----------
+    probabilities : numpy.ndarray
+        Probability mass function, non-negative and 1D.
+    offset : int, default=0
+        Integer value corresponding to index 0 of ``probabilities``.
     """
 
     probabilities: np.ndarray
@@ -42,14 +45,40 @@ class Distribution:
     # ---- Constructors ----
     @classmethod
     def from_pmf(cls, pmf: np.ndarray, offset: int = 0) -> "Distribution":
+        """Create a distribution from a PMF and offset.
+
+        Parameters
+        ----------
+        pmf : numpy.ndarray
+            1D non-negative array representing the probability mass function.
+        offset : int, default=0
+            Integer value corresponding to index 0 of ``pmf``.
+
+        Returns
+        -------
+        Distribution
+            A new distribution with the specified PMF and offset.
+        """
         p = np.asarray(pmf, dtype=float).copy()
         _validate_pmf(p)
         return cls(probabilities=p, offset=int(offset))
 
     @classmethod
-    def from_poisson_with_cap(
-        cls, lambda_param: float, max_k: int
-    ) -> "Distribution":
+    def from_poisson_with_cap(cls, lambda_param: float, max_k: int) -> "Distribution":
+        """Create a Poisson distribution truncated at ``max_k``.
+
+        Parameters
+        ----------
+        lambda_param : float
+            Poisson rate parameter (mean and variance).
+        max_k : int
+            Maximum value to include in the PMF (inclusive).
+
+        Returns
+        -------
+        Distribution
+            Truncated Poisson distribution over ``[0, max_k]``.
+        """
         if lambda_param < 0:
             raise ValueError("Poisson lambda must be non-negative")
         if lambda_param == 0.0:
@@ -60,6 +89,20 @@ class Distribution:
 
     @staticmethod
     def poisson_cap_from_k_sigma(lambda_param: float, k_sigma: float) -> int:
+        """Return a k-sigma truncation cap for a Poisson distribution.
+
+        Parameters
+        ----------
+        lambda_param : float
+            Poisson rate parameter.
+        k_sigma : float
+            Number of standard deviations to include above the mean.
+
+        Returns
+        -------
+        int
+            Truncation value ``k_max`` such that support is ``[0, k_max]``.
+        """
         if lambda_param <= 0:
             return 0
[…]
