### 2025-10-20 a4adcc4
implement router to route distributions into cohorts
