### 2025-10-20 6552b93
implement router to route distributions into cohorts
