### 2025-10-20 c2922c9
Revert "implement router to route distributions into cohorts"
This reverts commit 6552b93dbfa74c900919d228029ff6c7b028e230.
