+++ src/patientflow/predict/subspecialty.py
@@ -49,12 +49,47 @@ from patientflow.calculate.admission_in_prediction_window import (
     calculate_admission_probability_from_survival_curve,
 )
 from patientflow.model_artifacts import TrainedClassifier
-from patientflow.predict.routing import IdentityCohortRouter, FlowSpec
 
 
 @dataclass(frozen=True)
 class FlowInputs:
-    """Deprecated. Use FlowSpec in patientflow.predict.routing instead."""
+    """Represents a single source of patient flow.
+
+    This class encapsulates a flow of patients (either arriving or departing)
+    with its distribution type and parameters. It provides a uniform interface
+    for both probability mass functions and Poisson-distributed flows.
+
+    Attributes
+    ----------
+    flow_id : str
+        Unique identifier for this flow (e.g., "ed_current", "transfers_in")
+    flow_type : str
+        Type of distribution: "pmf" for probability mass function or "poisson" for Poisson
+    distribution : np.ndarray or float
+        For "pmf": numpy array where distribution[k] = P(k patients)
+        For "poisson": float representing the Poisson rate parameter (lambda)
+    display_name : str, optional
+        Human-readable name for display purposes. If not provided, flow_id will be
+        formatted automatically (underscores replaced with spaces, title cased).
+
+    Examples
+    --------
+    >>> # PMF flow (e.g., current ED patients)
+    >>> ed_flow = FlowInputs(
+    ...     flow_id="ed_current",
+    ...     flow_type="pmf",
+    ...     distribution=np.array([0.5, 0.3, 0.2]),
+    ...     display_name="Admissions from current ED"
+    ... )
+
+    >>> # Poisson flow (e.g., yet-to-arrive patients)
+    >>> yta_flow = FlowInputs(
+    ...     flow_id="ed_yta",
+    ...     flow_type="poisson",
+    ...     distribution=2.5,
+    ...     display_name="ED yet-to-arrive admissions"
+    ... )
+    """
 
     flow_id: str
     flow_type: str
@@ -62,7 +97,16 @@ class FlowInputs:
     display_name: Optional[str] = None
 
     def get_display_name(self) -> str:
-        return self.display_name or self.flow_id.replace("_", " ").title()
+        """Get human-readable display name.
+
+        Returns
+        -------
+        str
+            Display name if provided, otherwise formatted flow_id.
+        """
+        if self.display_name:
+            return self.display_name
+        return self.flow_id.replace("_", " ").title()
 
 
 @dataclass(frozen=True)
@@ -581,13 +625,13 @@ def build_subspecialty_data(
         # Build FlowInputs objects for inflows and outflows
         # INFLOWS: All sources of patient arrivals to this subspecialty
         inflows_dict = {
-            "ed_current": FlowSpec(
+            "ed_current": FlowInputs(
                 flow_id="ed_current",
                 flow_type="pmf",
                 distribution=np.array(agg_predicted_in_ed["agg_proba"]),
                 display_name="Admissions from current ED",
             ),
-            "ed_yta": FlowSpec(
+            "ed_yta": FlowInputs(
                 flow_id="ed_yta",
                 flow_type="poisson",
                 distribution=float(
@@ -597,7 +641,7 @@ def build_subspecialty_data(
                 ),
                 display_name="ED yet-to-arrive admissions",
             ),
-            "non_ed_yta": FlowSpec(
+            "non_ed_yta": FlowInputs(
                 flow_id="non_ed_yta",
                 flow_type="poisson",
                 distribution=float(
@@ -607,7 +651,7 @@ def build_subspecialty_data(
                 ),
                 display_name="Non-ED emergency admissions",
             ),
-            "elective_yta": FlowSpec(
+            "elective_yta": FlowInputs(
[…]
