### 2025-10-20 a53f51e
remove obsolete functions from subspecialty.py
