+++ src/patientflow/predict/subspecialty.py
@@ -44,6 +44,7 @@ from patientflow.aggregate import (
     model_input_to_pred_proba,
     pred_proba_to_agg_predicted,
 )
+from patientflow.predict.distribution import Distribution
 from patientflow.calculate.admission_in_prediction_window import (
     calculate_probability,
     calculate_admission_probability_from_survival_curve,
@@ -215,7 +216,7 @@ class SubspecialtyPredictionInputs:
         return "\n".join(lines)
 
 
-def build_subspecialty_data(
+def _validate_models_and_data(
     models: Tuple[
         TrainedClassifier,
         TrainedClassifier,
@@ -233,63 +234,10 @@ def build_subspecialty_data(
     prediction_time: Tuple[int, int],
     ed_snapshots: pd.DataFrame,
     inpatient_snapshots: pd.DataFrame,
-    specialties: List[str],
     prediction_window,
-    x1: float,
-    y1: float,
-    x2: float,
-    y2: float,
-    cdf_cut_points: Optional[List[float]] = None,
-    use_admission_in_window_prob: bool = True,
-) -> Dict[str, SubspecialtyPredictionInputs]:
-    """Build per-subspecialty inputs for downstream roll-up.
-
-    This function processes current patient snapshots through trained models and
-    computes, for each subspecialty, the probability distribution of admissions
-    from current ED patients, departures from current inpatients, the expected
-    means of yet-to-arrive admissions, and transfer arrival distributions.
-
-    Parameters
-    ----------
-    models : tuple
-        Tuple of seven trained models:
-
-        - ed_classifier: TrainedClassifier for ED admission probability prediction
-        - inpatient_classifier: TrainedClassifier for inpatient departure probability prediction
-        - spec_model: SequenceToOutcomePredictor | ValueToOutcomePredictor | MultiSubgroupPredictor
-          for specialty assignment probabilities
-        - ed_yta_model: ParametricIncomingAdmissionPredictor | EmpiricalIncomingAdmissionPredictor
-          for ED yet-to-arrive predictions
-        - non_ed_yta_model: DirectAdmissionPredictor for non-ED emergency predictions
-        - elective_yta_model: DirectAdmissionPredictor for elective predictions
-        - transfer_model: TransferProbabilityEstimator for internal transfer predictions
-    prediction_time : tuple of (int, int)
-        Hour and minute for inference time
-    ed_snapshots : pandas.DataFrame
-        DataFrame of current ED patients. Must include 'elapsed_los' column as timedelta.
-        Each row represents a patient currently in the ED.
-    inpatient_snapshots : pandas.DataFrame
-        DataFrame of current inpatients. Must include 'elapsed_los' column as timedelta.
-        Each row represents a patient currently in a subspecialty ward.
-    specialties : list of str
-        List of subspecialties to prepare inputs for
-    prediction_window : datetime.timedelta
-        Time window over which to predict admissions
-    x1, y1, x2, y2 : float
-        Parameters for the parametric admission-in-window curve. Used when
-        ed_yta_model is parametric and for computing in-ED window probabilities.
-    cdf_cut_points : list of float, optional
-        Ignored in this function; present for API compatibility. If provided,
-        has no effect on output.
-    use_admission_in_window_prob : bool, default=True
-        Whether to weight current ED admissions by their probability of being
-        admitted within the prediction window.
-
-    Returns
-    -------
-    dict of str to SubspecialtyPredictionInputs
-        Dictionary mapping subspecialty_id to SubspecialtyPredictionInputs dataclass.
[…]
