### 2025-10-20 1facf84
remove obsolete functions from subspecialty.py
