+++ src/patientflow/predict/__init__.py
@@ -8,14 +8,10 @@ from patientflow.predict.subspecialty import (
     build_subspecialty_data,
     SubspecialtyPredictionInputs,
     compute_transfer_arrivals,
-    scale_pmf_by_probability,
-    convolve_pmfs,
 )
 
 __all__ = [
     "build_subspecialty_data",
     "SubspecialtyPredictionInputs",
     "compute_transfer_arrivals",
-    "scale_pmf_by_probability",
-    "convolve_pmfs",
 ]
