### 2025-10-20 f3108f9
remove obsolete functions from subspecialty.py
