### 2025-10-20 60def2e
predict transfers by admission type
