+++ tests/test_transfer_predictor.py
@@ -169,6 +169,173 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
             predictor.fit(X_unknown, self.subspecialties)
         self.assertIn("unknown_subspecialty", str(context.exception))
 
+    def test_cohort_functionality_basic(self):
+        """Test basic cohort functionality with separate models for each cohort."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology", "cardiology", "surgery", "surgery"],
+                "next_subspecialty": ["surgery", None, None, "cardiology"],
+                "admission_type": ["elective", "elective", "emergency", "emergency"],
+            }
+        )
+        predictor = TransferProbabilityEstimator(cohort_col="admission_type")
+        predictor.fit(X, self.subspecialties)
+
+        # Check that cohorts are properly identified
+        cohorts = predictor.get_available_cohorts()
+        self.assertEqual(cohorts, {"elective", "emergency"})
+
+        # Check elective cohort: cardiology 1/2 transfers, surgery 0/0 transfers
+        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology", "elective"), 0.5)
+        self.assertAlmostEqual(predictor.get_transfer_prob("surgery", "elective"), 0.0)
+
+        # Check emergency cohort: cardiology 0/0 transfers, surgery 1/2 transfers
+        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology", "emergency"), 0.0)
+        self.assertAlmostEqual(predictor.get_transfer_prob("surgery", "emergency"), 0.5)
+
+    def test_cohort_functionality_no_cohorts(self):
+        """Test that predictor works without cohorts (backward compatibility)."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology", "cardiology"],
+                "next_subspecialty": ["surgery", None],
+            }
+        )
+        predictor = TransferProbabilityEstimator()  # No cohort_col
+        predictor.fit(X, self.subspecialties)
+
+        # Should create "all" cohort automatically
+        cohorts = predictor.get_available_cohorts()
+        self.assertEqual(cohorts, {"all"})
+
+        # Should work without specifying cohort
+        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology"), 0.5)
+
+    def test_cohort_functionality_single_cohort_auto_select(self):
+        """Test automatic cohort selection when only one cohort exists."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology", "cardiology"],
+                "next_subspecialty": ["surgery", None],
+                "admission_type": ["elective", "elective"],  # Only one cohort
+            }
+        )
+        predictor = TransferProbabilityEstimator(cohort_col="admission_type")
+        predictor.fit(X, self.subspecialties)
+
+        # Should automatically use the single cohort
+        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology"), 0.5)
+        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology", "elective"), 0.5)
+
+    def test_cohort_functionality_multiple_cohorts_require_specification(self):
+        """Test that multiple cohorts require explicit specification."""
+        X = pd.DataFrame(
+            {
+                "current_subspecialty": ["cardiology", "cardiology"],
+                "next_subspecialty": ["surgery", None],
+                "admission_type": ["elective", "emergency"],
+            }
+        )
+        predictor = TransferProbabilityEstimator(cohort_col="admission_type")
+        predictor.fit(X, self.subspecialties)
+
+        # Should raise error when no cohort specified with multiple cohorts
+        with self.assertRaises(ValueError) as context:
+            predictor.get_transfer_prob("cardiology")
[…]
