### 2025-10-20 66f6311
predict transfers by admission type
