+++ src/patientflow/predictors/transfer_predictor.py
@@ -6,6 +6,11 @@ subspecialties based on historical movement patterns. It computes the probabilit
 that a departing patient will transfer to another subspecialty versus being
 discharged, and the distribution of destination subspecialties for transfers.
 
+The predictor supports cohort-based analysis, allowing separate transfer probability
+models to be trained for different patient groups (e.g., elective vs emergency
+admissions). This enables more accurate modeling of patient flows when transfer
+patterns differ significantly between cohorts.
+
 The predictor follows sklearn conventions with fit() and predict() methods, allowing
 it to be saved and loaded as a model artifact alongside other predictors in the
 patientflow ecosystem.
@@ -19,6 +24,9 @@ external discharges and internal transfers.
 The predictor stores transition probabilities between subspecialties, which can be
 used at prediction time to model internal patient flows and compute arrival
 distributions for each subspecialty from transfers.
+
+When cohort_col is specified, separate probability models are trained for each
+cohort. When cohort_col is None, all data is used together to train a single model.
 """
 
 import pandas as pd
@@ -45,6 +53,13 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     destination_col : str, default='next_subspecialty'
         Name of the column containing the destination subspecialty (where patient is going to).
         None/NaN values indicate discharge rather than transfer.
+    cohort_col : str, optional
+        Name of the column containing cohort information (e.g., 'admission_type').
+        If None, all data is used together. If specified, separate models are trained
+        for each cohort.
+    cohort_values : list, optional
+        List of expected cohort values for validation. If provided, the fit method
+        will validate that only these values appear in the cohort_col.
 
     Attributes
     ----------
@@ -52,21 +67,37 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         Name of the source subspecialty column
     destination_col : str
         Name of the destination subspecialty column
+    cohort_col : str or None
+        Name of the cohort column, or None if not using cohorts
+    cohort_values : list or None
+        List of expected cohort values for validation, or None
     transfer_probabilities : dict
-        Nested dictionary containing transfer statistics for each subspecialty:
+        Nested dictionary containing transfer statistics. Structure depends on
+        whether cohorts are used:
+        
+        If cohort_col is None:
         {
-            'source_subspecialty': {
-                'prob_transfer': float,
-                    Probability that a departure from this subspecialty results
-                    in a transfer to another subspecialty (vs discharge)
-                'destination_distribution': dict,
-                    Distribution of destinations given a transfer occurs.
-                    Maps destination subspecialty names to probabilities that
-                    sum to 1.0
+            'all': {
+                'source_subspecialty': {
+                    'prob_transfer': float,
+                    'destination_distribution': dict
+                }
+            }
+        }
+        
+        If cohort_col is specified:
+        {
+            'cohort_name': {
+                'source_subspecialty': {
+                    'prob_transfer': float,
+                    'destination_distribution': dict
+                }
             }
         }
     subspecialties : set
         Set of all subspecialties in the system
+    cohorts : set or None
+        Set of all cohorts in the system, or None if not using cohorts
     is_fitted_ : bool
         Whether the predictor has been fitted
 
@@ -75,23 +106,24 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     >>> # Prepare training data with subspecialty movements
[…]
