### 2025-10-21 31df80a
separate elective and emergency in subspecialty inputs
