+++ src/patientflow/predict/subspecialty.py
@@ -136,12 +136,14 @@ class SubspecialtyPredictionInputs:
         - "ed_yta": Yet-to-arrive ED patients who will be admitted (Poisson)
         - "non_ed_yta": Yet-to-arrive non-ED emergency admissions (Poisson)
         - "elective_yta": Yet-to-arrive elective admissions (Poisson)
-        - "transfers_in": Patients transferring from other subspecialties (PMF)
+        - "elective_transfers": Elective patients transferring from other subspecialties (PMF)
+        - "emergency_transfers": Emergency patients transferring from other subspecialties (PMF)
     outflows : Dict[str, FlowInputs]
         Dictionary mapping flow identifiers to FlowInputs objects for departures.
         Standard keys include:
 
-        - "departures": Current inpatients who will depart (PMF)
+        - "elective_departures": Current elective inpatients who will depart (PMF)
+        - "emergency_departures": Current emergency inpatients who will depart (PMF)
 
         Future extensions may include "transfers_out", "deaths", etc.
 
@@ -485,10 +487,18 @@ def _prepare_base_probabilities(
         "elapsed_los"
     ].dt.total_seconds()
 
+    # Split inpatient snapshots into elective and emergency
+    elective_snapshots = inpatient_snapshots_temp[inpatient_snapshots_temp["admission_type"] == "elective"]
+    emergency_snapshots = inpatient_snapshots_temp[inpatient_snapshots_temp["admission_type"] == "emergency"]
+
     # Departure probability for current inpatients (per row)
-    prob_departure_after_inpatient = model_input_to_pred_proba(
-        inpatient_snapshots_temp, inpatient_pipeline
-    )
+    prob_departure_after_elective = model_input_to_pred_proba(
+        elective_snapshots, inpatient_pipeline
+    ) if not elective_snapshots.empty else pd.Series(dtype=float)
+    
+    prob_departure_after_emergency = model_input_to_pred_proba(
+        emergency_snapshots, inpatient_pipeline
+    ) if not emergency_snapshots.empty else pd.Series(dtype=float)
 
     # Specialty probabilities per row for ED patients
     if hasattr(spec_model, "predict_dataframe"):
@@ -557,7 +567,8 @@ def _prepare_base_probabilities(
         "ed_snapshots": ed_snapshots,
         "inpatient_snapshots": inpatient_snapshots,
         "prob_admission_after_ed": prob_admission_after_ed,
-        "prob_departure_after_inpatient": prob_departure_after_inpatient,
+        "prob_departure_after_elective": prob_departure_after_elective,
+        "prob_departure_after_emergency": prob_departure_after_emergency,
         "prob_admission_in_window": prob_admission_in_window,
         "specialty_to_subgroups": specialty_to_subgroups,
         "ed_masks_by_func": ed_masks_by_func,
@@ -616,42 +627,54 @@ def _process_ed_patients_for_specialty(
     }
 
 
-def _process_inpatients_for_specialty(
+def _process_inpatients_for_specialty_by_admission_type(
     spec: str,
     inpatient_snapshots: pd.DataFrame,
-    prob_departure_after_inpatient: pd.Series,
+    prob_departure_series: pd.Series,
+    admission_type: str,
 ) -> Dict[str, Any]:
-    """Process inpatients for a specific specialty.
+    """Process inpatients for a specific specialty and admission type.
+
+    Parameters
+    ----------
+    spec : str
+        The subspecialty to process
+    inpatient_snapshots : pd.DataFrame
+        DataFrame containing inpatient snapshot data
+    prob_departure_series : pd.Series
+        Series containing departure probabilities for the admission type
+    admission_type : str
+        The admission type to process ("elective" or "emergency")
 
     Returns
     -------
     dict
[…]
