### 2025-10-21 60a08c9
add new tests for subspecialty input change
