### 2025-10-21 821d200
add new tests for subspecialty input change
