+++ tests/test_subspecialty.py
@@ -50,6 +50,7 @@ def _create_random_df(n=1000, include_consults=False):
     )
     sex = np.random.choice(["M", "F"], size=n)
     is_admitted = np.random.choice([0, 1], size=n)
+    admission_type = np.random.choice(["elective", "emergency"], size=n)
 
     df = pd.DataFrame(
         {
@@ -59,6 +60,7 @@ def _create_random_df(n=1000, include_consults=False):
             "arrival_method": arrival_method,
             "sex": sex,
             "is_admitted": is_admitted,
+            "admission_type": admission_type,
         }
     )
 
@@ -354,16 +356,22 @@ class TestBuildSubspecialtyData(unittest.TestCase):
             self.assertIn("ed_yta", spec_data.inflows)
             self.assertIn("non_ed_yta", spec_data.inflows)
             self.assertIn("elective_yta", spec_data.inflows)
-            self.assertIn("transfers_in", spec_data.inflows)
+            self.assertIn("elective_transfers", spec_data.inflows)
+            self.assertIn("emergency_transfers", spec_data.inflows)
             # Check outflows
-            self.assertIn("departures", spec_data.outflows)
+            self.assertIn("elective_departures", spec_data.outflows)
+            self.assertIn("emergency_departures", spec_data.outflows)
             # Check values
             ed_pmf = np.asarray(spec_data.inflows["ed_current"].distribution)
-            inpatient_pmf = np.asarray(spec_data.outflows["departures"].distribution)
-            transfer_pmf = np.asarray(spec_data.inflows["transfers_in"].distribution)
+            elective_departure_pmf = np.asarray(spec_data.outflows["elective_departures"].distribution)
+            emergency_departure_pmf = np.asarray(spec_data.outflows["emergency_departures"].distribution)
+            elective_transfer_pmf = np.asarray(spec_data.inflows["elective_transfers"].distribution)
+            emergency_transfer_pmf = np.asarray(spec_data.inflows["emergency_transfers"].distribution)
             self.assertGreater(len(ed_pmf), 0)
-            self.assertGreater(len(inpatient_pmf), 0)
-            self.assertGreater(len(transfer_pmf), 0)
+            self.assertGreater(len(elective_departure_pmf), 0)
+            self.assertGreater(len(emergency_departure_pmf), 0)
+            self.assertGreater(len(elective_transfer_pmf), 0)
+            self.assertGreater(len(emergency_transfer_pmf), 0)
             self.assertIsInstance(spec_data.inflows["ed_yta"].distribution, float)
             self.assertIsInstance(spec_data.inflows["non_ed_yta"].distribution, float)
             self.assertIsInstance(spec_data.inflows["elective_yta"].distribution, float)
@@ -546,17 +554,27 @@ class TestComputeTransferArrivals(unittest.TestCase):
         subspecialty_data = {
             "cardiology": {
                 "outflows": {
-                    "departures": FlowInputs(
-                        flow_id="departures",
+                    "elective_departures": FlowInputs(
+                        flow_id="elective_departures",
                         flow_type="pmf",
-                        distribution=np.array([0.0, 0.0, 1.0]),  # 2 departures certain
+                        distribution=np.array([1.0, 0.0]),  # No elective departures
+                    ),
+                    "emergency_departures": FlowInputs(
+                        flow_id="emergency_departures",
+                        flow_type="pmf",
+                        distribution=np.array([0.0, 0.0, 1.0]),  # 2 emergency departures certain
                     )
                 }
             },
             "surgery": {
                 "outflows": {
-                    "departures": FlowInputs(
-                        flow_id="departures",
+                    "elective_departures": FlowInputs(
+                        flow_id="elective_departures",
+                        flow_type="pmf",
[…]
