### 2025-10-21 3aad1ec
update hierarchy.py for subspecialty input change
