+++ src/patientflow/predictors/transfer_predictor.py
@@ -74,7 +74,7 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     transfer_probabilities : dict
         Nested dictionary containing transfer statistics. Structure depends on
         whether cohorts are used:
-        
+
         If cohort_col is None:
         {
             'all': {
@@ -84,7 +84,7 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
                 }
             }
         }
-        
+
         If cohort_col is specified:
         {
             'cohort_name': {
@@ -155,7 +155,9 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     def __repr__(self) -> str:
         class_name = self.__class__.__name__
         if not self.is_fitted_:
-            cohort_info = f",\n    cohort_col='{self.cohort_col}'" if self.cohort_col else ""
+            cohort_info = (
+                f",\n    cohort_col='{self.cohort_col}'" if self.cohort_col else ""
+            )
             return (
                 f"{class_name}(\n"
                 f"    source_col='{self.source_col}',\n"
@@ -177,7 +179,11 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
                 if stats["prob_transfer"] > 0
             )
 
-        cohort_info = f",\n    cohort_col='{self.cohort_col}',\n    n_cohorts={n_cohorts}" if self.cohort_col else ""
+        cohort_info = (
+            f",\n    cohort_col='{self.cohort_col}',\n    n_cohorts={n_cohorts}"
+            if self.cohort_col
+            else ""
+        )
         return (
             f"{class_name}(\n"
             f"    source_col='{self.source_col}',\n"
@@ -226,7 +232,7 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         required_columns = {self.source_col, self.destination_col}
         if self.cohort_col is not None:
             required_columns.add(self.cohort_col)
-            
+
         if not required_columns.issubset(X.columns):
             missing = required_columns - set(X.columns)
             raise ValueError(
@@ -257,7 +263,7 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         if self.cohort_col is not None:
             # Extract cohorts from data
             self.cohorts = set(X[self.cohort_col].dropna().unique())
-            
+
             # Validate cohort values if provided
             if self.cohort_values is not None:
                 unknown_cohorts = self.cohorts - set(self.cohort_values)
@@ -267,20 +273,22 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
                         f"cohort_values parameter: {sorted(unknown_cohorts)}. "
                         f"Expected cohorts: {sorted(self.cohort_values)}"
                     )
-            
+
             # Compute transfer probabilities for each cohort
             self.transfer_probabilities = {}
             for cohort in self.cohorts:
                 cohort_data = X[X[self.cohort_col] == cohort]
-                self.transfer_probabilities[cohort] = self._prepare_transfer_probabilities(
-                    self.subspecialties, cohort_data
+                self.transfer_probabilities[cohort] = (
+                    self._prepare_transfer_probabilities(
+                        self.subspecialties, cohort_data
+                    )
                 )
         else:
             # No cohort processing - use all data
             self.cohorts = None
-            self.transfer_probabilities = {"all": self._prepare_transfer_probabilities(
-                self.subspecialties, X
-            )}
+            self.transfer_probabilities = {
+                "all": self._prepare_transfer_probabilities(self.subspecialties, X)
+            }
 
         self.is_fitted_ = True
         return self
[…]
