+++ src/patientflow/predict/hierarchy.py
@@ -8,6 +8,7 @@ The main components are:
 
 - DemandPrediction: A dataclass representing prediction results with probabilities,
   expected values, and percentiles
+- FlowSelection: Configuration for selecting which patient flows to include in predictions
 - DemandPredictor: Core prediction engine using convolution of probability distributions
 - HospitalHierarchy: Represents organizational structure of a hospital
 - HierarchicalPredictor: High-level interface for making predictions across all levels
@@ -29,6 +30,12 @@ from dataclasses import dataclass
 
 from patientflow.predict.subspecialty import SubspecialtyPredictionInputs, FlowInputs
 from patientflow.predict.distribution import Distribution
+
+
+# Constants for magic numbers
+DEFAULT_PERCENTILES = [50, 75, 90, 95, 99]
+DEFAULT_PRECISION = 3
+DEFAULT_MAX_PROBS = 10
 
 
 @dataclass
@@ -98,7 +105,9 @@ class DemandPrediction:
         """
         return len(self.probabilities) - 1
 
-    def to_pretty(self, max_probs: int = 10, precision: int = 3) -> str:
+    def to_pretty(
+        self, max_probs: int = DEFAULT_MAX_PROBS, precision: int = DEFAULT_PRECISION
+    ) -> str:
         """Return a concise, human-friendly string representation.
 
         Parameters
@@ -152,119 +161,95 @@ class DemandPrediction:
 class FlowSelection:
     """Configuration for which flows to include in predictions.
 
-    This class controls which patient flows (arrivals and departures) are included
-    when making predictions. It allows users to customize predictions based on their
-    specific use cases, such as excluding transfers or focusing only on certain types
-    of admissions.
+    This defines booleans to include each flow family and a
+    single cohort selector ("all", "elective", "emergency").
 
-    Attributes
-    ----------
-    inflow_keys : List[str]
-        List of inflow identifiers to include in predictions.
-        Standard values: ["ed_current", "ed_yta", "non_ed_yta", "elective_yta", "transfers_in"]
-    outflow_keys : List[str]
-        List of outflow identifiers to include in predictions.
-        Standard values: ["departures"]
-        Future extensions may include: ["transfers_out", "deaths"]
-
-    Examples
+    Families
     --------
-    >>> # Include all flows (arrivals and departures)
-    >>> selection = FlowSelection.default()
+    Inflows families:
+    - include_ed_current: current ED cohort admissions (emergency only)
+    - include_ed_yta: yet-to-arrive ED admissions (emergency only)
+    - include_non_ed_yta: yet-to-arrive non-ED emergency admissions (emergency only)
+    - include_elective_yta: yet-to-arrive elective admissions (elective only)
+    - include_transfers_in: internal transfers into subspecialty (both cohorts)
 
-    >>> # Only incoming flows (all arrivals, no departures)
-    >>> selection = FlowSelection.incoming_flow()
+    Outflows families:
+    - include_departures: inpatient departures (both cohorts)
 
-    >>> # Only outgoing flows (all departures, no arrivals)
-    >>> selection = FlowSelection.outgoing_flow()
-
-    >>> # Custom selection
-    >>> selection = FlowSelection.custom(
-    ...     include_inflows=["ed_current", "ed_yta"],
-    ...     include_outflows=["departures"]
-    ... )
+    Cohort
+    ------
+    - cohort: one of {"all", "elective", "emergency"}
     """
 
-    inflow_keys: List[str]
-    outflow_keys: List[str]
+    include_ed_current: bool = True
+    include_ed_yta: bool = True
+    include_non_ed_yta: bool = True
+    include_elective_yta: bool = True
+    include_transfers_in: bool = True
+    include_departures: bool = True
[…]

[…]
