### 2025-10-21 9db6d1f
update hierarchy.py for subspecialty input change
