+++ tests/test_transfer_predictor.py
@@ -173,7 +173,12 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
         """Test basic cohort functionality with separate models for each cohort."""
         X = pd.DataFrame(
             {
-                "current_subspecialty": ["cardiology", "cardiology", "surgery", "surgery"],
+                "current_subspecialty": [
+                    "cardiology",
+                    "cardiology",
+                    "surgery",
+                    "surgery",
+                ],
                 "next_subspecialty": ["surgery", None, None, "cardiology"],
                 "admission_type": ["elective", "elective", "emergency", "emergency"],
             }
@@ -186,11 +191,15 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
         self.assertEqual(cohorts, {"elective", "emergency"})
 
         # Check elective cohort: cardiology 1/2 transfers, surgery 0/0 transfers
-        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology", "elective"), 0.5)
+        self.assertAlmostEqual(
+            predictor.get_transfer_prob("cardiology", "elective"), 0.5
+        )
         self.assertAlmostEqual(predictor.get_transfer_prob("surgery", "elective"), 0.0)
 
         # Check emergency cohort: cardiology 0/0 transfers, surgery 1/2 transfers
-        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology", "emergency"), 0.0)
+        self.assertAlmostEqual(
+            predictor.get_transfer_prob("cardiology", "emergency"), 0.0
+        )
         self.assertAlmostEqual(predictor.get_transfer_prob("surgery", "emergency"), 0.5)
 
     def test_cohort_functionality_no_cohorts(self):
@@ -225,7 +234,9 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
 
         # Should automatically use the single cohort
         self.assertAlmostEqual(predictor.get_transfer_prob("cardiology"), 0.5)
-        self.assertAlmostEqual(predictor.get_transfer_prob("cardiology", "elective"), 0.5)
+        self.assertAlmostEqual(
+            predictor.get_transfer_prob("cardiology", "elective"), 0.5
+        )
 
     def test_cohort_functionality_multiple_cohorts_require_specification(self):
         """Test that multiple cohorts require explicit specification."""
@@ -268,7 +279,9 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
                 "next_subspecialty": [None],
             }
         )
-        predictor_with_cohort = TransferProbabilityEstimator(cohort_col="admission_type")
+        predictor_with_cohort = TransferProbabilityEstimator(
+            cohort_col="admission_type"
+        )
         with self.assertRaises(ValueError) as context:
             predictor_with_cohort.fit(X_no_cohort, self.subspecialties)
         self.assertIn("missing required columns", str(context.exception))
@@ -282,11 +295,10 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
                 "admission_type": ["elective"],
             }
         )
-        
+
         # Test with valid cohort_values
         predictor = TransferProbabilityEstimator(
-            cohort_col="admission_type", 
-            cohort_values=["elective", "emergency"]
+            cohort_col="admission_type", cohort_values=["elective", "emergency"]
         )
         predictor.fit(X, self.subspecialties)  # Should not raise
 
@@ -299,12 +311,14 @@ class TestTransferProbabilityEstimator(unittest.TestCase):
             }
         )
         predictor_with_validation = TransferProbabilityEstimator(
-            cohort_col="admission_type", 
-            cohort_values=["elective", "emergency"]
+            cohort_col="admission_type", cohort_values=["elective", "emergency"]
         )
[…]

[…]
