### 2025-10-21 a2a02eb
fix bug in FlowSelection
