### 2025-10-21 e226f11
add subgroups to transfer predictor
