### 2025-10-21 ee9161b
add subgroups to transfer predictor
