+++ src/patientflow/predictors/transfer_predictor.py
@@ -30,8 +30,9 @@ cohort. When cohort_col is None, all data is used together to train a single mod
 """
 
 import pandas as pd
-from typing import Dict, Set, Optional
+from typing import Dict, Set, Optional, Callable, Union
 from sklearn.base import BaseEstimator, TransformerMixin
+from patientflow.predictors.subgroup_definitions import create_subgroup_functions
 
 
 class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
@@ -60,6 +61,9 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     cohort_values : list, optional
         List of expected cohort values for validation. If provided, the fit method
         will validate that only these values appear in the cohort_col.
+    subgroup_functions : dict, optional
+        Dictionary mapping subgroup names to functions that identify patients in each subgroup.
+        If None, uses the standard 5 subgroup functions (paediatric, adult_male_young, etc.).
 
     Attributes
     ----------
@@ -91,6 +95,12 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
                 'source_subspecialty': {
                     'prob_transfer': float,
                     'destination_distribution': dict
+                },
+                'subgroup_name': {
+                    'source_subspecialty': {
+                        'prob_transfer': float,
+                        'destination_distribution': dict
+                    }
                 }
             }
         }
@@ -98,6 +108,8 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         Set of all subspecialties in the system
     cohorts : set or None
         Set of all cohorts in the system, or None if not using cohorts
+    subgroups : set or None
+        Set of all subgroups in the system, or None if not using subgroups
     is_fitted_ : bool
         Whether the predictor has been fitted
 
@@ -142,14 +154,17 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         destination_col: str = "next_subspecialty",
         cohort_col: Optional[str] = None,
         cohort_values: Optional[list] = None,
+        subgroup_functions: Optional[Dict[str, Callable[[Union[pd.Series, dict]], bool]]] = None,
     ):
         self.source_col = source_col
         self.destination_col = destination_col
         self.cohort_col = cohort_col
         self.cohort_values = cohort_values
-        self.transfer_probabilities: Optional[Dict[str, Dict[str, Dict]]] = None
+        self.subgroup_functions = subgroup_functions or create_subgroup_functions()
+        self.transfer_probabilities: Optional[Dict[str, Dict[str, Dict[str, Dict]]]] = None
         self.subspecialties: Optional[Set[str]] = None
         self.cohorts: Optional[Set[str]] = None
+        self.subgroups: Optional[Set[str]] = None
         self.is_fitted_ = False
 
     def __repr__(self) -> str:
@@ -278,17 +293,55 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
             self.transfer_probabilities = {}
             for cohort in self.cohorts:
                 cohort_data = X[X[self.cohort_col] == cohort]
-                self.transfer_probabilities[cohort] = (
-                    self._prepare_transfer_probabilities(
-                        self.subspecialties, cohort_data
-                    )
+                
+                # Train global model for this cohort (all data) - store directly at cohort level
+                self.transfer_probabilities[cohort] = self._prepare_transfer_probabilities(
+                    self.subspecialties, cohort_data
                 )
+                
+                # Train subgroup models for this cohort
+                for subgroup_name, subgroup_func in self.subgroup_functions.items():
+                    subgroup_data = cohort_data[cohort_data.apply(subgroup_func, axis=1)]
+                    if len(subgroup_data) > 0:  # Only train if subgroup has data
[…]
