### 2025-10-21 870232d
add subgroups to transfer predictor
