### 2025-10-21 03505ca
Improve transfer predictor __repr__ implementation
- Add subgroups support
- Fix bugs in representation
- Separate elective/emergency subgroup counts
- Make output more concise
