+++ src/patientflow/predictors/transfer_predictor.py
@@ -165,6 +165,7 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         self.subspecialties: Optional[Set[str]] = None
         self.cohorts: Optional[Set[str]] = None
         self.subgroups: Optional[Set[str]] = None
+        self.patient_counts: Optional[Dict[str, Dict[str, int]]] = None
         self.is_fitted_ = False
 
     def __repr__(self) -> str:
@@ -185,24 +186,83 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
             len(self.subspecialties) if self.subspecialties is not None else 0
         )
         n_cohorts = len(self.cohorts) if self.cohorts is not None else 0
+        n_subgroups = len(self.subgroups) if self.subgroups is not None else 0
+        
         n_with_transfers = 0
+        cohort_subspecialty_counts = {}
+        subspecialties_with_transfers = set()
+        
         if self.transfer_probabilities is not None:
-            n_with_transfers = sum(
-                1
-                for cohort_stats in self.transfer_probabilities.values()
-                for stats in cohort_stats.values()
-                if stats["prob_transfer"] > 0
-            )
+            # Process each cohort separately to get cohort-specific counts
+            for cohort_name, cohort_data in self.transfer_probabilities.items():
+                cohort_subspecialty_counts[cohort_name] = {}
+                for key, value in cohort_data.items():
+                    # Check if this is subspecialty data (has prob_transfer at top level)
+                    if isinstance(value, dict) and "prob_transfer" in value:
+                        if value["prob_transfer"] > 0:
+                            subspecialties_with_transfers.add(key)  # Track unique subspecialties with transfers
+                    # Check if this is subgroup data (contains nested subspecialty data)
+                    elif isinstance(value, dict) and any(
+                        isinstance(v, dict) and "prob_transfer" in v 
+                        for v in value.values()
+                    ):
+                        # Count subspecialties in this subgroup for this specific cohort
+                        # Only count subspecialties that have actual data (prob_transfer > 0)
+                        subgroup_total_subspecialties = sum(
+                            1 for v in value.values()
+                            if isinstance(v, dict) and "prob_transfer" in v and v["prob_transfer"] > 0
+                        )
+                        cohort_subspecialty_counts[cohort_name][key] = subgroup_total_subspecialties
+            
+            # Set final count of unique subspecialties with transfers
+            n_with_transfers = len(subspecialties_with_transfers)
+        
+        # Collect patient counts for subgroups by cohort
+        cohort_subgroup_counts = {}
+        if self.patient_counts is not None:
+            for cohort_name, cohort_counts in self.patient_counts.items():
+                cohort_subgroup_counts[cohort_name] = cohort_counts.copy()
 
-        cohort_info = (
-            f",\n    cohort_col='{self.cohort_col}',\n    n_cohorts={n_cohorts}"
-            if self.cohort_col
-            else ""
+        # Build single-line core info (source, destination, cohort)
+        core_info = (
+            f"    source_col='{self.source_col}', destination_col='{self.destination_col}'"
         )
+        if self.cohort_col:
+            core_info += f", cohort_col='{self.cohort_col}', n_cohorts={n_cohorts}"
+        
+        # Build subgroup info lines (each cohort on its own line)
+        subgroup_info = ""
+        if n_subgroups > 0:
+            if cohort_subgroup_counts and cohort_subspecialty_counts:
+                lines = [f"    n_subgroups={n_subgroups}"]
+                for cohort_name in sorted(cohort_subgroup_counts.keys()):
+                    entries = []
+                    for subgroup_name in sorted(cohort_subgroup_counts[cohort_name].keys()):
[…]
