+++ src/patientflow/predictors/transfer_predictor.py
@@ -54,6 +54,11 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     destination_col : str, default='next_subspecialty'
         Name of the column containing the destination subspecialty (where patient is going to).
         None/NaN values indicate discharge rather than transfer.
+    visit_col : str, optional
+        Name of the column identifying a visit/encounter. If provided, the
+        training data `X` will be de-duplicated to a single row per unique
+        combination of `(visit_col, source_col, destination_col)` to avoid
+        multiple rows representing the same transition for a visit.
     cohort_col : str, optional
         Name of the column containing cohort information (e.g., 'admission_type').
         If None, all data is used together. If specified, separate models are trained
@@ -152,16 +157,22 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         self,
         source_col: str = "current_subspecialty",
         destination_col: str = "next_subspecialty",
+        visit_col: Optional[str] = None,
         cohort_col: Optional[str] = None,
         cohort_values: Optional[list] = None,
-        subgroup_functions: Optional[Dict[str, Callable[[Union[pd.Series, dict]], bool]]] = None,
+        subgroup_functions: Optional[
+            Dict[str, Callable[[Union[pd.Series, dict]], bool]]
+        ] = None,
     ):
         self.source_col = source_col
         self.destination_col = destination_col
+        self.visit_col = visit_col
         self.cohort_col = cohort_col
         self.cohort_values = cohort_values
         self.subgroup_functions = subgroup_functions or create_subgroup_functions()
-        self.transfer_probabilities: Optional[Dict[str, Dict[str, Dict[str, Dict]]]] = None
+        self.transfer_probabilities: Optional[Dict[str, Dict[str, Dict[str, Dict]]]] = (
+            None
+        )
         self.subspecialties: Optional[Set[str]] = None
         self.cohorts: Optional[Set[str]] = None
         self.subgroups: Optional[Set[str]] = None
@@ -187,11 +198,11 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         )
         n_cohorts = len(self.cohorts) if self.cohorts is not None else 0
         n_subgroups = len(self.subgroups) if self.subgroups is not None else 0
-        
+
         n_with_transfers = 0
-        cohort_subspecialty_counts = {}
-        subspecialties_with_transfers = set()
-        
+        cohort_subspecialty_counts: Dict[str, Dict[str, int]] = {}
+        subspecialties_with_transfers: Set[str] = set()
+
         if self.transfer_probabilities is not None:
             # Process each cohort separately to get cohort-specific counts
             for cohort_name, cohort_data in self.transfer_probabilities.items():
@@ -199,24 +210,35 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
                 for key, value in cohort_data.items():
                     # Check if this is subspecialty data (has prob_transfer at top level)
                     if isinstance(value, dict) and "prob_transfer" in value:
-                        if value["prob_transfer"] > 0:
-                            subspecialties_with_transfers.add(key)  # Track unique subspecialties with transfers
+                        prob_transfer = value["prob_transfer"]
+                        if (
+                            isinstance(prob_transfer, (int, float))
+                            and prob_transfer > 0
+                        ):
+                            subspecialties_with_transfers.add(
+                                key
+                            )  # Track unique subspecialties with transfers
                     # Check if this is subgroup data (contains nested subspecialty data)
                     elif isinstance(value, dict) and any(
-                        isinstance(v, dict) and "prob_transfer" in v 
+                        isinstance(v, dict) and "prob_transfer" in v
                         for v in value.values()
[…]
