### 2025-10-23 f12d5c6
make hierarchy generic
