+++ src/patientflow/predict/hierarchy.py
@@ -1,8 +1,7 @@
 """Hierarchical demand prediction for hospital bed capacity management.
 
 This module provides classes and functions for predicting hospital bed demand at multiple
-hierarchical levels (subspecialty, reporting unit, division, board, hospital) using
-probability distributions and convolution operations.
+hierarchical levels using probability distributions and convolution operations.
 
 The main components are:
 
@@ -10,22 +9,24 @@ The main components are:
   expected values, and percentiles
 - FlowSelection: Configuration for selecting which patient flows to include in predictions
 - DemandPredictor: Core prediction engine using convolution of probability distributions
-- HospitalHierarchy: Represents organizational structure of a hospital
+- Hierarchy: Generic hierarchical structure that can represent any organizational hierarchy
 - HierarchicalPredictor: High-level interface for making predictions across all levels
-- Utility functions for populating hierarchy from DataFrames
+- Utility functions for populating hierarchy from DataFrames and YAML configuration
 
 Functions
 ---------
 populate_hierarchy_from_dataframe
-    Create HospitalHierarchy from pandas DataFrame with organizational structure
+    Populate Hierarchy from pandas DataFrame with organizational structure
 create_hierarchical_predictor
     Create complete HierarchicalPredictor from DataFrame and parameters
 """
 
 import numpy as np
 import pandas as pd
+import yaml
 from typing import List, Dict, Optional, Any
 from dataclasses import dataclass
+from enum import Enum
 
 
 from patientflow.predict.subspecialty import SubspecialtyPredictionInputs, FlowInputs
@@ -36,6 +37,44 @@ from patientflow.predict.distribution import Distribution
 DEFAULT_PERCENTILES = [50, 75, 90, 95, 99]
 DEFAULT_PRECISION = 3
 DEFAULT_MAX_PROBS = 10
+
+
+class EntityType:
+    """Represents an entity type in the hierarchy.
+    
+    This class is used to represent entity types dynamically based on the
+    hierarchy configuration.
+    """
+    
+    def __init__(self, name: str):
+        self.name = name
+    
+    def __str__(self) -> str:
+        return self.name
+    
+    def __repr__(self) -> str:
+        return f"EntityType('{self.name}')"
+    
+    def __eq__(self, other) -> bool:
+        if isinstance(other, EntityType):
+            return self.name == other.name
+        return False
+    
+    def __hash__(self) -> int:
+        return hash(self.name)
+    
+    @classmethod
+    def from_string(cls, value: str) -> "EntityType":
+        """Create EntityType from string."""
+        return cls(value)
+
+
+@dataclass
+class HierarchyLevel:
+    """Represents a level in the hierarchy with its configuration."""
+    entity_type: EntityType
+    parent_type: Optional[EntityType]
+    level_order: int  # 0 = bottom level, higher numbers = higher levels
 
 
 @dataclass
@@ -154,7 +193,254 @@ class DemandPrediction:
         )
 
     def __str__(self) -> str:
-        return self.to_pretty()
+        return self.to_pretty(        )
+
+
+class Hierarchy:
+    """Generic hierarchical structure that can represent any organizational hierarchy.
+    
+    This class eliminates the need for specific methods for each entity type by using
+    a generic approach that works with any hierarchical structure.
+    
+    Attributes
+    ----------
+    levels : Dict[EntityType, HierarchyLevel]
+        Configuration of each level in the hierarchy
+    relationships : Dict[str, str]
+        Mapping from child_id to parent_id for all entities
+    entity_types : Dict[str, EntityType]
+        Mapping from entity_id to its EntityType
+    """
+    
+    def __init__(self, levels: List[HierarchyLevel]):
+
[…]
