### 2025-10-23 a5def25
add __repr__ to heirarchy object
