### 2025-10-23 77eb8eb
handle levels having same name
