+++ src/patientflow/predict/hierarchy.py
@@ -291,36 +291,112 @@ class Hierarchy:
         return cls(levels)
     
     def add_entity(self, entity_id: str, entity_type: EntityType, parent_id: Optional[str] = None):
-        """Add an entity to the hierarchy.
+        """Add an entity to the hierarchy with type-prefixed unique ID.
         
         Parameters
         ----------
         entity_id : str
-            Unique identifier for the entity
+            Original identifier for the entity (will be prefixed with entity type)
         entity_type : EntityType
             Type of the entity
         parent_id : str, optional
             Parent entity identifier. Required for all entities except top-level ones.
+            Should be the original entity name (will be converted to prefixed ID internally).
         """
         if entity_type not in self.levels:
             raise ValueError(f"Unknown entity type: {entity_type}")
         
         level = self.levels[entity_type]
         
+        # Create unique ID by prefixing with entity type
+        unique_id = f"{entity_type.name}:{entity_id}"
+        
+        # Convert parent_id to prefixed format if provided
+        prefixed_parent_id = None
+        if parent_id is not None:
+            # Find the parent's entity type to create prefixed parent ID
+            parent_entity_type = self._find_entity_type_by_name(parent_id)
+            if parent_entity_type is not None:
+                # Check if the found parent type matches the expected parent type
+                if parent_entity_type == level.parent_type:
+                    prefixed_parent_id = f"{parent_entity_type.name}:{parent_id}"
+                else:
+                    # If the found parent type doesn't match, assume it's already prefixed or will be created later
+                    prefixed_parent_id = parent_id
+            else:
+                # If parent not found, assume it's already prefixed or will be created later
+                prefixed_parent_id = parent_id
+        
         # Validate parent relationship
         if level.parent_type is None:
-            if parent_id is not None:
+            if prefixed_parent_id is not None:
                 raise ValueError(f"Top-level entity {entity_id} cannot have a parent")
         else:
             # For non-top-level entities, parent_id is optional during initial creation
             # but will be validated when parent_id is provided
-            if parent_id is not None and parent_id in self.entity_types:
-                parent_type = self.entity_types[parent_id]
+            if prefixed_parent_id is not None and prefixed_parent_id in self.entity_types:
+                parent_type = self.entity_types[prefixed_parent_id]
                 if parent_type != level.parent_type:
                     raise ValueError(f"Parent {parent_id} is of type {parent_type}, expected {level.parent_type}")
         
-        self.relationships[entity_id] = parent_id
-        self.entity_types[entity_id] = entity_type
+        self.relationships[unique_id] = prefixed_parent_id
+        self.entity_types[unique_id] = entity_type
+    
+    def _find_entity_type_by_name(self, entity_name: str) -> Optional[EntityType]:
+        """Find the entity type for a given entity name by searching through existing entities.
+        
+        Parameters
+        ----------
+        entity_name : str
+            Original entity name to search for
+            
+        Returns
+        -------
+        Optional[EntityType]
+            Entity type if found, None otherwise
+        """
+        for unique_id, entity_type in self.entity_types.items():
+            # Extract original name from prefixed ID
+            if ":" in unique_id:
+                original_name = unique_id.split(":", 1)[1]
+                if original_name == entity_name:
+                    return entity_type
+        return None
+    
+    def _get_original_name(self, unique_id: str) -> str:
[…]
