+++ tests/test_hierarchy.py
@@ -1,12 +1,16 @@
 """Tests for hierarchical demand prediction functionality."""
 
 import numpy as np
+import pandas as pd
 import pytest
 
 from patientflow.predict.hierarchy import (
     DemandPredictor,
     DemandPrediction,
     FlowSelection,
+    Hierarchy,
+    EntityType,
+    create_hierarchical_predictor,
     DEFAULT_PERCENTILES,
     DEFAULT_PRECISION,
     DEFAULT_MAX_PROBS,
@@ -790,3 +794,261 @@ class TestFlowSelectionEdgeCases:
         # Should raise KeyError for missing outflow keys
         with pytest.raises(KeyError, match="Missing outflow keys"):
             predictor.predict_subspecialty("cardio", inputs)
+
+
+class TestHierarchyCollisionFix:
+    """Test the hierarchy collision fix for entities with same names across levels."""
+
+    def test_entity_collision_fix(self):
+        """Test that entities with the same name can coexist at different levels."""
+        # Create test data with collision scenario
+        data = {
+            'sub_specialty': ['Cardiology', 'Neurology', 'Cardiology'],  # Same name as reporting unit
+            'reporting_unit': ['Cardiology', 'Neurology', 'Cardiology'],  # Same name as subspecialty
+            'division': ['Medicine', 'Medicine', 'Medicine'],
+            'board': ['Clinical Board', 'Clinical Board', 'Clinical Board']
+        }
+        
+        df = pd.DataFrame(data)
+        
+        # Column mapping
+        column_mapping = {
+            'sub_specialty': 'subspecialty',
+            'reporting_unit': 'reporting_unit', 
+            'division': 'division',
+            'board': 'board'
+        }
+        
+        # Create hierarchical predictor
+        predictor = create_hierarchical_predictor(
+            hierarchy_df=df,
+            column_mapping=column_mapping,
+            top_level_id="Hospital"
+        )
+        
+        # Test that both "Cardiology" entities exist but are different types
+        subspecialties = predictor.hierarchy.get_entities_by_type(EntityType("subspecialty"))
+        reporting_units = predictor.hierarchy.get_entities_by_type(EntityType("reporting_unit"))
+        
+        assert "Cardiology" in subspecialties
+        assert "Cardiology" in reporting_units
+        assert "Neurology" in subspecialties
+        assert "Neurology" in reporting_units
+        
+        # Test that they have different entity types
+        cardio_subspecialty_type = predictor.hierarchy.get_entity_type("Cardiology")
+        cardio_reporting_unit_type = predictor.hierarchy.get_entity_type("Cardiology")
+        
+        # Since we have two entities with the same name, we need to check which one is returned
+        # The get_entity_type method should find the first match
+        assert cardio_subspecialty_type is not None
+        
+        # Test entity info for both entities
+        cardio_info = predictor.hierarchy.get_entity_info("Cardiology")
+        assert cardio_info is not None
+        assert cardio_info["entity_id"] == "Cardiology"
+        assert cardio_info["entity_type"] is not None
+        
+        # Test that hierarchy structure is correct
+        assert len(subspecialties) == 2
+        assert len(reporting_units) == 2
+        
+        # Test that no collisions occurred in internal storage
+        all_entities = predictor.hierarchy.get_all_entities()
+        assert "Cardiology" in all_entities
+        assert "Neurology" in all_entities
+        
+        # Test parent-child relationships work correctly
+        hospital_children = predictor.hierarchy.get_children("Hospital")
+        assert len(hospital_children) > 0
+        
+        # Test that the hierarchy can be represented correctly
+        hierarchy_str = str(predictor.hierarchy)
+        assert "subspecialty: 2" in hierarchy_str
+        assert "reporting_unit: 2" in hierarchy_str
+
[…]
