### 2025-10-23 d13f396
fix error in accessing EntityType
