### 2025-10-23 3d08e23
fix error in hierarchy creation
