+++ src/patientflow/predict/hierarchy.py
@@ -290,8 +290,11 @@ class Hierarchy:
         ]
         return cls(levels)
     
-    def add_entity(self, entity_id: str, entity_type: EntityType, parent_id: Optional[str] = None):
+    def add_entity(self, entity_id: str, entity_type: EntityType):
         """Add an entity to the hierarchy with type-prefixed unique ID.
+        
+        This method is used in Pass 1 to create entities without relationships.
+        Relationships are established separately in Pass 2.
         
         Parameters
         ----------
@@ -299,47 +302,14 @@ class Hierarchy:
             Original identifier for the entity (will be prefixed with entity type)
         entity_type : EntityType
             Type of the entity
-        parent_id : str, optional
-            Parent entity identifier. Required for all entities except top-level ones.
-            Should be the original entity name (will be converted to prefixed ID internally).
         """
         if entity_type not in self.levels:
             raise ValueError(f"Unknown entity type: {entity_type}")
         
-        level = self.levels[entity_type]
-        
         # Create unique ID by prefixing with entity type
         unique_id = f"{entity_type.name}:{entity_id}"
         
-        # Convert parent_id to prefixed format if provided
-        prefixed_parent_id = None
-        if parent_id is not None:
-            # Find the parent's entity type to create prefixed parent ID
-            parent_entity_type = self._find_entity_type_by_name(parent_id)
-            if parent_entity_type is not None:
-                # Check if the found parent type matches the expected parent type
-                if parent_entity_type == level.parent_type:
-                    prefixed_parent_id = f"{parent_entity_type.name}:{parent_id}"
-                else:
-                    # If the found parent type doesn't match, assume it's already prefixed or will be created later
-                    prefixed_parent_id = parent_id
-            else:
-                # If parent not found, assume it's already prefixed or will be created later
-                prefixed_parent_id = parent_id
-        
-        # Validate parent relationship
-        if level.parent_type is None:
-            if prefixed_parent_id is not None:
-                raise ValueError(f"Top-level entity {entity_id} cannot have a parent")
-        else:
-            # For non-top-level entities, parent_id is optional during initial creation
-            # but will be validated when parent_id is provided
-            if prefixed_parent_id is not None and prefixed_parent_id in self.entity_types:
-                parent_type = self.entity_types[prefixed_parent_id]
-                if parent_type != level.parent_type:
-                    raise ValueError(f"Parent {parent_id} is of type {parent_type}, expected {level.parent_type}")
-        
-        self.relationships[unique_id] = prefixed_parent_id
+        # Store entity with its type
         self.entity_types[unique_id] = entity_type
     
     def _find_entity_type_by_name(self, entity_name: str) -> Optional[EntityType]:
@@ -1512,8 +1482,10 @@ class HierarchicalPredictor:
         # Level 1: Bottom level (e.g., subspecialties)
         for entity_id, inputs in bottom_level_data.items():
             bundle = self.predictor.predict_subspecialty(entity_id, inputs, flow_selection)
-            results[entity_id] = bundle
-            self.cache[entity_id] = bundle
+            # Use prefixed entity ID as key to avoid collisions
+            prefixed_entity_id = f"{bottom_type.name}:{entity_id}"
+            results[prefixed_entity_id] = bundle
+            self.cache[prefixed_entity_id] = bundle
         
         # Process each level from bottom to top
         for level_type in levels[1:]:
@@ -1521,14 +1493,24 @@ class HierarchicalPredictor:
             
             for entity_id in entities_at_level:
                 children = self.hierarchy.get_children(entity_id)
-                child_bundles = [results[child_id] for child_id in children]
+                # Convert child IDs to prefixed format for lookup
+                child_bundles = []
+                for child_id in children:
+                    # Find the child's entity type to create prefixed ID
[…]
