### 2025-10-23 137a066
truncate only bottom level
