### 2025-10-23 7c9868b
remove hard-coded logic in hierarcy
