### 2025-10-23 81991ee
fix error in entity type name retrieval
