### 2025-10-27 92b5a89
improve __repr__ for SubspecialtyPredictionInputs
