### 2025-10-27 888044a
try adaptive cap for truncation
