+++ tests/test_hierarchy.py
@@ -8,7 +8,6 @@ from patientflow.predict.hierarchy import (
     DemandPredictor,
     DemandPrediction,
     FlowSelection,
-    Hierarchy,
     EntityType,
     PredictionBundle,
     create_hierarchical_predictor,
@@ -48,31 +47,31 @@ class TestDemandPredictor:
         """Test that truncate_only_bottom parameter is properly set."""
         # Test default behavior (truncate at all levels)
         predictor_default = DemandPredictor()
-        assert predictor_default.truncate_only_bottom == False
-        
+        assert not predictor_default.truncate_only_bottom
+
         # Test truncate only at bottom level
         predictor_bottom_only = DemandPredictor(truncate_only_bottom=True)
-        assert predictor_bottom_only.truncate_only_bottom == True
+        assert predictor_bottom_only.truncate_only_bottom
 
     def test_truncation_behavior_difference(self):
         """Test that truncate_only_bottom affects convolution behavior."""
         # Create two predictors with different truncation settings
         predictor_all_levels = DemandPredictor(k_sigma=2.0, truncate_only_bottom=False)
         predictor_bottom_only = DemandPredictor(k_sigma=2.0, truncate_only_bottom=True)
-        
+
         # Create distributions that would be truncated differently
         p1 = np.array([0.1, 0.2, 0.4, 0.2, 0.1])  # 5 elements
         p2 = np.array([0.1, 0.2, 0.4, 0.2, 0.1])  # 5 elements
-        
+
         # Test _convolve_multiple with both settings
         result_all_levels = predictor_all_levels._convolve_multiple([p1, p2])
         result_bottom_only = predictor_bottom_only._convolve_multiple([p1, p2])
-        
+
         # The results should be different due to truncation behavior
         # With truncate_only_bottom=True, no truncation is applied in _convolve_multiple
         # With truncate_only_bottom=False, truncation is applied
         assert len(result_all_levels) <= len(result_bottom_only)
-        
+
         # Verify that the bottom-only version preserves more of the original distribution
         assert np.sum(result_bottom_only) > 0  # Should have valid probabilities
 
@@ -836,65 +835,74 @@ class TestHierarchyCollisionFix:
         """Test that entities with the same name can coexist at different levels."""
         # Create test data with collision scenario
         data = {
-            'sub_specialty': ['Cardiology', 'Neurology', 'Cardiology'],  # Same name as reporting unit
-            'reporting_unit': ['Cardiology', 'Neurology', 'Cardiology'],  # Same name as subspecialty
-            'division': ['Medicine', 'Medicine', 'Medicine'],
-            'board': ['Clinical Board', 'Clinical Board', 'Clinical Board']
+            "sub_specialty": [
+                "Cardiology",
+                "Neurology",
+                "Cardiology",
+            ],  # Same name as reporting unit
+            "reporting_unit": [
+                "Cardiology",
+                "Neurology",
+                "Cardiology",
+            ],  # Same name as subspecialty
+            "division": ["Medicine", "Medicine", "Medicine"],
+            "board": ["Clinical Board", "Clinical Board", "Clinical Board"],
         }
-        
+
         df = pd.DataFrame(data)
-        
+
         # Column mapping
         column_mapping = {
-            'sub_specialty': 'subspecialty',
-            'reporting_unit': 'reporting_unit', 
-            'division': 'division',
-            'board': 'board'
+            "sub_specialty": "subspecialty",
+            "reporting_unit": "reporting_unit",
+            "division": "division",
+            "board": "board",
         }
-        
+
         # Create hierarchical predictor
         predictor = create_hierarchical_predictor(
-            hierarchy_df=df,
-            column_mapping=column_mapping,
-            top_level_id="Hospital"
[…]
