### 2025-10-27 867ffcb
try adaptive cap for truncation
