### 2025-10-27 c730616
change default k_sigma and truncation_only_bottom values
