### 2025-10-27 4969808
change default k_sigma and truncation_only_bottom values
