### 2025-10-27 fc97794
update __repr__ for subspecialty
