+++ src/patientflow/predict/subspecialty.py
@@ -163,10 +163,21 @@ class SubspecialtyPredictionInputs:
 
     def __repr__(self) -> str:
         def format_pmf(
-            arr: np.ndarray, max_display: int = 10, total_count: Optional[int] = None
+            arr: np.ndarray, max_display: int = 10, total_count: Optional[int] = None, 
+            custom_bracket_text: Optional[str] = None
         ) -> str:
             expectation = np.sum(np.arange(len(arr)) * arr)
-            total_str = f" of {total_count}" if total_count is not None else ""
+            
+            # Use custom bracket text if provided, otherwise use total_count
+            if custom_bracket_text is not None:
+                if custom_bracket_text == "":
+                    total_str = ""
+                else:
+                    total_str = f" {custom_bracket_text}"
+            elif total_count is not None:
+                total_str = f" of {total_count}"
+            else:
+                total_str = ""
 
             if len(arr) <= max_display:
                 values = ", ".join(f"{v:.3f}" for v in arr)
@@ -192,7 +203,20 @@ class SubspecialtyPredictionInputs:
             if flow.flow_type == "pmf":
                 assert isinstance(flow.distribution, np.ndarray)
                 total_count = len(flow.distribution) - 1
-                return format_pmf(flow.distribution, total_count=total_count)
+                
+                # Customize bracket text based on flow type
+                custom_bracket_text = None
+                if flow.flow_id == "ed_current":
+                    custom_bracket_text = f"of {total_count} patients in ED"
+                elif flow.flow_id in ["elective_transfers", "emergency_transfers"]:
+                    # Remove 'of N' for transfers - just show expectation
+                    custom_bracket_text = ""
+                elif flow.flow_id == "emergency_departures":
+                    custom_bracket_text = f"of {total_count} emergency patients in subspec"
+                elif flow.flow_id == "elective_departures":
+                    custom_bracket_text = f"of {total_count} elective patients in subspec"
+                
+                return format_pmf(flow.distribution, total_count=total_count, custom_bracket_text=custom_bracket_text)
             elif flow.flow_type == "poisson":
                 return f"λ = {flow.distribution:.3f}"
             else:
