### 2025-11-11 ee6d099
add default for missing subgroup_to_spec mapping
