+++ src/patientflow/api/main.py
@@ -1,0 +1,295 @@
+"""Simple FastAPI application exposing cached hierarchical predictions."""
+
+from __future__ import annotations
+
+import logging
+import os
+import pickle
+from enum import Enum
+from pathlib import Path
+from typing import Any, Dict, Iterable, Optional
+
+from fastapi import FastAPI, HTTPException, Query
+
+from patientflow.predict.hierarchy import (
+    DemandPrediction,
+    EntityType,
+    HierarchicalPredictor,
+    PredictionBundle,
+)
+
+LOGGER = logging.getLogger(__name__)
+
+app = FastAPI(title="PatientFlow Hierarchical Predictions")
+
+_predictor: Optional[HierarchicalPredictor] = None
+
+
+class FlowType(str, Enum):
+    arrivals = "arrivals"
+    departures = "departures"
+    net_flow = "net_flow"
+    all = "all"
+
+
+class Cohort(str, Enum):
+    all = "all"
+    elective = "elective"
+    emergency = "emergency"
+
+
+class DetailLevel(str, Enum):
+    summary = "summary"
+    full = "full"
+
+
+def _load_predictor() -> HierarchicalPredictor:
+    """Load the pickled ``HierarchicalPredictor`` using environment configuration."""
+    path_value = os.getenv("PREDICTOR_PICKLE_PATH")
+    if not path_value:
+        raise RuntimeError("PREDICTOR_PICKLE_PATH environment variable is not set")
+
+    path = Path(path_value).expanduser()
+    if not path.is_file():
+        raise RuntimeError(f"PREDICTOR_PICKLE_PATH does not point to a file: {path}")
+
+    try:
+        with path.open("rb") as file:
+            predictor = pickle.load(file)
+    except Exception as exc:  # pragma: no cover - defensive logging
+        LOGGER.exception("Failed to load predictor pickle from %s", path)
+        raise RuntimeError("Unable to load predictor pickle") from exc
+
+    if not isinstance(predictor, HierarchicalPredictor):
+        raise RuntimeError(
+            "Loaded object is not a HierarchicalPredictor; check the pickle content"
+        )
+
+    LOGGER.info(
+        "Loaded HierarchicalPredictor with %s cached bundles",
+        len(predictor.cache),
+    )
+    return predictor
+
+
+def _get_predictor() -> HierarchicalPredictor:
+    if _predictor is None:
+        raise RuntimeError("Predictor not initialised")
+    return _predictor
+
+
+@app.on_event("startup")
+def _on_startup() -> None:
+    global _predictor
+    _predictor = _load_predictor()
+
+
+def _validate_entity_type(predictor: HierarchicalPredictor, entity_type: str) -> None:
+    available_types = predictor.hierarchy.get_entity_type_names()
+    if entity_type not in available_types:
+        raise HTTPException(
+            status_code=404,
+            detail={
+                "message": f"Unknown entity type '{entity_type}'",
+                "available_types": available_types,
+            },
+        )
+
+
+def _prefixed_id(entity_type: str, entity_id: str) -> str:
+    return f"{entity_type}:{entity_id}"
+
+
+def _get_bundle(entity_type: str, entity_id: str) -> PredictionBundle:
+    predictor = _get_predictor()
+    _validate_entity_type(predictor, entity_type)
+
+    cache_key = _prefixed_id(entity_type, entity_id)
+    bundle = predictor.cache.get(cache_key)
+    if bundle is None:
+        raise HTTPException(
+            status_code=404,
+            detail={"message": f"No prediction cached for {entity_type}/{entity_id}"},
+        )
+    return bundle
+
+
+def _serialize_prediction(
+    prediction: DemandPrediction, detail: DetailLevel
+) -> Dict[str, Any]:
+    payload: Dict[str, Any] = {
[…]
