+++ pyproject.toml
@@ -1,115 +1,117 @@
 [build-system]
 build-backend = "setuptools.build_meta"
 requires = [
-    "setuptools",
-    "setuptools-scm",
+  "setuptools",
+  "setuptools-scm",
 ]
 
 [dependency-groups]
 dev = [
-    "patientflow",
+  "patientflow",
 ]
 
 [project]
 authors = [
-    {email = "zella.king@ucl.ac.uk", name = "Zella King"},
+  {email = "zella.king@ucl.ac.uk", name = "Zella King"},
 ]
 classifiers = [
-    "Operating System :: POSIX",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Typing :: Typed",
+  "Operating System :: POSIX",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Typing :: Typed",
 ]
 dependencies = [
-    "PyYAML==6.0.2",
-    "joblib>=1.4.2",
-    "matplotlib>=3.1.3",
-    "numpy>=1.18.1",
-    "pandas>=1.0.1",
-    "requests",
-    "scikit-learn>=1.4.0",
-    "scipy>=1.4.1",
-    "seaborn>=0.10.0", # "shap>=0.42.1",
-    "sympy>=1.12",
-    "tabulate==0.9.0",
-    "xgboost>=2.0.3",
+  "fastapi>=0.111.0",
+  "PyYAML==6.0.2",
+  "joblib>=1.4.2",
+  "matplotlib>=3.1.3",
+  "numpy>=1.18.1",
+  "pandas>=1.0.1",
+  "requests",
+  "scikit-learn>=1.4.0",
+  "scipy>=1.4.1",
+  "seaborn>=0.10.0", # "shap>=0.42.1",
+  "sympy>=1.12",
+  "tabulate==0.9.0",
+  "xgboost>=2.0.3",
+  "uvicorn[standard]>=0.30.0",
 ]
 description = "Code and explanatory notebooks for predicting short-term hospital bed capacity using real-time data"
 dynamic = [
-    "version",
+  "version",
 ]
 keywords = []
+license.file = "LICENSE.md"
 name = "patientflow"
 optional-dependencies = {dev = [
-    "build",
-    "mypy",
-    "pre-commit",
-    "ruff",
-    "tox",
-    "twine",
+  "build",
+  "mypy",
+  "pre-commit",
+  "ruff",
+  "tox",
+  "twine",
 ], docs = [
-    "griffe",
-    "mkdocs",
-    "mkdocs-autorefs",
-    "mkdocs-awesome-pages-plugin",
-    "mkdocs-include-markdown-plugin",
-    "mkdocs-material",
-    "mkdocstrings",
-    "mkdocstrings-python",
-    "nbconvert",
+  "griffe",
+  "mkdocs",
+  "mkdocs-autorefs",
+  "mkdocs-awesome-pages-plugin",
+  "mkdocs-include-markdown-plugin",
+  "mkdocs-material",
+  "mkdocstrings",
+  "mkdocstrings-python",
+  "nbconvert",
 ], test = [
-    "pytest",
-    "pytest-cov",
-    "pyyaml",
+  "pytest",
+  "pytest-cov",
+  "pyyaml",
 ]}
 readme = "README.md"
 requires-python = ">=3.10"
[…]
