### 2025-11-17 ac02164
modify API to receive elective, emergency and all flows
