+++ src/patientflow/api/main.py
@@ -1,4 +1,36 @@
-"""Simple FastAPI application exposing cached hierarchical predictions."""
+"""Simple FastAPI application exposing cached hierarchical predictions.
+
+Usage
+=====
+1. Build a pickle file that contains a dictionary mapping cohort names to
+   ``HierarchicalPredictor`` instances. Example structure::
+
+       {
+           "all": HierarchicalPredictor(...),
+           "elective": HierarchicalPredictor(...),
+           "emergency": HierarchicalPredictor(...),
+       }
+
+   Each predictor should already have its cache populated, typically by calling
+   ``predict_all_levels`` after constructing the predictor.
+
+2. Point the API to that pickle via the ``PREDICTOR_PICKLE_PATH`` environment
+   variable before starting uvicorn, e.g.::
+
+       export PREDICTOR_PICKLE_PATH=/path/to/hierarchical_predictors.pkl
+       uvicorn patientflow.api.main:app
+
+3. Call the HTTP endpoints with the desired cohort. For example::
+
+       GET /api/predictions/subspecialty/Cardiology?cohort=elective
+
+   The API selects the correct predictor dictionary entry and returns the
+   cached arrivals/departures/net-flow bundle for that entity.
+
+   cURL example::
+
+       curl "http://localhost:8000/api/predictions/subspecialty/Cardiology?cohort=all&flow_type=arrivals"
+"""
 
 from __future__ import annotations
 
@@ -7,22 +39,26 @@ import os
 import pickle
 from enum import Enum
 from pathlib import Path
-from typing import Any, Dict, Iterable, Optional
+import numpy as np
+from typing import Any, Dict, Iterable
 
 from fastapi import FastAPI, HTTPException, Query
 
 from patientflow.predict.hierarchy import (
+    DEFAULT_MAX_PROBS,
+    # DEFAULT_PRECISION,
     DemandPrediction,
     EntityType,
     HierarchicalPredictor,
     PredictionBundle,
 )
 
+DEFAULT_PRECISION = 2
 LOGGER = logging.getLogger(__name__)
 
 app = FastAPI(title="PatientFlow Hierarchical Predictions")
 
-_predictor: Optional[HierarchicalPredictor] = None
+_predictors: Dict[str, HierarchicalPredictor] = {}
 
 
 class FlowType(str, Enum):
@@ -43,8 +79,8 @@ class DetailLevel(str, Enum):
     full = "full"
 
 
-def _load_predictor() -> HierarchicalPredictor:
-    """Load the pickled ``HierarchicalPredictor`` using environment configuration."""
+def _load_predictors() -> Dict[str, HierarchicalPredictor]:
+
     path_value = os.getenv("PREDICTOR_PICKLE_PATH")
     if not path_value:
         raise RuntimeError("PREDICTOR_PICKLE_PATH environment variable is not set")
@@ -55,33 +91,25 @@ def _load_predictor() -> HierarchicalPredictor:
 
     try:
         with path.open("rb") as file:
-            predictor = pickle.load(file)
+            predictors = pickle.load(file)
     except Exception as exc:  # pragma: no cover - defensive logging
         LOGGER.exception("Failed to load predictor pickle from %s", path)
         raise RuntimeError("Unable to load predictor pickle") from exc
 
-    if not isinstance(predictor, HierarchicalPredictor):
-        raise RuntimeError(
-            "Loaded object is not a HierarchicalPredictor; check the pickle content"
-        )
-
-    LOGGER.info(
-        "Loaded HierarchicalPredictor with %s cached bundles",
-        len(predictor.cache),
-    )
-    return predictor
+    LOGGER.info("Loaded %s cohort predictors", len(predictors))
+    return predictors
 
 
-def _get_predictor() -> HierarchicalPredictor:
-    if _predictor is None:
[…]
