### 2025-11-20 15f8a7d
apply top down truncation of distributions
