### 2025-11-20 e8b2d6e
apply top down truncation of distributions
