+++ tests/test_hierarchy.py
@@ -9,6 +9,7 @@ from patientflow.predict.hierarchy import (
     DemandPrediction,
     FlowSelection,
     EntityType,
+    Hierarchy,
     PredictionBundle,
     create_hierarchical_predictor,
     DEFAULT_PERCENTILES,
@@ -43,37 +44,122 @@ class TestDemandPrediction:
 class TestDemandPredictor:
     """Test the DemandPredictor class."""
 
-    def test_truncate_only_bottom_parameter(self):
-        """Test that truncate_only_bottom parameter is properly set."""
-        # Test default behavior (truncate only at bottom level)
-        predictor_default = DemandPredictor()
-        assert predictor_default.truncate_only_bottom
+    def test_apply_cap_with_renormalization_truncates_tail(self):
+        """Truncating with a cap should fold remaining mass into the last bin."""
+        predictor = DemandPredictor()
+        pmf = np.array([0.2, 0.3, 0.1, 0.4])
 
-        # Test truncate at all levels
-        predictor_all_levels = DemandPredictor(truncate_only_bottom=False)
-        assert not predictor_all_levels.truncate_only_bottom
+        truncated = predictor._apply_cap_with_renormalization(pmf, 2)
 
-    def test_truncation_behavior_difference(self):
-        """Test that truncate_only_bottom affects convolution behavior."""
-        # Create two predictors with different truncation settings
-        predictor_all_levels = DemandPredictor(k_sigma=2.0, truncate_only_bottom=False)
-        predictor_bottom_only = DemandPredictor(k_sigma=2.0, truncate_only_bottom=True)
+        assert len(truncated) == 3
+        assert pytest.approx(truncated.sum(), rel=1e-9) == 1.0
+        # Last bin should contain its own mass plus the tail.
+        assert pytest.approx(truncated[2], rel=1e-9) == 0.5
 
-        # Create distributions that would be truncated differently
-        p1 = np.array([0.1, 0.2, 0.4, 0.2, 0.1])  # 5 elements
-        p2 = np.array([0.1, 0.2, 0.4, 0.2, 0.1])  # 5 elements
+    def test_apply_cap_with_renormalization_negative_cap(self):
+        """Negative caps should collapse the PMF into a single bucket."""
+        predictor = DemandPredictor()
+        pmf = np.array([0.2, 0.3])
 
-        # Test _convolve_multiple with both settings
-        result_all_levels = predictor_all_levels._convolve_multiple([p1, p2])
-        result_bottom_only = predictor_bottom_only._convolve_multiple([p1, p2])
+        truncated = predictor._apply_cap_with_renormalization(pmf, -1)
 
-        # The results should be different due to truncation behavior
-        # With truncate_only_bottom=True, no truncation is applied in _convolve_multiple
-        # With truncate_only_bottom=False, truncation is applied
-        assert len(result_all_levels) <= len(result_bottom_only)
+        assert len(truncated) == 1
+        assert pytest.approx(truncated[0], rel=1e-9) == 0.5
 
-        # Verify that the bottom-only version preserves more of the original distribution
-        assert np.sum(result_bottom_only) > 0  # Should have valid probabilities
+    def test_predict_flow_total_respects_max_support(self):
+        """predict_flow_total should respect the provided max_support cap."""
+        predictor = DemandPredictor()
+        flows = [
+            FlowInputs(
+                flow_id="pmf_flow",
+                flow_type="pmf",
+                distribution=np.array([0.6, 0.4]),
+            ),
+            FlowInputs(flow_id="pois_flow", flow_type="poisson", distribution=3.0),
+        ]
+
+        prediction = predictor.predict_flow_total(
[…]
