+++ src/patientflow/predict/hierarchy.py
@@ -24,7 +24,7 @@ create_hierarchical_predictor
 import numpy as np
 import pandas as pd
 import yaml
-from typing import List, Dict, Optional, Any
+from typing import List, Dict, Optional, Any, Tuple
 from dataclasses import dataclass
 
 
@@ -836,22 +836,17 @@ class DemandPredictor:
     Parameters
     ----------
     k_sigma : float, default=8.0
-        Cap width measured in standard deviations. Final (and intermediate)
-        distributions are hard-clipped using an adaptive approach that prevents
-        over-truncation for small lambda values. For lambda < 0.1, uses 2x k_sigma;
-        for lambda < 1.0, uses 1.5x k_sigma; otherwise uses the original k_sigma.
-        Net-flow uses asymmetric caps around the mean with the same adaptive
-        k_sigma multiplier and physical bounds.
-    truncate_only_bottom : bool, default=True
-        If True, only apply truncation at the bottom level (subspecialties).
-        If False, apply truncation at all hierarchical levels.
+        Cap width measured in standard deviations. Maximum support for each
+        hierarchical level is calculated top-down using statistical caps:
+        mean + k_sigma * sqrt(sum of variances). This ensures bounded array
+        sizes while maintaining statistical accuracy. Net flow is computed
+        naively from already-capped arrivals and departures, with no additional
+        truncation.
 
     Attributes
     ----------
     k_sigma : float
         Number of standard deviations used to cap supports
-    truncate_only_bottom : bool
-        Whether to truncate only at bottom level or all levels
     cache : dict
         Cache for storing computed predictions (currently unused)
 
@@ -865,8 +860,9 @@ class DemandPredictor:
     4. Computing statistics (expected value, percentiles) for each level
 
     The class uses discrete convolution to combine probability distributions.
-    Supports are clamped using adaptive k-sigma caps to prevent exponential
-    growth in array sizes while avoiding over-truncation for small lambda values.
+    Supports are clamped using top-down statistical caps calculated before
+    convolution, ensuring bounded array sizes while maintaining statistical
+    accuracy. Truncation uses renormalization to preserve probability mass.
 
     Flow Selection
     --------------
@@ -875,52 +871,202 @@ class DemandPredictor:
     cohorts (elective/emergency/all) to include in predictions.
     """
 
-    def __init__(self, k_sigma: float = 8.0, truncate_only_bottom: bool = True):
+    def __init__(self, k_sigma: float = 8.0):
         self.k_sigma = k_sigma
-        self.truncate_only_bottom = truncate_only_bottom
         self.cache: Dict[str, DemandPrediction] = {}
 
-    def adaptive_cap(self, lam: float, k_sigma: float) -> int:
-        """Calculate adaptive cap to prevent over-truncation for small lambda values.
+    def _calculate_hierarchical_stats(
+        self,
+        entity_id: str,
+        entity_type: EntityType,
+        bottom_level_data: Dict[str, SubspecialtyPredictionInputs],
+        hierarchy: "Hierarchy",
+        flow_type: str,
+        flow_selection: Optional[FlowSelection] = None,
+    ) -> Tuple[float, float, int]:
+        """Calculate sum of means, combined SD, and maximum support for an entity.
 
-        This method uses higher k_sigma multipliers for small lambda values to avoid
-        over-truncation that could occur with the standard k-sigma approach.
+        This method recursively gathers all bottom-level entities in the subtree
+        and calculates statistical caps based on the sum of means and combined
+        variance of all distributions in the subtree.
 
         Parameters
         ----------
-        lam : float
-            Lambda parameter (mean) of the distribution
-        k_sigma : float
-            Base k_sigma value for cap calculation
+        entity_id : str
+            Unique identifier for the entity
+        entity_type : EntityType
+            Type of entity being analyzed
[…]
