+++ tests/test_k_sigma_sensitivity.py
@@ -1,0 +1,318 @@
+"""Test sensitivity of prediction generation time to k_sigma parameter.
+
+This script benchmarks the time taken to generate hierarchical predictions
+across a range of k_sigma values. It loads pickled HierarchicalPredictor objects
+and measures performance metrics for each k_sigma value.
+
+To run it: "uv run python tests/test_k_sigma_sensitivity.py /path/to/file/hierarchical_predictors_with_inputs.pkl 10 50 100"
+"""
+
+import pickle
+import time
+from typing import Dict, Any, List, Optional
+import numpy as np
+import pandas as pd
+
+from patientflow.predict.hierarchy import (
+    HierarchicalPredictor,
+    FlowSelection,
+    create_hierarchical_predictor,
+)
+from patientflow.predict.subspecialty import SubspecialtyPredictionInputs
+
+
+def load_pickled_data(pickle_path: str) -> Dict[str, Dict[str, Any]]:
+    """Load pickled data containing predictors and associated metadata.
+    
+    Parameters
+    ----------
+    pickle_path : str
+        Path to the pickle file
+        
+    Returns
+    -------
+    dict[str, dict]
+        Dictionary with keys 'all', 'elective', 'emergency', where each value
+        contains: 'predictor', 'subspecialty_data', 'hierarchy_df', 
+        'column_mapping', 'top_level_id'
+    """
+    with open(pickle_path, 'rb') as f:
+        data = pickle.load(f)
+    return data
+
+
+def extract_hierarchy_info(
+    predictor: HierarchicalPredictor
+) -> tuple[Any, Optional[FlowSelection]]:
+    """Extract hierarchy information from a HierarchicalPredictor.
+    
+    Parameters
+    ----------
+    predictor : HierarchicalPredictor
+        The predictor to extract info from
+        
+    Returns
+    -------
+    tuple
+        (hierarchy, flow_selection) - hierarchy object and flow selection
+    """
+    hierarchy = predictor.hierarchy
+    # Get flow_selection from first cached bundle
+    flow_selection = None
+    if predictor.cache:
+        first_bundle = next(iter(predictor.cache.values()))
+        flow_selection = first_bundle.flow_selection
+    
+    return hierarchy, flow_selection
+
+
+def measure_prediction_time(
+    predictor: HierarchicalPredictor,
+    bottom_level_data: Dict[str, SubspecialtyPredictionInputs],
+    flow_selection: FlowSelection,
+    n_iterations: int = 10,
+) -> tuple[float, Dict[str, Any]]:
+    """Measure time to generate predictions across multiple iterations.
+    
+    Parameters
+    ----------
+    predictor : HierarchicalPredictor
+        Configured hierarchical predictor
+    bottom_level_data : dict[str, SubspecialtyPredictionInputs]
+        Bottom-level prediction inputs
+    flow_selection : FlowSelection
+        Flow selection configuration
+    n_iterations : int, default=10
+        Number of iterations to run for timing
+        
+    Returns
+    -------
+    tuple
+        (mean_time_seconds, metrics_dict)
+        metrics_dict contains timing statistics and PMF size information
+    """
+    times = []
+    pmf_lengths = []
+    
+    for i in range(n_iterations):
+        # Clear cache to ensure fresh computation
+        predictor.cache.clear()
+        
+        start = time.perf_counter()
+        results = predictor.predict_all_levels(
+            bottom_level_data,
+            flow_selection=flow_selection
+        )
+        end = time.perf_counter()
+        
+        elapsed = end - start
+        times.append(elapsed)
+        
+        # Collect PMF length metrics on first iteration
+        if i == 0:
[…]
