### 2025-11-24 eab83f8
update docs following JOSS review
