### 2025-11-28 49f42c6
wording changes to aggregate.py docstrings
