### 2025-11-28 UCL-CORU/patientflow#111 @zmek
Input validation using schemas (pydantic/dataclasses) to safeguard snapshot construction
### What needs to be done?

Suggested by @dmeghana24 in [JOSS review](https://github.com/openjournals/joss-reviews/issues/9187#issuecomment-3567065611)

With schemas (Pydantic/dataclasses + validation):

- The user gets early, explicit failures when inputs don’t match expectations.
- Improves robustness, reproducibility and safety
