### 2025-11-28 UCL-CORU/patientflow#112 @zmek
Documentation changes and references to examples
### What needs to be done?

Suggestions from JOSS reviewers

- [ ] Clarify in docs when aggregation switches between exact and normal approximation for large sample sizes; suggestion from @dmeghana24 [here](https://github.com/openjournals/joss-reviews/issues/9187#issuecomment-3393598180)
- [ ] Add explicit reference to examples in JOSS paper; suggestion from @limengbinggz [here](https://github.com/openjournals/joss-reviews/issues/9187#issuecomment-3420215735)
