### 2025-11-28 9826648
add backticks to equations
