### 2025-11-28 #9187 @zmek
Regarding the suggestions made by [dmeghana24](https://github.com/dmeghana24):

### Consider adding Windows CI tests or noting OS limitations more prominently.

I have updated a pre-existing [issue regarding windows compatibility](https://github.com/UCL-CORU/patientflow/issues/48) with a task list for what needs to be done. We will work on this as an enhancement to the package in due course.

### A Dockerfile or container recipe would enhance reproducibility for future users.

I have noted this suggestion from dmeghana24 in a [pre-existing issue proposing a Dockerfile](https://github.com/UCL-CORU/patientflow/issues/66) as a potential enhancement that would help people using our synthetic or public datasets. While JOSS requires automated dependency management (which our package provides), a standalone Dockerfile is not required. In our hospital implementation, patientflow is itself a dependency in a Dockerised application that extracts snapshots from the electronic patient record and formats and sends predictions to users of the output.

I### Input validation using schemas (pydantic/dataclasses) could further safeguard snapshot construction.

This is a very good suggestion and would definitely improve the experience and the error handling for users. I have created an issue for this as a potential enhancement, but we take it from @dmeghana24 that we don't need to implement this as a requirement for acceptance by JOSS.

### A CHANGELOG.md summarizing releases (not just GitHub Releases) would aid transparency.

We agree completely with the intent of making releases as clear as possible, but are concerned that having a separate CHANGELOG.md risks these two sources of information getting out of sync. We feel that the Github Release page is the best source and will endeavour to provide detailed release notes, as can be seen [here](https://github.com/UCL-CORU/patientflow/releases), in all future releases.

### Clarify in docs when aggregation switches between exact and normal approximation for large sample sizes.

This is a very useful suggestion. We have implemented a change in the current version of the API documentation for the patientflow.aggregate module [here](https://ucl-coru.github.io/patientflow/api/#patientflow.aggregate).

We are extremely grateful to @dmeghana24 and @limengbinggz for the time they have taken to review our package and for their excellent suggestions.

@majensen I hope you will agree that we have addressed all of their suggestions, but please let me know if you would like to see any further amendments.
