### 2025-11-30 b37dfd8
update notebooks for MultiSubgroupPredictor
