+++ docs/notebooks/2a_Create_patient_snapshots.md
@@ -9,32 +9,34 @@
 - Group snapshot: A set of patient snapshots. Each group snapshot has a date and a prediction time associated with it.
 - Prediction window: A period of hours that begins at the prediction time.
 
-To use `patientflow` your data should be in snapshot form.
+To use `patientflow` your data should be in snapshot form. 
 
 In this notebook I suggest how to you might prepare your data, starting from data on finished hospital visits. I start with fake data on Emergency Department visits, and demonstrate how to convert it into snapshots. There are two examples
 
-- A simple example of creating snapshots assuming you have one flat table of hospital visits
-- An example of creating snapshots from data structured as a relational database.
+* A simple example of creating snapshots assuming you have one flat table of hospital visits
+* An example of creating snapshots from data structured as a relational database. 
 
 ## A note on creating your own shapshots
 
-The snapshot creation shown here is designed to work with fake data generated below. You would need to create your own version of this process, to handle the data you have.
+The snapshot creation shown here is designed to work with fake data generated below. You would need to create your own version of this process, to handle the data you have. 
 
-In practice, determining from data _whether a patient was admitted after the ED visit_, and _when they were ready to be admitted_, can be tricky. How do you account for the fact that the patient may wait in the ED for a bed, due to lack of available beds? Likewise, if you are trying to predict discharge at the end of a hospital visit, should that that be the time they were ready to leave, or the time they actually left? Discharge delays are common, due to waiting for medication or transport, or waiting for onward care provision to become available.
+In practice, determining from data *whether a patient was admitted after the ED visit*, and *when they were ready to be admitted*, can be tricky. How do you account for the fact that the patient may wait in the ED for a bed, due to lack of available beds? Likewise, if you are trying to predict discharge at the end of a hospital visit, should that that be the time they were ready to leave, or the time they actually left? Discharge delays are common, due to waiting for medication or transport, or waiting for onward care provision to become available. 
 
-The outcome that you are aiming for will depend on your setting, and the information needs of the bed managers you are looking to support. You may have to infer when a patient was ready from available data. Suffice to say, think carefully about what it is you are trying to predict, and how you will identify that outcome in data.
+The outcome that you are aiming for will depend on your setting, and the information needs of the bed managers you are looking to support. You may have to infer when a patient was ready from available data. Suffice to say, think carefully about what it is you are trying to predict, and how you will identify that outcome in data. 
 
 ## Creating fake finished visits
 
-I'll start by loading some fake data resembling the structure of EHR data on Emergency Department (ED) visits, using a function called `create_fake_finished_visits`. In my fake data, each visit has one row, with an arrival time at the ED, a discharge time from the ED, the patient's age and an outcome of whether they were admitted after the ED visit.
+I'll start by loading some fake data resembling the structure of EHR data on Emergency Department (ED) visits, using a function called `create_fake_finished_visits`. In my fake data, each visit has one row, with an arrival time at the ED, a discharge time from the ED, the patient's age and an outcome of whether they were admitted after the ED visit. 
 
-The `is_admitted` column is our label, indicating the outcome in this imaginary case.
[…]
