+++ docs/notebooks/3d_Predict_bed_counts_for_subgroups.md
@@ -1,25 +1,24 @@
 # 3d. Predict bed count distributions for subgroups
 
-It is often the case that pressures build up on certain areas of the hospital, or for subgroups of patients, due to random fluctuations. Sometimes one demographic group seems to be showing up more than usual, such as older males, putting pressure on male geriatric beds.
+It is often the case that pressures build up on certain areas of the hospital, or for subgroups of patients, due to random fluctuations. Sometimes one demographic group seems to be showing up more than usual, such as older males, putting pressure on male geriatric beds. 
 
-In this notebook I show how `patientflow` can be used to create predictions for subgroups of patients. These groups might be
+In this notebook I show how `patientflow` can be used to create predictions for subgroups of patients. These groups might be 
+* specific clinical areas of the hospital (eg medical or paediatric beds)
+* subgroups defined by a demographic characteristic (eg sex) 
 
-- specific clinical areas of the hospital (eg medical or paediatric beds)
-- subgroups defined by a demographic characteristic (eg sex)
-
-### Predicting which subgroup a patient will belong to
+### Predicting which subgroup a patient will belong to 
 
 My focus initially is on subgroups defined by the clinical area patients are admitted to after ED; I refer to the clinical areas as specialties.
 
 I demonstrate the use of a `SequenceToOutcomePredictor` class, that can be used to predict each patient's probability of admission to a specialty if they are admitted. I load real patient data and show how the `SequenceToOutcomePredictor` is trained using sequences of consult requests made while patients are in the ED.
 
-The `SequenceToOutcomePredictor` approach could be used with other sequence data, such as sequences of locations or procedures, if you deem these likely to be associated with a patient being admitted to a clinical area. The key assumption of the sequence design is that the order is meaningful; for example, a surgical consult following a medical consult, or vice versa, is meaningful for the patient's likelihood of being admitted under surgery.
+The `SequenceToOutcomePredictor` approach could be used with other sequence data, such as sequences of locations or procedures, if you deem these likely to be associated with a patient being admitted to a clinical area. The key assumption of the sequence design is that the order is meaningful; for example, a surgical consult following a medical consult, or vice versa, is meaningful for the patient's likelihood of being admitted under surgery. 
 
-If you don't have sequences for predicting sub-groups, you might have more simple data, such as a single reason code for each visit, entered on triage (eg heart problem, broken bone), that suggests which specialty a patient will end up in. I demonstrate a `ValueToOutcomePredictor` which can be used with such data.
+If you don't have sequences for predicting sub-groups, you might have more simple data, such as a single reason code for each visit, entered on triage (eg heart problem, broken bone), that suggests which specialty a patient will end up in. I demonstrate a `ValueToOutcomePredictor` which can be used with such data. 
 
 ### Combining specialty prediction with admission prediction
 
-I then combine the specialty prediction model with the admission probability model (shown in previous notebooks) to calculate the joint probability that a patient will both be admitted and require a specific specialty. Formally, this derives P(admitted AND specialty X) = P(admitted) × P(specialty X | admitted). This joint probability approach means we can generate specialty-specific bed count predictions. I demonstrate the joint probability for one group snapshot.
[…]

[…]
