### 2025-11-30 bdfb0df
update notebooks for MultiSubgroupPredictor
