+++ docs/notebooks/3c_Predict_bed_counts_without_using_patient_snapshots.md
@@ -2,38 +2,40 @@
 
 There are situations where we might want to predict bed count distributions without having details of the patients, for example when predicting the number of beds needed for incoming patients such as:
 
-- Incoming patients who haven't yet arrived at the Emergency Department, but who will need a bed within a prediction window
-- Emergency patients who arrive via other routes than the ED, and become inpatients (such as emergency transfers from other hospitals)
-- Elective admissions of patients. Planned elective admissions are recorded on a 'To Come In' (TCI) list by patient number. In data, the hospital visit (known as an encounter) for the elective procedure begins at the moment a patient arrives, but may not be recorded in data prior to that point. In a simple case without making use of any data on TCI lists, we might want to predict bed requirements for such patients based on past patterns of such arrivals between a prediction time (eg 12:00) and the end of a prediction window (eg 8 hours later).
+* Incoming patients who haven't yet arrived at the Emergency Department, but who will need a bed within a prediction window
+* Emergency patients who arrive via other routes than the ED, and become inpatients (such as emergency transfers from other hospitals)
+* Elective admissions of patients. Planned elective admissions are recorded on a 'To Come In' (TCI) list by patient number. In data, the hospital visit (known as an encounter) for the elective procedure begins at the moment a patient arrives, but may not be recorded in data prior to that point. In a simple case without making use of any data on TCI lists, we might want to predict bed requirements for such patients based on past patterns of such arrivals between a prediction time (eg 12:00) and the end of a prediction window (eg 8 hours later).  
 
 For these situations, we can use `patientflow` to learn patterns from past data, and use these to predict a bed count distribution at the aggregate level.
 
-In this notebook, I'll use the example of predicting the number of beds needed for incoming patients (patients yet to arrive to the Emergency Department who will need a bed within a prediction window). I'll show three approaches:
+In this notebook, I'll use the example of predicting the number of beds needed for incoming patients (patients yet to arrive to the Emergency Department who will need a bed within a prediction window). I'll show three approaches: 
 
-- a Poisson model trained on past arrival rates of patients who both arrived and were admitted within a prediction window
-- a weighted Poisson model using an empirical survival curve; arrival rates of patients who were admitted (at some point), are weighted by their probability of being admitted within a prediction window, calculated from a survival curve learned form past data
-- a weighted Poisson model using an aspirational approach; instead of using a survival curve learned form past data, it is assumed that the ED is meeting 4-hour targets for time to admission.
+* a Poisson model trained on past arrival rates of patients who both arrived and were admitted within a prediction window
+* a weighted Poisson model using an empirical survival curve; arrival rates of patients who were admitted (at some point), are weighted by their probability of being admitted within a prediction window, calculated from a survival curve learned form past data
+* a weighted Poisson model using an aspirational approach; instead of using a survival curve learned form past data, it is assumed that the ED is meeting 4-hour targets for time to admission. 
 
-I also demonstrate making predictions by specialty for demand from incoming patients.
+I also demonstrate making predictions by specialty for demand from incoming patients. 
+
 
 ```python
 # Reload functions every time
-%load_ext autoreload
+%load_ext autoreload 
 %autoreload 2
 ```
 
[…]
