+++ docs/notebooks/2c_Evaluate_patient_snapshot_models.md
@@ -1,42 +1,47 @@
 # 2c. Evaluate models trained on patient-level snapshots
 
-## Things to consider
+## Things to consider 
 
-In the last notebook, I showed how to train models on patient snapshots using `patientflow`. Now, let's think about how to evaluate those models.
+In the last notebook, I showed how to train models on patient snapshots using `patientflow`. Now, let's think about how to evaluate those models. 
 
 When evaluating patient snapshots, we focus on:
 
-- How well calibrated the predicted probabilities are.
-- How well the probabilities discriminate between patients with and without the outcome.
+* How well calibrated the predicted probabilities are.
+* How well the probabilities discriminate between patients with and without the outcome.
 
-We don't focus as much on typical classification metrics like Area under the ROC curve, accuracy or precision/recall.
+We don't focus as much on typical classification metrics like Area under the ROC curve, accuracy or precision/recall. 
 
-### Why don't we focus on typical classification metrics?
+### Why don't we focus on typical classification metrics? 
 
 The ultimate goal is to predict bed count distributions for groups of patients. Bed count distributions will be calculated in two steps
 
 1. First, we predict the probability of the outcome we are interested in (admission or discharge) for each individual patient, as shown in previous notebooks.
 2. Then, we use these probabilities in Bernoulli trials to get bed count distributions. The Bernouill trials step will be shown in later notebooks.
 
-Because of this approach, the accuracy of the probability values matters more than correct classification. That is why we use log loss to optimise our classifiers.
+Because of this approach, the accuracy of the probability values matters more than correct classification. That is why we use log loss to optimise our classifiers. 
 
 ### About the data used in this notebook
 
 I'm going to use real patient data from visits to the Emergency Department (ED) and Same Day Emergency Care (SDEC) unit at University College London Hospital (UCLH) to demonstrate the evaluation. For more information about the data, see [the data exploration notebook](2d_Explore_the_datasets_provided.md).
 
-The methods shown will work on any data in the same structure.
+The methods shown will work on any data in the same structure. 
 
 You can request the datasets that are used here on [Zenodo](https://zenodo.org/records/14866057). Alternatively you can use the synthetic data that has been created from the distributions of real patient data. If you don't have the public data, change the argument in the cell below from `data_folder_name='data-public'` to `data_folder_name='data-synthetic'`.
 
+
+
+
+
 ```python
 # Reload functions every time
-%load_ext autoreload
+%load_ext autoreload 
 %autoreload 2
 ```
 
 ## Loading real patient data
 
-I load the data using a `load_data` function that will sort the data and return the tuple columns as tuples rather than strings or lists. If you run the cell below without the public dataset, you will need to change the `data_folder_name` or (better, since it will solve the problem for all notebooks) copy the synthetic data from `data-synthetic` to `data-public`.
+I load the data using a `load_data` function that will sort the data and return the tuple columns as tuples rather than strings or lists. If you run the cell below without the public dataset, you will need to change the `data_folder_name` or (better, since it will solve the problem for all notebooks) copy the synthetic data from `data-synthetic` to `data-public`. 
+
 
 ```python
 from patientflow.load import set_file_paths, load_data
@@ -47,23 +52,25 @@ project_root = set_project_root()
 
 # set file paths
 data_file_path, media_file_path, model_file_path, config_path = set_file_paths(
-        project_root,
+        project_root, 
[…]
