+++ docs/notebooks/4d_Predict_emergency_demand_with_special_categories.md
@@ -1,20 +1,22 @@
 # 4d. Evaluate predictions with special categories
 
-In the previous notebook I demonstrated how we evaluate the models used at UCLH. As a final step, I now show the same implementation in code, but including extra functionality to handle certain sub-groups differently from others.
+In the previous notebook I demonstrated how we evaluate the models used at UCLH. As a final step, I now show the same implementation in code, but including extra functionality to handle certain sub-groups differently from others. 
 
-At UCLH, it is standard practice to admit paediatric patients (defined as patients under 18 on the day of arrival at the ED) to paediatric wards, and not to admit adult patients (18 or over) to paediatric wards.
+At UCLH, it is standard practice to admit paediatric patients (defined as patients under 18 on the day of arrival at the ED) to paediatric wards, and not to admit adult patients (18 or over) to paediatric wards. 
 
-The two models that enable prediction by sub-groups (specialty of admission, and yet-to-arrive by specialty) offer parameters that allow you to specify that certain groups are handled differently. In the UCLH example, this means disregarding any consult requests for patients in the ED when predicting which specialty they will be admitted to, and counting all yet-to-arrive patients under 18 as paediatric admissions.
+The two models that enable prediction by sub-groups (specialty of admission, and yet-to-arrive by specialty) offer parameters that allow you to specify that certain groups are handled differently. In the UCLH example, this means disregarding any consult requests for patients in the ED when predicting which specialty they will be admitted to, and counting all yet-to-arrive patients under 18 as paediatric admissions. 
 
-Most of the code below is the same as in the previous notebook. I limit the narrative here to pointing out how the special sub-groups are handled.
+Most of the code below is the same as in the previous notebook. I limit the narrative here to pointing out how the special sub-groups are handled. 
 
 ## Set up the notebook environment
 
+
 ```python
 # Reload functions every time
-%load_ext autoreload
+%load_ext autoreload 
 %autoreload 2
 ```
+
 
 ```python
 from patientflow.load import set_project_root
@@ -24,11 +26,13 @@ project_root = set_project_root()
 
     Inferred project root: /Users/zellaking/Repos/patientflow
 
+
 ## Set file paths and load data
 
-I'm going to use real patient data from UCLH to demonstrate the implementation.
+I'm going to use real patient data from UCLH to demonstrate the implementation. 
 
 As noted previously, you can request the datasets that are used here on [Zenodo](https://zenodo.org/records/14866057). Alternatively you can use the synthetic data that has been created from the distributions of real patient data. If you don't have the public data, change the argument in the cell below from `data_folder_name='data-public'` to `data_folder_name='data-synthetic'`.
+
 
 ```python
 from patientflow.load import set_file_paths
@@ -38,25 +42,26 @@ data_folder_name = 'data-public'
 data_file_path = project_root / data_folder_name
 
 data_file_path, media_file_path, model_file_path, config_path = set_file_paths(
-    project_root,
+    project_root, 
     data_folder_name=data_folder_name,
     config_file = 'config.yaml', verbose=False)
[…]

[…]
