### 2025-12-02 3cf5771
update notebook markdown
