+++ docs/notebooks/2d_Explore_the_datasets_provided.md
@@ -1,16 +1,18 @@
 # 2d. Explore the datasets provided
 
-Two datasets have been provided with this repository. 
-- `ed_visits.csv` 
+Two datasets have been provided with this repository.
+
+- `ed_visits.csv`
 - `inpatient_arrivals.csv`
 
-These accompany my fully worked example of the modelling of emergency demand for beds. And they are also useful to illustrate what patient snapshots might be made up of. 
+These accompany my fully worked example of the modelling of emergency demand for beds. And they are also useful to illustrate what patient snapshots might be made up of.
 
-This notebook does some data exploration by plotting charts of all relevant variables in each dataset. 
+This notebook does some data exploration by plotting charts of all relevant variables in each dataset.
 
 The `inpatient_arrivals` dataset contains arrival times of all patients who visited the UCLH Emergency Department (ED) and the Same Day Emergency Care (SDEC) unit, over the period of the data, and were later admitted. It includes their sex, child status (whether adult or child), and which specialty they were admitted to.
 
-The `ed_visits` database contains a set of snapshots of patients who visited the ED and SDEC over the period of the data, including both admitted and discharged patients. Each snapshot includes information known at the time of the snapshot, and excludes anything that was recorded later, except the variables that serve a 'labels' for model training. These are:   
+The `ed_visits` database contains a set of snapshots of patients who visited the ED and SDEC over the period of the data, including both admitted and discharged patients. Each snapshot includes information known at the time of the snapshot, and excludes anything that was recorded later, except the variables that serve a 'labels' for model training. These are:
+
 - `is_admitted` - whether the visit ended in admission to a ward
 - `final_sequence` - the sequence of consultations the patient had during the visit
 - `specialty` - the specialty of the admission, if the patient was admitted
@@ -19,28 +21,22 @@ See the [data dictionaries](https://github.com/UCL-CORU/patientflow/tree/main/da
 
 ## Learn more about the data
 
-I recorded a webinar to demonstrate how we converted data from the UCLH Electronic Health Record in a form suitable for this modelling. If you click on the image below, the video will open at the point where I provide detail about the datasets 
+I recorded a webinar to demonstrate how we converted data from the UCLH Electronic Health Record in a form suitable for this modelling. If you click on the image below, the video will open at the point where I provide detail about the datasets
 
 <a href="https://www.youtube.com/watch?v=ha_zckz3_rU&t=262s" target="_blank">
     <img src="img/thumbnail_NHSR_webinar.jpg" alt="Link to webinar on how to turn your EHR data into predictions of demand for emergency beds" width="600"/>
 </a>
 
-
-
 ## Set up the notebook environment
-
-
 
 ```python
 # Reload functions every time
-%load_ext autoreload 
+%load_ext autoreload
 %autoreload 2
 ```
 
     The autoreload extension is already loaded. To reload it, use:
       %reload_ext autoreload
-
-
 
 ```python
 from patientflow.load import set_project_root
@@ -49,52 +45,45 @@ project_root = set_project_root()
 
     Inferred project root: /Users/zellaking/Repos/patientflow
 
-
 ## Load parameters and set file paths
 
[…]

[…]
