### 2025-12-02 7daed4e
update notebook markdown
