+++ docs/notebooks/2b_Predict_using_patient_snapshots.md
@@ -1,6 +1,6 @@
 # 2b. Make predictions using patient-level snapshots
 
-Now that the data have been prepared in snapshot form, we have a dataset of unfinished visits. The ultimate goal is to make predictions about whether an outcome of interest (e.g. admission, discharge) will happen within a prediction window. For now, I will use a simple outcome variable of admission or not, without worrying about when that admission happened. 
+Now that the data have been prepared in snapshot form, we have a dataset of unfinished visits. The ultimate goal is to make predictions about whether an outcome of interest (e.g. admission, discharge) will happen within a prediction window. For now, I will use a simple outcome variable of admission or not, without worrying about when that admission happened.
 
 Everything shown here is standard modelling, but there are some important considerations when working with unfinished hospital visits.
 
@@ -12,36 +12,30 @@ When dividing your data into training, validation and test sets, a random alloca
 
 **Multiple snapshots per visit**
 
-To use `patientflow` your data should be in snapshot form. I showed how to create this in the last notebook. I defined a series of prediction times, and then sampled finished visits to get snapshots that represent those visits while still in progress. When you follow this method, you may end up with multiple snapshots per patient visit. Is this OK, for your analysis? You will need to decide whether you include all snapshots from a single visit into the training, validation or testing of a predictive model. Snapshots from the same visit are inherently correlated, which may violate assumptions of the statistical or machine learning methods you are using. For this reason, we chose to sample only one snapshot per patient visit. 
+To use `patientflow` your data should be in snapshot form. I showed how to create this in the last notebook. I defined a series of prediction times, and then sampled finished visits to get snapshots that represent those visits while still in progress. When you follow this method, you may end up with multiple snapshots per patient visit. Is this OK, for your analysis? You will need to decide whether you include all snapshots from a single visit into the training, validation or testing of a predictive model. Snapshots from the same visit are inherently correlated, which may violate assumptions of the statistical or machine learning methods you are using. For this reason, we chose to sample only one snapshot per patient visit.
 
 **Multiple visits per patient**
 
-The patient identifier is also important, because if the same patient appears in training and test sets, there is the potential for data leakage. We took the decision to probabilistically allocate each patient to training, validation or test set, where the probability of being allocated to each set is in proportion to the number of visits they made in any of those time periods. 
+The patient identifier is also important, because if the same patient appears in training and test sets, there is the potential for data leakage. We took the decision to probabilistically allocate each patient to training, validation or test set, where the probability of being allocated to each set is in proportion to the number of visits they made in any of those time periods.
 
 `patientflow` includes functions to handle all of these considerations. I demonstrate them here.
 
-
-
 ```python
 # Reload functions every time
-%load_ext autoreload 
+%load_ext autoreload
 %autoreload 2
 ```
 
 ## Create fake snapshots
 
-See the previous notebook for more information about how this is done. 
-
+See the previous notebook for more information about how this is done.
 
 ```python
 from patientflow.generate import create_fake_snapshots
-prediction_times = [(6, 0), (9, 30), (12, 0), (15, 30), (22, 0)] 
+prediction_times = [(6, 0), (9, 30), (12, 0), (15, 30), (22, 0)]
 snapshots_df=create_fake_snapshots(prediction_times=prediction_times, start_date='2023-01-01', end_date='2023-04-01')
 snapshots_df.head()
 ```
-
-
-
 
 <div>
 <style scoped>
@@ -56,6 +50,7 @@ snapshots_df.head()
[…]
