+++ docs/notebooks/4b_Predict_emergency_demand.md
@@ -1,16 +1,14 @@
 # 4b. Predict emergency demand
 
-This notebook demonstrates the full implementation in code. I show how we used these building blocks in our application at University College London Hospital (UCLH) to predict emergency demand for beds, by specialty, over the next 8 hours. The predictions are aspirational; they assume that ED four-hour targets are met. 
+This notebook demonstrates the full implementation in code. I show how we used these building blocks in our application at University College London Hospital (UCLH) to predict emergency demand for beds, by specialty, over the next 8 hours. The predictions are aspirational; they assume that ED four-hour targets are met.
 
 ## Set up the notebook environment
 
-
 ```python
 # Reload functions every time
-%load_ext autoreload 
+%load_ext autoreload
 %autoreload 2
 ```
-
 
 ```python
 from patientflow.load import set_project_root
@@ -20,13 +18,11 @@ project_root = set_project_root()
 
     Inferred project root: /Users/zellaking/Repos/patientflow
 
-
 ## Set file paths and load data
 
-I'm going to use real patient data from UCLH to demonstrate the implementation. 
+I'm going to use real patient data from UCLH to demonstrate the implementation.
 
 You can request the datasets that are used here on [Zenodo](https://zenodo.org/records/14866057). Alternatively you can use the synthetic data that has been created from the distributions of real patient data. If you don't have the public data, change the argument in the cell below from `data_folder_name='data-public'` to `data_folder_name='data-synthetic'`.
-
 
 ```python
 from patientflow.load import set_file_paths
@@ -36,26 +32,25 @@ data_folder_name = 'data-public'
 data_file_path = project_root / data_folder_name
 
 data_file_path, media_file_path, model_file_path, config_path = set_file_paths(
-    project_root, 
+    project_root,
     data_folder_name=data_folder_name,
     config_file = 'config.yaml', verbose=False)
 ```
-
 
 ```python
 import pandas as pd
 from patientflow.load import load_data
 
 # load ED snapshots data
-ed_visits = load_data(data_file_path, 
-                    file_name='ed_visits.csv', 
+ed_visits = load_data(data_file_path,
+                    file_name='ed_visits.csv',
                     index_column = 'snapshot_id',
-                    sort_columns = ["visit_number", "snapshot_date", "prediction_time"], 
+                    sort_columns = ["visit_number", "snapshot_date", "prediction_time"],
                     eval_columns = ["prediction_time", "consultation_sequence", "final_sequence"])
 ed_visits.snapshot_date = pd.to_datetime(ed_visits.snapshot_date).dt.date
 
 # load data on inpatient arrivals
-inpatient_arrivals = inpatient_arrivals = load_data(data_file_path, 
+inpatient_arrivals = inpatient_arrivals = load_data(data_file_path,
                     file_name='inpatient_arrivals.csv')
 inpatient_arrivals['arrival_datetime'] = pd.to_datetime(inpatient_arrivals['arrival_datetime'], utc = True)
 
@@ -65,7 +60,6 @@ inpatient_arrivals['arrival_datetime'] = pd.to_datetime(inpatient_arrivals['arri
 ## Set modelling parameters
 
 The parameters are used in training or inference. They are set in config.json in the root of the repository and loaded by `load_config_file()`
-
 
 ```python
 # load params
@@ -79,38 +73,31 @@ print(f'Validation set starts on {start_validation_set} and ends on {start_test_
 print(f'Test set starts on {start_test_set} and ends on {end_test_set- pd.Timedelta(days=1)} inclusive' )
 ```
 
-    
     Training set starts 2031-03-01 and ends on 2031-08-31 inclusive
     Validation set starts on 2031-09-01 and ends on 2031-09-30 inclusive
[…]

[…]
