### 2025-12-02 aa78e6a
update notebook markdown
