+++ docs/notebooks/4c_Evaluate_emergency_demand_predictions.md
@@ -1,29 +1,28 @@
 # 4c. Evaluate predictions of emergency demand
 
-The previous notebook demonstrated the full implementation in code of training models for use at UCLH using the functions provided in `patientflow`. I demonstrated how we create predictions for a single snapshot. 
+The previous notebook demonstrated the full implementation in code of training models for use at UCLH using the functions provided in `patientflow`. I demonstrated how we create predictions for a single snapshot.
 
 This notebooks show evaluate the predicted distributions against observed numbers, for the whole test set. I will evaluate the predictions by specialty using the approaches covered in a [previous notebook](3b_Evaluate_group_snapshots.md).
 
-* Histograms of observed versus expected values
-* Adjusted QQ plots
+- Histograms of observed versus expected values
+- Adjusted QQ plots
 
 As the predictions for yet-to-arrive patients are aspirational, these cannot be directed evaluated against observed numbers of admissions in the prediction window. In reality, due to poor ED performance, few may have been admitted within the window. Similarly for the group of patients in the ED, we calculate the predicted number of beds needed within the prediction window, but the observed numbers will not reflect the targets.
 
-We can, however, evaluate the predictions in a slightly different way. 
+We can, however, evaluate the predictions in a slightly different way.
 
-* For the patients in ED, we can compare the predicted bed counts needed for each specialty against observed numbers admitted to each specialty from among patients comprising each group snapshot in the test set period, without taking into account how long it takes each patient to be admitted.
-* For the yet-to-arrive patients, we can compare the predicted with the observed arrival rates within the prediction window.
+- For the patients in ED, we can compare the predicted bed counts needed for each specialty against observed numbers admitted to each specialty from among patients comprising each group snapshot in the test set period, without taking into account how long it takes each patient to be admitted.
+- For the yet-to-arrive patients, we can compare the predicted with the observed arrival rates within the prediction window.
 
 ## Loading data and parameters
-### Set up the notebook environment
 
+### Set up the notebook environment
 
 ```python
 # Reload functions every time
-%load_ext autoreload 
+%load_ext autoreload
 %autoreload 2
 ```
-
 
 ```python
 from patientflow.load import set_project_root
@@ -33,13 +32,11 @@ project_root = set_project_root()
 
     Inferred project root: /Users/zellaking/Repos/patientflow
 
-
 ### Set file paths and load data
 
-I'm going to use real patient data from UCLH to demonstrate the implementation. 
+I'm going to use real patient data from UCLH to demonstrate the implementation.
 
 As noted previously, you can request the datasets that are used here on [Zenodo](https://zenodo.org/records/14866057). Alternatively you can use the synthetic data that has been created from the distributions of real patient data. If you don't have the public data, change the argument in the cell below from `data_folder_name='data-public'` to `data_folder_name='data-synthetic'`.
-
 
 ```python
 from patientflow.load import set_file_paths
@@ -49,26 +46,25 @@ data_folder_name = 'data-public'
 data_file_path = project_root / data_folder_name
 
 data_file_path, media_file_path, model_file_path, config_path = set_file_paths(
-    project_root, 
+    project_root,
     data_folder_name=data_folder_name,
     config_file = 'config.yaml', verbose=False)
 ```
-
 
 ```python
 import pandas as pd
 from patientflow.load import load_data
 
 # load ED snapshots data
-ed_visits = load_data(data_file_path, 
-                    file_name='ed_visits.csv', 
[…]
