### 2025-12-02 af668fa
results from k_sigma_sensitivity_test
