+++ k_sigma_sensitivity_results.txt
@@ -1,0 +1,51 @@
+/Users/zellaking/Repos/patientflow/tests/test_k_sigma_sensitivity.py:41: DeprecationWarning: numpy.core.numeric is deprecated and has been renamed to numpy._core.numeric. The numpy._core namespace contains private NumPy internals and its use is discouraged, as NumPy internals can change without warning in any release. In practice, most real-world usage of numpy.core is to access functionality in the public NumPy API. If that is the case, use the public NumPy API. If not, you are using NumPy internals. If you would still like to access an internal attribute, use numpy._core.numeric._frombuffer.
+  data = pickle.load(f)
+Loading pickled data from /Users/zellaking/Downloads/hierarchical_predictors_with_inputs.pkl...
+Using FlowSelection: all
+Found 291 subspecialties
+Top level ID: uclh
+
+Testing k_sigma = 2.0...
+  Mean time: 0.2092s
+  Total PMF elements: 23002
+
+Testing k_sigma = 4.0...
+  Mean time: 0.2151s
+  Total PMF elements: 24040
+
+Testing k_sigma = 6.0...
+  Mean time: 0.2295s
+  Total PMF elements: 25072
+
+Testing k_sigma = 8.0...
+  Mean time: 0.2259s
+  Total PMF elements: 26082
+
+Testing k_sigma = 10.0...
+  Mean time: 0.2290s
+  Total PMF elements: 27108
+
+Testing k_sigma = 12.0...
+  Mean time: 0.2398s
+  Total PMF elements: 28142
+
+================================================================================
+K_SIGMA SENSITIVITY TEST RESULTS
+================================================================================
+
+k_sigma    Mean Time (s)   Std Time (s)    Total PMF Elements  
+--------------------------------------------------------------------------------
+2.00       0.2092          0.0066          23002               
+4.00       0.2151          0.0042          24040               
+6.00       0.2295          0.0149          25072               
+8.00       0.2259          0.0119          26082               
+10.00      0.2290          0.0040          27108               
+12.00      0.2398          0.0089          28142               
+
+--------------------------------------------------------------------------------
+Speedup relative to smallest k_sigma:
+  k_sigma=2.00 vs k_sigma=4.00: 0.97x
+  k_sigma=2.00 vs k_sigma=6.00: 0.91x
+  k_sigma=2.00 vs k_sigma=8.00: 0.93x
+  k_sigma=2.00 vs k_sigma=10.00: 0.91x
+  k_sigma=2.00 vs k_sigma=12.00: 0.87x
