### 2025-12-05 4dcc16d
correct lengths of pmfs - cap departure pmf at physical max - check that departure input type is always pmf and never poisson - return pmf of length 1 when zero patients exist in a specialty
