### 2025-12-05 41e83ab
add detailed logging to k_sigma test
