+++ tests/test_k_sigma_sensitivity.py
@@ -112,6 +112,9 @@ def measure_prediction_time(
         # Collect PMF length metrics on first iteration
         if i == 0:
             for entity_id, bundle in results.items():
+                # Calculate maximum value (not just length)
+                arrivals_max = len(bundle.arrivals.probabilities) - 1 if len(bundle.arrivals.probabilities) > 0 else 0
+                departures_max = len(bundle.departures.probabilities) - 1 if len(bundle.departures.probabilities) > 0 else 0
                 pmf_lengths.append({
                     'entity_id': entity_id,
                     'arrivals_pmf_len': len(bundle.arrivals.probabilities),
@@ -119,6 +122,8 @@ def measure_prediction_time(
                     'net_flow_pmf_len': len(bundle.net_flow.probabilities),
                     'arrivals_expected': bundle.arrivals.expected_value,
                     'departures_expected': bundle.departures.expected_value,
+                    'arrivals_max': arrivals_max,
+                    'departures_max': departures_max,
                 })
     
     mean_time = np.mean(times)
@@ -139,6 +144,9 @@ def measure_prediction_time(
             m['arrivals_pmf_len'] + m['departures_pmf_len'] + m['net_flow_pmf_len']
             for m in pmf_lengths
         ),
+        'arrivals_pmf_elements': sum(m['arrivals_pmf_len'] for m in pmf_lengths),
+        'departures_pmf_elements': sum(m['departures_pmf_len'] for m in pmf_lengths),
+        'net_flow_pmf_elements': sum(m['net_flow_pmf_len'] for m in pmf_lengths),
     }
     
     return mean_time, metrics
@@ -237,13 +245,165 @@ def test_k_sigma_sensitivity(
             n_iterations=n_iterations,
         )
         
+        # Analyze individual flows from input data to see how physical_maxes accumulate
+        if k_sigma == k_sigma_values[0]:  # Only do this detailed analysis once
+            print(f"\n  Analyzing individual departure flows from input data:")
+            elective_totals = []
+            emergency_totals = []
+            
+            for spec_id, inputs in subspecialty_data.items():
+                # Get individual flows
+                elective_flow = inputs.outflows.get("elective_departures")
+                emergency_flow = inputs.outflows.get("emergency_departures")
+                
+                if elective_flow and elective_flow.flow_type == "pmf":
+                    pmf_array = elective_flow.distribution
+                    if isinstance(pmf_array, np.ndarray):
+                        physical_max = len(pmf_array) - 1
+                        elective_totals.append(physical_max)
+                
+                if emergency_flow and emergency_flow.flow_type == "pmf":
+                    pmf_array = emergency_flow.distribution
+                    if isinstance(pmf_array, np.ndarray):
+                        physical_max = len(pmf_array) - 1
+                        emergency_totals.append(physical_max)
+            
+            total_elective_max = sum(elective_totals)
+            total_emergency_max = sum(emergency_totals)
+            total_individual_max = total_elective_max + total_emergency_max
+            
+            print(f"    Sum of elective_departures physical_maxes: {total_elective_max}")
+            print(f"    Sum of emergency_departures physical_maxes: {total_emergency_max}")
+            print(f"    Total (elective + emergency): {total_individual_max}")
+            print(f"    Expected (total inpatients): 305")
+            print(f"    Difference: {total_individual_max - 305}")
+            
+            # Show distribution
+            elective_dist = {}
+            for val in elective_totals:
+                elective_dist[val] = elective_dist.get(val, 0) + 1
+            
+            emergency_dist = {}
+            for val in emergency_totals:
[…]
