### 2025-12-05 c000357
update departures test to respect new pmf rules
