### 2025-12-06 6897d06
script to build and save one-off examples of hierarchical predictor
