### 2025-12-09 #9187 @majensen
@zmek This is a clear and succinct paper, and IMO an unusual and useful tool. 2 minor suggestions:

l. 45: "fake and synthetic data" - I would just say 'synthetic' unless their is a distinction you have in mind.

l. 74 (ref): "Npj" -> "NPJ"

Once the archive is up I will add that info to the ticket and start the final review process (with one of our Assoc. Editors in Chief)
