### 2025-12-09 9addea0
update paper for JOSS
