+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{king2022machine,
     author = {King, Z and Farrington, J. and Utley, M. and others},
     title = {Machine learning for real-time aggregated prediction of hospital admission for emergency patients},
-    journal = {npj Digital Medicine},
+    journal = {NPJ Digital Medicine},
     volume = {5},
     number = {104},
     year = {2022},
