### 2025-12-09 #9187 @zmek
@majensen Many thanks for your speedy response.

I have created a permanent archive on Zenodo for the reviewed software version (v1.0.3), which is associated with the final DOI 10.5281/zenodo.15723614

Also I have

- updated the title of the software in that DOI to match the title of the paper
- made the changes you suggested
- updated the affiliation of one of the authors, who was shown in the incorrect unit
- removed the generated paper/paper.pdf file from the repository
