### 2025-12-10 4b60589
fix to k sigma sensitivity test
