+++ tests/test_k_sigma_sensitivity.py
@@ -226,24 +226,133 @@ def test_k_sigma_sensitivity(
     # Test each k_sigma value
     results = {}
     
+    # Process all flow types
+    flow_types = ['all', 'elective', 'emergency']
+    flow_type_data = {}
+    for flow_type in flow_types:
+        if flow_type in pickled_data:
+            flow_type_data[flow_type] = pickled_data[flow_type]
+    
     for k_sigma in k_sigma_values:
         print(f"\nTesting k_sigma = {k_sigma}...")
         
-        # Create predictor with this k_sigma value
-        hierarchical_predictor = create_hierarchical_predictor(
-            hierarchy_df,
-            column_mapping,
-            top_level_id,
-            k_sigma=k_sigma,
-        )
+        # Process each flow type
+        flow_type_results = {}
         
-        # Measure prediction time
-        mean_time, metrics = measure_prediction_time(
-            hierarchical_predictor,
-            subspecialty_data,
-            flow_selection,
-            n_iterations=n_iterations,
-        )
+        for flow_type in flow_types:
+            if flow_type not in flow_type_data:
+                continue
+                
+            print(f"\n  Flow type: {flow_type}")
+            flow_data = flow_type_data[flow_type]
+            flow_subspecialty_data = flow_data['subspecialty_data']
+            flow_hierarchy_df = flow_data['hierarchy_df']
+            flow_column_mapping = flow_data['column_mapping']
+            flow_top_level_id = flow_data['top_level_id']
+            
+            # Extract flow_selection from the predictor
+            flow_predictor = flow_data['predictor']
+            _, flow_flow_selection = extract_hierarchy_info(flow_predictor)
+            
+            if flow_flow_selection is None:
+                flow_flow_selection = FlowSelection.default()
+            
+            # Create predictor with this k_sigma value
+            hierarchical_predictor = create_hierarchical_predictor(
+                flow_hierarchy_df,
+                flow_column_mapping,
+                flow_top_level_id,
+                k_sigma=k_sigma,
+            )
+            
+            # Measure prediction time
+            mean_time, metrics = measure_prediction_time(
+                hierarchical_predictor,
+                flow_subspecialty_data,
+                flow_flow_selection,
+                n_iterations=n_iterations,
+            )
+            
+            flow_type_results[flow_type] = {
+                'mean_time': mean_time,
+                **metrics
+            }
+            
+            # Group PMF metrics by hierarchy level
+            level_sums = {
+                'subspecialty': {'arrivals': 0, 'departures': 0, 'net_flow': 0, 'count': 0},
+                'reporting_unit': {'arrivals': 0, 'departures': 0, 'net_flow': 0, 'count': 0},
+                'division': {'arrivals': 0, 'departures': 0, 'net_flow': 0, 'count': 0},
+                'board': {'arrivals': 0, 'departures': 0, 'net_flow': 0, 'count': 0},
+                'hospital': {'arrivals': 0, 'departures': 0, 'net_flow': 0, 'count': 0},
+            }
+            
+            for pmf_info in metrics['pmf_metrics']:
+                entity_id = pmf_info['entity_id']
+                # Extract level from entity_id prefix (e.g., "subspecialty:..." -> "subspecialty")
+                if ':' in entity_id:
+                    level = entity_id.split(':')[0]
+                    if level in level_sums:
+                        level_sums[level]['arrivals'] += pmf_info['arrivals_pmf_len']
+                        level_sums[level]['departures'] += pmf_info['departures_pmf_len']
+                        level_sums[level]['net_flow'] += pmf_info['net_flow_pmf_len']
+                        level_sums[level]['count'] += 1
+            
[…]
