+++ tests/test_k_sigma_sensitivity.py
@@ -184,7 +184,6 @@ def test_k_sigma_sensitivity(
         Dictionary mapping k_sigma values to their timing/metrics results
     """
     # Load pickled data
-    print(f"Loading pickled data from {pickle_path}...")
     pickled_data = load_pickled_data(pickle_path)
     
     if 'all' not in pickled_data:
@@ -216,12 +215,6 @@ def test_k_sigma_sensitivity(
     
     if flow_selection is None:
         flow_selection = FlowSelection.default()
-        print("Using default FlowSelection (could not extract from predictor)")
-    else:
-        print(f"Using FlowSelection: {flow_selection.cohort}")
-    
-    print(f"Found {len(subspecialty_data)} subspecialties")
-    print(f"Top level ID: {top_level_id}")
     
     # Test each k_sigma value
     results = {}
@@ -234,8 +227,6 @@ def test_k_sigma_sensitivity(
             flow_type_data[flow_type] = pickled_data[flow_type]
     
     for k_sigma in k_sigma_values:
-        print(f"\nTesting k_sigma = {k_sigma}...")
-        
         # Process each flow type
         flow_type_results = {}
         
@@ -243,7 +234,6 @@ def test_k_sigma_sensitivity(
             if flow_type not in flow_type_data:
                 continue
                 
-            print(f"\n  Flow type: {flow_type}")
             flow_data = flow_type_data[flow_type]
             flow_subspecialty_data = flow_data['subspecialty_data']
             flow_hierarchy_df = flow_data['hierarchy_df']
@@ -297,36 +287,6 @@ def test_k_sigma_sensitivity(
                         level_sums[level]['departures'] += pmf_info['departures_pmf_len']
                         level_sums[level]['net_flow'] += pmf_info['net_flow_pmf_len']
                         level_sums[level]['count'] += 1
-            
-            # Display PMF length sums by level
-            print(f"    PMF Length Sums by Hierarchy Level:")
-            print(f"    {'Level':<20} {'Count':<10} {'Arrivals':<12} {'Departures':<12} {'Net Flow':<12}")
-            print(f"    {'-'*20} {'-'*10} {'-'*12} {'-'*12} {'-'*12}")
-            for level in ['subspecialty', 'reporting_unit', 'division', 'board', 'hospital']:
-                if level_sums[level]['count'] > 0:
-                    print(f"    {level:<20} {level_sums[level]['count']:<10} "
-                          f"{level_sums[level]['arrivals']:<12} {level_sums[level]['departures']:<12} "
-                          f"{level_sums[level]['net_flow']:<12}")
-            
-            # Print departures PMF length for each subspecialty
-            subspecialty_departures = []
-            for pmf_info in metrics['pmf_metrics']:
-                entity_id = pmf_info['entity_id']
-                if entity_id.startswith('subspecialty:'):
-                    subspecialty_name = entity_id.replace('subspecialty:', '')
-                    subspecialty_departures.append({
-                        'name': subspecialty_name,
-                        'departures_pmf_len': pmf_info['departures_pmf_len']
-                    })
-            
-            # Sort by subspecialty name
-            subspecialty_departures.sort(key=lambda x: x['name'])
-            
-            print(f"\n    Departures PMF Length for Each Subspecialty:")
-            print(f"    {'Subspecialty':<60} {'Departures PMF Length':<25}")
-            print(f"    {'-'*60} {'-'*25}")
-            for sub in subspecialty_departures:
-                print(f"    {sub['name']:<60} {sub['departures_pmf_len']:<25}")
         
         # Store results for all flow types
         results[k_sigma] = {
@@ -354,167 +314,11 @@ def test_k_sigma_sensitivity(
         if analysis_flow_selection is None:
[…]
