### 2025-12-11 d89e97c
use child level type to avoid name collisions
