### 2025-12-11 9173ee1
pass departures cap to bottom level processing
