### 2025-12-11 76f82d5
calculate departures physical cap from child pmf len
