### 2025-12-11 1c6b50b
track truncated mass and save
