+++ src/patientflow/predict/hierarchy.py
@@ -24,7 +24,7 @@ create_hierarchical_predictor
 import numpy as np
 import pandas as pd
 import yaml
-from typing import List, Dict, Optional, Any, Tuple
+from typing import List, Dict, Optional, Any, Tuple, Union
 from dataclasses import dataclass
 
 
@@ -874,6 +874,9 @@ class DemandPredictor:
     def __init__(self, k_sigma: float = 8.0):
         self.k_sigma = k_sigma
         self.cache: Dict[str, DemandPrediction] = {}
+        # Track truncated mass: {(entity_id, flow_type): truncated_mass}
+        # flow_type is 'arrivals' or 'departures'
+        self.truncated_mass: Dict[Tuple[str, str], float] = {}
 
     def _calculate_hierarchical_stats(
         self,
@@ -1146,8 +1149,13 @@ class DemandPredictor:
 
         # Apply cap with renormalization if max_support is provided
         if max_support is not None:
-            truncated_pmf = self._apply_cap_with_renormalization(
-                dist_total.probabilities, max_support
+            truncated_pmf, truncated_mass = self._apply_cap_with_renormalization(
+                dist_total.probabilities, max_support, return_truncated_mass=True
+            )
+            # Track truncated mass for this entity and flow type
+            key = (entity_id, entity_type)
+            self.truncated_mass[key] = (
+                self.truncated_mass.get(key, 0.0) + truncated_mass
             )
             return self._create_prediction(entity_id, entity_type, truncated_pmf)
         else:
@@ -1282,7 +1290,7 @@ class DemandPredictor:
         selected_outflows = [
             inputs.outflows[k] for k in outflow_keys if outflow_allowed(k)
         ]
-        
+
         # Calculate physical cap for departures (sum of physical maxes from PMF-based flows)
         # Departures are always PMF-based (physically bounded by current patients)
         departures_physical_cap = 0
@@ -1296,10 +1304,13 @@ class DemandPredictor:
                 # Departures should never be Poisson (enforced elsewhere)
                 # But if somehow one exists, it's unbounded, so we skip it
                 pass
-        
+
         # Apply physical cap to departures prediction
         departures = self.predict_flow_total(
-            selected_outflows, subspecialty_id, "departures", max_support=departures_physical_cap
+            selected_outflows,
+            subspecialty_id,
+            "departures",
+            max_support=departures_physical_cap,
         )
 
         # Renormalize arrivals and departures before net-flow
@@ -1336,6 +1347,7 @@ class DemandPredictor:
         entity_type: EntityType,
         child_predictions: List[DemandPrediction],
         max_support: Optional[int] = None,
+        flow_type: Optional[str] = None,
     ) -> DemandPrediction:
         """Generic method for hierarchical prediction at any level.
 
@@ -1353,6 +1365,9 @@ class DemandPredictor:
         max_support : int, optional
             Pre-calculated maximum support value. If provided, the result will
             be truncated to this value with renormalization.
+        flow_type : str, optional
+            Flow type ('arrivals' or 'departures') for truncated mass tracking.
+            If None, truncated mass will not be tracked for this aggregation.
 
         Returns
         -------
@@ -1366,7 +1381,9 @@ class DemandPredictor:
         uses renormalization to ensure probability conservation.
         """
         distributions = [p.probabilities for p in child_predictions]
-        p_total = self._convolve_multiple(distributions, max_support)
+        p_total = self._convolve_multiple(
+            distributions, max_support, entity_id=entity_id, flow_type=flow_type
+        )
         return self._create_prediction(entity_id, entity_type.name, p_total)
 
     def _create_bundle_from_children(
@@ -1401,9 +1418,13 @@ class DemandPredictor:
         departures_preds = [b.departures for b in child_bundles]
 
         arrivals = self.predict_hierarchical_level(
[…]
