### 2025-12-11 217a21a
track truncated mass in k_sigma tests
