+++ tests/test_hierarchy.py
@@ -1219,9 +1219,11 @@ class TestHierarchicalPredictor:
 
         def poisson_flow(flow_id: str, lam: float) -> FlowInputs:
             return FlowInputs(flow_id=flow_id, flow_type="poisson", distribution=lam)
-        
+
         def pmf_flow(flow_id: str, probabilities: np.ndarray) -> FlowInputs:
-            return FlowInputs(flow_id=flow_id, flow_type="pmf", distribution=probabilities)
+            return FlowInputs(
+                flow_id=flow_id, flow_type="pmf", distribution=probabilities
+            )
 
         def make_inputs(subspecialty_id: str) -> SubspecialtyPredictionInputs:
             return SubspecialtyPredictionInputs(
@@ -1238,9 +1240,13 @@ class TestHierarchicalPredictor:
                 outflows={
                     # Departures must be PMF-based (physically bounded by current patients)
                     # For mean=0.5, use PMF [0.5, 0.5] (P(0)=0.5, P(1)=0.5)
-                    "elective_departures": pmf_flow("elective_departures", np.array([0.5, 0.5])),
+                    "elective_departures": pmf_flow(
+                        "elective_departures", np.array([0.5, 0.5])
+                    ),
                     # For mean=0.0, use PMF [1.0] (0 patients)
-                    "emergency_departures": pmf_flow("emergency_departures", np.array([1.0])),
+                    "emergency_departures": pmf_flow(
+                        "emergency_departures", np.array([1.0])
+                    ),
                 },
             )
 
