### 2025-12-11 0982d61
remove debug statements from k_sigma test
