### 2025-12-18 #9187 @Kevin-Mattheus-Moerman
👉 Please remove the additional occurrence of the title around lines 10-11.
👉 Check `conbined` that should be `combined`
