### 2025-12-18 2cef5a3
fix paper typos for JOSS
