### 2025-12-18 #9187 @zmek
@Kevin-Mattheus-Moerman thanks for alerting me to these issues. They are now fixed in a new version pushed to main
