### 2025-12-23 1dcdbbc
refactor hierarchy module
