+++ src/patientflow/predict/hierarchy/types.py
@@ -1,0 +1,359 @@
+"""Type definitions for hierarchical demand prediction."""
+
+from dataclasses import dataclass
+from typing import Dict, Any, Tuple
+import numpy as np
+
+# Constants for magic numbers
+DEFAULT_PERCENTILES = [50, 75, 90, 95, 99]
+DEFAULT_PRECISION = 3
+DEFAULT_MAX_PROBS = 10
+
+
+@dataclass
+class DemandPrediction:
+    """Result of a demand prediction at any hierarchical level.
+
+    This dataclass encapsulates the complete prediction results for a single
+    organizational entity (subspecialty, reporting unit, division, board, or hospital).
+    It contains the probability mass function, expected value, and key percentiles.
+
+    Parameters
+    ----------
+    entity_id : str
+        Unique identifier for the organizational entity (e.g., subspecialty name)
+    entity_type : str
+        Type of entity ('subspecialty', 'reporting_unit', 'division', 'board', 'hospital')
+    probabilities : numpy.ndarray
+        Probability mass function for bed demand (probability of 0, 1, 2, ... beds)
+    expected_value : float
+        Expected number of beds needed (mean of the distribution)
+    percentiles : dict[int, int]
+        Dictionary mapping percentile values (50, 75, 90, 95, 99) to bed counts
+    offset : int, optional
+        Offset for the support of the distribution. Index i corresponds to value (i + offset).
+        Default is 0 for non-negative distributions. For net flow, offset is typically negative.
+
+    Attributes
+    ----------
+    entity_id : str
+        Unique identifier for the organizational entity
+    entity_type : str
+        Type of entity in the hierarchy
+    probabilities : numpy.ndarray
+        Probability mass function for bed demand
+    expected_value : float
+        Expected number of beds needed
+    percentiles : dict[int, int]
+        Percentile values for bed demand
+    offset : int
+        Offset for the support (index 0 corresponds to value = offset)
+    max_beds : int
+        Maximum number of beds in the probability distribution
+
+    Notes
+    -----
+    The probabilities array represents P(X=k) where X is the number of beds needed
+    and k ranges from offset to (offset + max_beds). The sum of all probabilities should equal 1.0.
+    For non-negative distributions (arrivals, departures), offset=0.
+    For net flow distributions, offset can be negative (e.g., -10 means support starts at -10).
+    """
+
+    entity_id: str
+    entity_type: str
+    probabilities: np.ndarray
+    expected_value: float
+    percentiles: Dict[int, int]
+    offset: int = 0
+
+    @property
+    def max_beds(self) -> int:
+        """Maximum number of beds in the probability distribution.
+
+        Returns
+        -------
+        int
+            Maximum bed count (length of probabilities array minus 1)
+        """
+        return len(self.probabilities) - 1
+
+    def to_pretty(
+        self, max_probs: int = DEFAULT_MAX_PROBS, precision: int = DEFAULT_PRECISION
+    ) -> str:
+        """Return a concise, human-friendly string representation.
+
+        Parameters
+        ----------
+        max_probs : int, default 10
+            Maximum number of head probabilities to display from the PMF
+        precision : int, default 3
+            Number of decimal places for numeric values
+
+        Returns
+        -------
+        str
+            Formatted summary string
+        """
+        probs = self.probabilities
+
+        # Find where the probability mass is concentrated
+        if len(probs) <= max_probs:
+            start_idx = 0
+            end_idx = len(probs)
+        else:
+            mode_idx = int(np.argmax(probs))
+            half_window = max_probs // 2
+            start_idx = max(0, mode_idx - half_window)
+            end_idx = min(len(probs), start_idx + max_probs)
[…]
