+++ src/patientflow/predict/hierarchy/orchestrator.py
@@ -1,0 +1,248 @@
+"""Hierarchical prediction orchestrator."""
+
+from typing import Dict, Optional, Tuple, Any
+
+from patientflow.predict.subspecialty import SubspecialtyPredictionInputs
+from .structure import Hierarchy
+from .types import PredictionBundle, FlowSelection
+from .calculation import DemandPredictor
+
+
+class HierarchicalPredictor:
+    """Orchestrates the hierarchical prediction process.
+
+    This class manages the interaction between the hierarchy structure and the demand predictor,
+    providing a method running predictions across all levels of the organization.
+
+    The prediction process follows a 3-phase algorithm:
+
+    Phase 1: Bottom-up Stats & Top-down Capping
+    - Recursively calculate statistical properties (sum of means, combined variance) from bottom up
+    - Calculate max_support caps for each node (statistical cap)
+    - The top-level cap is driven by its total statistical properties
+    - Intermediate caps respect both their own statistical properties and are sufficient to support children
+
+    Phase 2: Bottom-level Prediction
+    - Compute full PMF predictions for all bottom-level entities (subspecialties) using the caps from Phase 1
+
+    Phase 3: Aggregation
+    - Aggregate predictions upwards through the hierarchy using convolution
+    - Apply caps at each level during aggregation to keep distribution sizes manageable
+    - Compute net flow at each level
+
+    Attributes
+    ----------
+    hierarchy : Hierarchy
+        The organisational hierarchy structure
+    predictor : DemandPredictor
+        The calculation engine
+    """
+
+    def __init__(self, hierarchy: Hierarchy, predictor: DemandPredictor):
+        self.hierarchy = hierarchy
+        self.predictor = predictor
+        # Store computed bundles: entity_id -> PredictionBundle
+        self.prediction_results: Dict[str, PredictionBundle] = {}
+
+    def predict_all_levels(
+        self,
+        bottom_level_data: Dict[str, SubspecialtyPredictionInputs],
+        top_level_id: Optional[str] = None,
+        flow_selection: Optional[FlowSelection] = None,
+    ) -> Dict[str, PredictionBundle]:
+        """Run predictions for the entire hierarchy.
+
+        This method executes the full 3-phase prediction algorithm (capping,
+        bottom-level prediction, aggregation) for the specified subtree.
+
+        Parameters
+        ----------
+        bottom_level_data : Dict[str, SubspecialtyPredictionInputs]
+            Dictionary mapping bottom-level entity IDs to their prediction inputs
+        top_level_id : str, optional
+            Root entity to start prediction from. If None, predicts for all
+            top-level entities in the hierarchy.
+        flow_selection : FlowSelection, optional
+            Selection specifying which flows to include. If None, uses
+            FlowSelection.default() which includes all flows.
+
+        Returns
+        -------
+        Dict[str, PredictionBundle]
+            Dictionary containing prediction bundles for all entities in the
+            hierarchy (bottom-level and aggregated), keyed by entity ID.
+        """
+        if flow_selection is None:
+            flow_selection = FlowSelection.default()
+        
+        # Validate flow selection configuration
+        flow_selection.validate()
+
+        self.prediction_results.clear()
+        self.predictor.clear_truncated_mass()
+
+        if top_level_id is None:
+            # Predict for all top-level entities
+            top_level_type = self.hierarchy.get_top_level_type()
+            top_entities = self.hierarchy.get_entities_by_type(top_level_type)
+            for entity_id in top_entities:
+                self._predict_subtree(entity_id, bottom_level_data, flow_selection)
+        else:
+            self._predict_subtree(top_level_id, bottom_level_data, flow_selection)
+
+        return self.prediction_results
+
+    def _predict_subtree(
+        self,
[…]

[…]
