### 2025-12-23 de75d65
refactor hierarchy module
