+++ src/patientflow/predict/hierarchy/structure.py
@@ -1,0 +1,632 @@
+"""Hierarchical structure definitions."""
+
+from dataclasses import dataclass
+from typing import Dict, List, Optional, Any
+import yaml
+import pandas as pd
+
+
+class EntityType:
+    """Represents an entity type in the hierarchy.
+
+    This class is used to represent entity types dynamically based on the
+    hierarchy configuration.
+    """
+
+    def __init__(self, name: str):
+        self.name = name
+
+    def __str__(self) -> str:
+        return self.name
+
+    def __repr__(self) -> str:
+        return f"EntityType('{self.name}')"
+
+    def __eq__(self, other) -> bool:
+        if isinstance(other, EntityType):
+            return self.name == other.name
+        return False
+
+    def __hash__(self) -> int:
+        return hash(self.name)
+
+    @classmethod
+    def from_string(cls, value: str) -> "EntityType":
+        """Create EntityType from string."""
+        return cls(value)
+
+
+@dataclass
+class HierarchyLevel:
+    """Represents a level in the hierarchy with its configuration."""
+
+    entity_type: EntityType
+    parent_type: Optional[EntityType]
+    level_order: int  # 0 = bottom level, higher numbers = higher levels
+
+
+class Hierarchy:
+    """Generic hierarchical structure that can represent any organizational hierarchy.
+
+    This class eliminates the need for specific methods for each entity type by using
+    a generic approach that works with any hierarchical structure.
+
+    Attributes
+    ----------
+    levels : Dict[EntityType, HierarchyLevel]
+        Configuration of each level in the hierarchy
+    relationships : Dict[str, str]
+        Mapping from child_id to parent_id for all entities
+    entity_types : Dict[str, EntityType]
+        Mapping from entity_id to its EntityType
+    """
+
+    def __init__(self, levels: List[HierarchyLevel]):
+        self.levels = {level.entity_type: level for level in levels}
+        self.relationships: Dict[str, str] = {}  # child_id -> parent_id
+        self.entity_types: Dict[str, EntityType] = {}  # entity_id -> EntityType
+
+        # Validate that levels form a proper hierarchy
+        self._validate_levels()
+
+    def _validate_levels(self):
+        """Validate that the hierarchy levels are properly configured."""
+        # Check that there's exactly one top level (no parent)
+        top_levels = [
+            level for level in self.levels.values() if level.parent_type is None
+        ]
+        if len(top_levels) != 1:
+            raise ValueError("Hierarchy must have exactly one top level")
+
+        # Check that all parent types exist
+        for level in self.levels.values():
+            if level.parent_type is not None and level.parent_type not in self.levels:
+                raise ValueError(f"Parent type {level.parent_type} not found in levels")
+
+    @classmethod
+    def from_yaml(cls, config_path: str) -> "Hierarchy":
+        """Create hierarchy from YAML configuration file.
+
+        Parameters
+        ----------
+        config_path : str
+            Path to YAML configuration file
+
+        Returns
+        -------
+        Hierarchy
+            Configured hierarchy instance
+        """
+        with open(config_path, "r") as f:
+            config = yaml.safe_load(f)
+
+        levels = []
+        for level_config in config["levels"]:
+            entity_type = EntityType.from_string(level_config["entity_type"])
+            parent_type = None
+            if level_config.get("parent_type"):
+                parent_type = EntityType.from_string(level_config["parent_type"])
+
+            level = HierarchyLevel(
+                entity_type, parent_type, level_config["level_order"]
+            )
+            levels.append(level)
+
+        return cls(levels)
+
+    @classmethod
+    def create_default_hospital(cls) -> "Hierarchy":
+        """Create default hospital hierarchy.
+
+        Returns
+        -------
+        Hierarchy
+            Default hospital hierarchy with standard levels
+        """
[…]

[…]
