### 2026-01-12 96540d7
Update notebook 4b to reflect refactoring including detailed changes to - change DemandPrediction class to include mode and expectation - remove check of service_id since this is implicit in predict_service - correct docstrings for find_probability_threshold_index - add min_beds_with_probability to DemandPrediction
