### 2026-01-12 90bce7b
Update PyYAML dependency version constraint
