+++ pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
     "Typing :: Typed",
 ]
 dependencies = [
-    "PyYAML==6.0.2",
+    "PyYAML>=6.0.2",
     "joblib>=1.4.2",
     "matplotlib>=3.1.3",
     "numpy>=1.18.1",
