+++ src/patientflow/predict/__init__.py
@@ -4,14 +4,5 @@ This module provides functions for making predictions about future patient flow,
 including emergency demand forecasting and other predictive analytics.
 """
 
-from patientflow.predict.service import (
-    build_service_data,
-    ServicePredictionInputs,
-    compute_transfer_arrivals,
-)
-
-__all__ = [
-    "build_service_data",
-    "ServicePredictionInputs",
-    "compute_transfer_arrivals",
-]
+# No exports - users import directly from submodules
+__all__ = []
