### 2026-01-13 255a7b6
correct pmf when empty
