### 2026-01-13 cbae7e9
return self.prediction_results as copy to avoid overwriting
