+++ src/patientflow/predict/hierarchy/orchestrate.py
@@ -4,15 +4,99 @@ This module provides HierarchyPredictor, which orchestrates the 3-phase
 prediction algorithm across all levels of the organizational hierarchy.
 """
 
-from typing import Dict, Optional, Any
+from typing import Dict, Optional, Any, List
 
 from patientflow.predict.service import ServicePredictionInputs
-from patientflow.predict.hierarchy.structure import Hierarchy, populate_hierarchy_from_dataframe
+from patientflow.predict.hierarchy.structure import Hierarchy, populate_hierarchy_from_dataframe, EntityType
 from patientflow.predict.types import PredictionBundle, FlowSelection
 from patientflow.predict.demand import DemandPredictor
 from patientflow.predict.hierarchy.calculate import calculate_hierarchical_stats
 
 import pandas as pd
+
+
+class PredictionResults(dict):
+    """Dictionary-like container for prediction results with flexible key access.
+    
+    Supports two key formats:
+    1. Prefixed ID: "hospital:UCLH" (always works)
+    2. Original name: "UCLH" (works if unique, raises KeyError if ambiguous)
+    
+    This class prevents data loss from name collisions while allowing convenient
+    access using original entity names when they are unambiguous.
+    
+    Examples
+    --------
+    >>> results = predictor.predict_all_levels(bottom_level_data)
+    >>> # Prefixed ID (always works)
+    >>> bundle = results["hospital:UCLH"]
+    >>> 
+    >>> # Original name (only if unique)
+    >>> bundle = results["UCLH"]  # Works if "UCLH" appears only once
+    >>> bundle = results["Acute Medicine"]  # Raises KeyError if ambiguous
+    
+    Notes
+    -----
+    Standard dict methods (get(), keys(), items(), values(), in operator)
+    work as expected. The get() method uses standard dict behavior (returns
+    None or default if key not found).
+    """
+    
+    def __init__(self, data: Dict[str, PredictionBundle]):
+        """Initialize PredictionResults.
+        
+        Parameters
+        ----------
+        data : Dict[str, PredictionBundle]
+            Dictionary keyed by prefixed IDs (e.g., "hospital:UCLH")
+        """
+        super().__init__(data)
+        # Build cache: original_name -> [list of prefixed_ids]
+        self._name_to_prefixed: Dict[str, List[str]] = {}
+        for prefixed_id in data.keys():
+            if ':' in prefixed_id:
+                _, original_name = prefixed_id.split(':', 1)
+                if original_name not in self._name_to_prefixed:
+                    self._name_to_prefixed[original_name] = []
+                self._name_to_prefixed[original_name].append(prefixed_id)
+    
+    def __getitem__(self, key: str) -> PredictionBundle:
+        """Get prediction bundle using prefixed ID or original name.
+        
+        Parameters
+        ----------
+        key : str
+            Prefixed ID ("hospital:UCLH") or original name ("UCLH")
+        
+        Returns
+        -------
+        PredictionBundle
+            The prediction bundle
+        
+        Raises
+        ------
+        KeyError
+            If key not found or ambiguous (with helpful message)
+        """
+        # Prefixed ID (contains ':')
+        if ':' in key:
+            return super().__getitem__(key)
+        
+        # Original name - check if unambiguous
+        if key in self._name_to_prefixed:
+            matching_ids = self._name_to_prefixed[key]
+            if len(matching_ids) == 1:
+                # Unique match
+                return super().__getitem__(matching_ids[0])
+            else:
+                # Ambiguous - raise helpful error
+                raise KeyError(
+                    f"Ambiguous entity name '{key}'. Found at multiple levels: "
+                    f"{', '.join(matching_ids)}. Use prefixed ID instead."
+                )
+        
+        # Not found
[…]
