### 2026-01-14 bd3f8fd
ensure full dict inheritance in PredictionResults class
