### 2026-01-19 0ed2553
fix madcap plot error in null values
