### 2026-01-20 2d74b4a
Address modelling issues: - do not one-hot encode boolean - add validation and error message when no snapshots match prediction time
 Please enter the commit message for your changes. Lines starting
