### 2026-01-21 b762087
handle missing cols inside classifier pipeline
