### 2026-02-06 631bd1d
save exclude_from_training_data list with model
