+++ src/patientflow/viz/shap.py
@@ -24,11 +24,12 @@ from pathlib import Path
 def plot_shap(
     trained_models: list[TrainedClassifier] | dict[str, TrainedClassifier],
     test_visits,
-    exclude_from_training_data: Optional[List[str]] = None,
     media_file_path: Optional[Path] = None,
     file_name: Optional[str] = None,
     return_figure=False,
     label_col: str = "is_admitted",
+    *,
+    exclude_from_training_data: Optional[List[str]] = None,
 ):
     """Generate SHAP plots for multiple trained models.
 
@@ -42,9 +43,6 @@ def plot_shap(
         List of trained classifier objects or dictionary with TrainedClassifier values.
     test_visits : pandas.DataFrame
         DataFrame containing the test visit data.
-    exclude_from_training_data : List[str], optional, deprecated
-        This parameter is deprecated and ignored. Column selection is now handled
-        automatically by the pipeline's FeatureColumnTransformer.
     media_file_path : Path, optional
         Directory path where the generated plots will be saved. If None, plots are
         only displayed.
@@ -54,6 +52,9 @@ def plot_shap(
         If True, returns the figure instead of displaying it.
     label_col : str, default="is_admitted"
         Name of the column containing the target labels.
+    exclude_from_training_data : List[str], optional, deprecated
+        This parameter is deprecated and ignored. Column selection is now handled
+        automatically by the pipeline's FeatureColumnTransformer.
 
     Returns
     -------
