### 2026-02-06 4eaacb6
fix: handle SettingWithCopyWarning removal in pandas 3.0
