### 2026-02-09 ee02e83
check documentation and notebooks
