+++ src/patientflow/predict/hierarchy/__init__.py
@@ -1,220 +1,82 @@
 """Hierarchical demand prediction module.
 
 This module provides tools for predicting hospital bed demand across different
-organisational levels (subspecialty, division, hospital, etc.) using a
-hierarchical structure.
+organisational levels (e.g. subspecialty, division, hospital) using a
+hierarchical structure. Predictions are generated at the bottom level
+(services) and aggregated upward through convolution of probability
+distributions, with statistical capping at each level to control
+distribution size.
 
-Usage
------
-The typical workflow is:
-
-1. Create or load a hierarchy structure:
-   >>> hierarchy = Hierarchy.create_default_hospital()
-   >>> # or load from YAML configuration
-   >>> hierarchy = Hierarchy.from_yaml("config.yaml")
-
-2. Populate the hierarchy with organizational data:
-   >>> import pandas as pd
-   >>> # The DataFrame must represent a proper tree structure where each child
-   >>> # entity has exactly one parent. Each row represents one path through
-   >>> # the hierarchy from bottom to top.
-   >>> hierarchy_df = pd.DataFrame({
-   ...     'subspecialty': ['Cardiology', 'Cardiology', 'Neurology', 'Neurology'],
-   ...     'reporting_unit': ['Cardiac Unit', 'Cardiac Unit', 'Neuro Unit', 'Neuro Unit'],
-   ...     'division': ['Medical Division', 'Medical Division', 'Medical Division', 'Medical Division'],
-   ... })
-   >>> # Important: The hierarchy must form a tree structure. Each child entity
-   >>> # can only have one parent. If the same child appears with different
-   >>> # parents in different rows, the last relationship processed will
-   >>> # overwrite previous ones, silently creating an invalid hierarchy.
-   >>> # For example, this would be INVALID:
-   >>> #   Row 1: subspecialty='Cardiology', reporting_unit='Unit X', division='Division A'
-   >>> #   Row 2: subspecialty='Cardiology', reporting_unit='Unit X', division='Division B'
-   >>> # This would result in 'Unit X' being assigned only to 'Division B'
-   >>> # (the last one processed), overwriting the relationship to 'Division A'.
-   >>> # This violates the tree structure requirement and should be avoided.
-   >>> # Note: Only entity types with a column_mapping entry will have entities
-   >>> # created from the DataFrame. The top_level_id ensures a single top-level
-   >>> # entity exists (created if not in DataFrame, consolidated if present).
-   >>> column_mapping = {
-   ...     'subspecialty': 'subspecialty',
-   ...     'reporting_unit': 'reporting_unit',
-   ...     'division': 'division',
-   ...     # 'hospital' not included, so top-level entity will be created separately
-   ... }
-   >>> populate_hierarchy_from_dataframe(
-   ...     hierarchy, hierarchy_df, column_mapping, top_level_id="uclh"
-   ... )
-
-3. Prepare prediction inputs (using functions from predict.service):
-   >>> from patientflow.predict.service import build_service_data
-   >>> bottom_level_data = build_service_data(
-   ...     models=(ed_classifier, inpatient_classifier, spec_model, ...),
-   ...     prediction_time=(12, 0),
-   ...     ed_snapshots=ed_snapshots_df,
-   ...     inpatient_snapshots=inpatient_snapshots_df,
-   ...     specialties=['Cardiology', 'Neurology', ...],
-   ...     prediction_window=timedelta(hours=8),
-   ...     x1=4.0, y1=0.95, x2=8.0, y2=0.99
-   ... )
-
-4. Create predictor and run predictions:
-   >>> predictor = create_hierarchical_predictor(
-   ...     hierarchy_df=hierarchy_df,
-   ...     column_mapping=column_mapping,
-   ...     top_level_id="Hospital",
[…]
