### 2026-02-09 bfbb47f
Revert "add tests for viz module"
This reverts commit 39fbe956cc111530e8a9bce457bedd3dfe801f4d.
