### 2026-02-09 8d0f84a
add viz tests
