+++ tests/test_viz.py
@@ -1,0 +1,412 @@
+"""Tests for the patientflow.viz module.
+
+Tier 1: Import smoke tests — verify every viz module can be imported.
+Tier 2: Unit tests for pure (non-plotting) functions in the viz module.
+"""
+
+import unittest
+import numpy as np
+import pandas as pd
+
+from patientflow.viz.utils import clean_title_for_filename, format_prediction_time
+from patientflow.viz.madcap import classify_age, DEFAULT_AGE_CATEGORIES
+from patientflow.viz.probability_distribution import _calculate_probability_thresholds
+from patientflow.viz.epudd import _calculate_cdf_values
+from patientflow.viz.randomised_pit import _prob_to_cdf
+
+
+# ---------------------------------------------------------------------------
+# Tier 1 – Import smoke tests
+# ---------------------------------------------------------------------------
+
+
+class TestVizImports(unittest.TestCase):
+    """Verify that every viz module can be imported without errors."""
+
+    def test_import_viz_package(self):
+        import patientflow.viz  # noqa: F401
+
+    def test_import_utils(self):
+        from patientflow.viz.utils import (  # noqa: F401
+            clean_title_for_filename,
+            format_prediction_time,
+        )
+
+    def test_import_arrival_rates(self):
+        from patientflow.viz.arrival_rates import (  # noqa: F401
+            plot_arrival_rates,
+            plot_cumulative_arrival_rates,
+        )
+
+    def test_import_aspirational_curve(self):
+        from patientflow.viz.aspirational_curve import plot_curve  # noqa: F401
+
+    def test_import_calibration(self):
+        from patientflow.viz.calibration import plot_calibration  # noqa: F401
+
+    def test_import_data_distribution(self):
+        from patientflow.viz.data_distribution import (  # noqa: F401
+            plot_data_distribution,
+        )
+
+    def test_import_epudd(self):
+        from patientflow.viz.epudd import plot_epudd  # noqa: F401
+
+    def test_import_estimated_probabilities(self):
+        from patientflow.viz.estimated_probabilities import (  # noqa: F401
+            plot_estimated_probabilities,
+        )
+
+    def test_import_features(self):
+        from patientflow.viz.features import plot_features  # noqa: F401
+
+    def test_import_madcap(self):
+        from patientflow.viz.madcap import (  # noqa: F401
+            classify_age,
+            plot_madcap,
+            plot_madcap_by_group,
+        )
+
+    def test_import_observed_against_expected(self):
+        from patientflow.viz.observed_against_expected import (  # noqa: F401
+            plot_deltas,
+            plot_arrival_delta_single_instance,
+            plot_arrival_deltas,
+        )
+
+    def test_import_probability_distribution(self):
+        from patientflow.viz.probability_distribution import (  # noqa: F401
+            plot_prob_dist,
+        )
+
+    def test_import_quantile_quantile(self):
+        from patientflow.viz.quantile_quantile import qq_plot  # noqa: F401
+
+    def test_import_randomised_pit(self):
+        from patientflow.viz.randomised_pit import (  # noqa: F401
+            plot_randomised_pit,
+        )
+
+    def test_import_shap(self):
+        try:
+            from patientflow.viz.shap import plot_shap  # noqa: F401
[…]
