### 2026-02-09 03b56dc
add Tier 3 tests for viz modules
