+++ tests/test_viz.py
@@ -2,104 +2,85 @@
 
 Tier 1: Import smoke tests — verify every viz module can be imported.
 Tier 2: Unit tests for pure (non-plotting) functions in the viz module.
+Tier 3: Render smoke tests — verify plotting functions produce figures without errors.
 """
 
 import unittest
 import numpy as np
 import pandas as pd
+import matplotlib
+
+matplotlib.use("Agg")  # Non-interactive backend for headless testing
+import matplotlib.pyplot as plt  # noqa: E402
+from matplotlib.figure import Figure  # noqa: E402
 
 from patientflow.viz.utils import clean_title_for_filename, format_prediction_time
-from patientflow.viz.madcap import classify_age, DEFAULT_AGE_CATEGORIES
-from patientflow.viz.probability_distribution import _calculate_probability_thresholds
+from patientflow.viz.madcap import classify_age
+from patientflow.viz.probability_distribution import (
+    _calculate_probability_thresholds,
+    plot_prob_dist,
+)
 from patientflow.viz.epudd import _calculate_cdf_values
 from patientflow.viz.randomised_pit import _prob_to_cdf
+from patientflow.viz.aspirational_curve import plot_curve
+from patientflow.viz.survival_curve import plot_admission_time_survival_curve
+from patientflow.viz.data_distribution import plot_data_distribution
+from patientflow.viz.observed_against_expected import plot_deltas
+from patientflow.viz.arrival_rates import (
+    plot_arrival_rates,
+    plot_cumulative_arrival_rates,
+)
+from patientflow.viz.trial_results import plot_trial_results
+from patientflow.model_artifacts import HyperParameterTrial
 
 
 # ---------------------------------------------------------------------------
 # Tier 1 – Import smoke tests
 # ---------------------------------------------------------------------------
 
+_VIZ_IMPORTS = [
+    ("patientflow.viz", None),
+    ("patientflow.viz.utils", ["clean_title_for_filename", "format_prediction_time"]),
+    (
+        "patientflow.viz.arrival_rates",
+        ["plot_arrival_rates", "plot_cumulative_arrival_rates"],
+    ),
+    ("patientflow.viz.aspirational_curve", ["plot_curve"]),
+    ("patientflow.viz.calibration", ["plot_calibration"]),
+    ("patientflow.viz.data_distribution", ["plot_data_distribution"]),
+    ("patientflow.viz.epudd", ["plot_epudd"]),
+    ("patientflow.viz.estimated_probabilities", ["plot_estimated_probabilities"]),
+    ("patientflow.viz.features", ["plot_features"]),
+    ("patientflow.viz.madcap", ["classify_age", "plot_madcap", "plot_madcap_by_group"]),
+    ("patientflow.viz.observed_against_expected", ["plot_deltas"]),
+    ("patientflow.viz.probability_distribution", ["plot_prob_dist"]),
+    ("patientflow.viz.quantile_quantile", ["qq_plot"]),
+    ("patientflow.viz.randomised_pit", ["plot_randomised_pit"]),
+    ("patientflow.viz.survival_curve", ["plot_admission_time_survival_curve"]),
+    ("patientflow.viz.trial_results", ["plot_trial_results"]),
+]
+
 
 class TestVizImports(unittest.TestCase):
     """Verify that every viz module can be imported without errors."""
 
-    def test_import_viz_package(self):
-        import patientflow.viz  # noqa: F401
+    def test_all_viz_modules_import(self):
+        import importlib
 
-    def test_import_utils(self):
-        from patientflow.viz.utils import (  # noqa: F401
-            clean_title_for_filename,
-            format_prediction_time,
-        )
-
-    def test_import_arrival_rates(self):
[…]
