### 2026-02-11 d54cdc2
add defensive validation in case spec model is none
