+++ src/patientflow/predict/service.py
@@ -451,6 +451,13 @@ def _validate_models_and_data(
     ):
         raise ValueError("Transfer model has not been fit")
 
+    # Validate that a specialty model is provided when needed
+    if spec_model is None and specialties and ed_snapshots is not None:
+        raise ValueError(
+            "Specialty model (spec_model) is required when specialties are requested "
+            "and ED snapshots are provided, but spec_model is None"
+        )
+
     # Validate specialties alignment
     if yet_to_arrive_model is not None and hasattr(yet_to_arrive_model, "filters"):
         if not set(yet_to_arrive_model.filters.keys()) == set(specialties):
@@ -646,11 +653,11 @@ def _prepare_base_probabilities(
                 special_category_dict=special_category_dict,
             )
         else:
-            # If no spec model, assuming 0 probability for all specialties
-            # This is effectively "unknown"
-            ed_snapshots.loc[:, "specialty_prob"] = [
-                {} for _ in range(len(ed_snapshots))
-            ]
+            raise ValueError(
+                "Cannot compute specialty probabilities: spec_model is None "
+                "but ED snapshots are present. This should have been caught "
+                "by validation — ensure _validate_models_and_data is called first."
+            )
 
     # Probability of being admitted within window (per row) for ED patients
     if (
