### 2026-02-11 4d586e9
fix issue pickling objects with callable functions
