+++ src/patientflow/train/emergency_demand.py
@@ -18,7 +18,13 @@ test_real_time_predictions : Test real-time prediction functionality
 
 train_all_models : Complete training pipeline
     Trains all three model types (admissions, specialty, yet-to-arrive)
-    with proper validation and optional model saving.
+    with proper validation and optional model saving. Returns the trained
+    models as a dictionary.
+
+prepare_prediction_inputs : Convenience wrapper for interactive use
+    Loads configuration and data, then calls ``train_all_models`` and
+    returns everything needed for inference. Designed for notebook /
+    pipeline use.
 
 main : Entry point for training pipeline
     Loads configuration, data, and runs the complete training process.
@@ -45,24 +51,28 @@ from patientflow.train.utils import save_model
 from patientflow.predictors.sequence_to_outcome_predictor import (
     SequenceToOutcomePredictor,
 )
+from patientflow.predictors.subgroup_predictor import MultiSubgroupPredictor
+from patientflow.predictors.subgroup_definitions import (
+    create_paediatric_adult_subgroup_functions,
+)
 from patientflow.predictors.incoming_admission_predictors import (
     ParametricIncomingAdmissionPredictor,
 )
 from patientflow.predict.emergency_demand import create_predictions
 
 from patientflow.train.classifiers import train_multiple_classifiers
-from patientflow.train.sequence_predictor import train_sequence_predictor
 from patientflow.train.incoming_admission_predictor import (
     train_parametric_admission_predictor,
 )
 from patientflow.model_artifacts import TrainedClassifier
+from patientflow.prepare import select_one_snapshot_per_visit
 
 
 def test_real_time_predictions(
     visits,
     models: Tuple[
         Dict[str, TrainedClassifier],
-        SequenceToOutcomePredictor,
+        "SequenceToOutcomePredictor | MultiSubgroupPredictor",
         ParametricIncomingAdmissionPredictor,
     ],
     prediction_window,
@@ -229,7 +239,12 @@ def train_all_models(
 
     Returns
     -------
-    None
+    dict
+        Dictionary containing the trained models with keys:
+
+        - ``"admission_models"`` : dict of str to TrainedClassifier
+        - ``"specialty_model"`` : MultiSubgroupPredictor
+        - ``"yta_model"`` : ParametricIncomingAdmissionPredictor
 
     Raises
     ------
@@ -313,15 +328,25 @@ def train_all_models(
     if save_models:
         save_model(admission_models, model_names["admissions"], model_file_path)
 
-    # Train specialty model
-    specialty_model = train_sequence_predictor(
-        train_visits=train_visits,
-        model_name=model_names["specialty"],
+    # Train specialty model using MultiSubgroupPredictor
+    # This trains separate SequenceToOutcomePredictor instances per subgroup
+    # (e.g. paediatric vs adult) for more accurate specialty prediction
+
+    # Select one snapshot per visit for specialty training
+    visits_single = select_one_snapshot_per_visit(train_visits, visit_col)
+    admitted = visits_single[
+        (visits_single.is_admitted) & ~(visits_single.specialty.isnull())
+    ]
+
+    specialty_model = MultiSubgroupPredictor(
+        subgroup_functions=create_paediatric_adult_subgroup_functions(),
+        base_predictor_class=SequenceToOutcomePredictor,
         input_var="consultation_sequence",
         grouping_var="final_sequence",
         outcome_var="specialty",
-        visit_col=visit_col,
+        min_samples=50,
     )
+    specialty_model = specialty_model.fit(admitted)
 
     # Save specialty model if requested
     if save_models:
[…]
