### 2026-02-11 eca2fe9
notebook for pipeline plots
