+++ src/patientflow/viz/pipeline_plots.py
@@ -1,0 +1,492 @@
+"""Visualization functions for the whole-pipeline overview figures.
+
+This module provides functions for creating the pipeline overview plots
+that illustrate the end-to-end patient flow prediction process: from
+a snapshot of patients currently in the ED, through individual admission
+predictions, to aggregate probability distributions of bed demand.
+
+Functions
+---------
+create_colour_dict : function
+    Build a colour mapping dictionary for specialty categories.
+prepare_snapshot_data : function
+    Filter and prepare a single snapshot of ED visits for plotting.
+add_specialty_predictions : function
+    Attach a per-row specialty probability dict to a snapshot DataFrame.
+build_pipeline_prediction_inputs : function
+    Build per-service prediction inputs for a given snapshot using
+    ``build_service_data`` from ``patientflow.predict.service``.
+in_ed_now_plot : function
+    Scatter plot of patients currently in the ED, optionally coloured by
+    predicted probability of admission.
+"""
+
+import os
+from pathlib import Path
+
+import matplotlib.colors as mcolors
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+
+
+DEFAULT_LOCATION_MAPPING = {
+    "utc": "Minors",
+    "majors": "Majors/Resus",
+    "sdec": "Other",
+    "paeds": "Other",
+    "waiting": "Other",
+    "rat": "Majors/Resus",
+    "sdec_waiting": "Other",
+    "resus": "Majors/Resus",
+    "taf": "Other",
+}
+"""dict : Default mapping from raw ``current_location_type`` values to
+simplified display categories used on the y-axis of pipeline plots."""
+
+DEFAULT_CATEGORY_ORDER = ["Majors/Resus", "Minors", "Other"]
+"""list of str : Default ordered list of display category names for the
+y-axis of pipeline plots."""
+
+
+def create_colour_dict():
+    """Build a colour mapping dictionary for specialty categories.
+
+    Returns a nested dictionary with two keys:
+
+    * ``"single"`` – maps each specialty name to a single hex colour.
+    * ``"spectrum"`` – maps each specialty name to a
+      :class:`~matplotlib.colors.LinearSegmentedColormap` that runs from
+      a pale tint to the corresponding single colour at full intensity.
+
+    Returns
+    -------
+    dict
+        ``{"single": {<specialty>: <hex>}, "spectrum": {<specialty>: <cmap>}}``
+    """
+    spec_colour_dict = {
+        "single": {
+            "medical": "#ED7D31",
+            "surgical": "#70AD47",
+            "haem/onc": "#FFC000",
+            "paediatric": "#5B9BD5",
+            "all": "#44546A",
+            "window": "#A9A9A9",
+        },
+        "spectrum": {},
+    }
+
+    def _generate_continuous_colormap(base_color, start_fraction=0.2):
+        base = mcolors.to_rgb(base_color)
+        # Blend from a tinted white towards the full base colour.
+        # start_fraction=0 gives pure white at the low end;
+        # higher values (e.g. 0.2–0.6) make the pale end more saturated.
+        start = tuple(1.0 - start_fraction * (1.0 - c) for c in base)
+        cdict = {
+            "red": [(0.0, start[0], start[0]), (1.0, base[0], base[0])],
+            "green": [(0.0, start[1], start[1]), (1.0, base[1], base[1])],
+            "blue": [(0.0, start[2], start[2]), (1.0, base[2], base[2])],
+        }
+        return mcolors.LinearSegmentedColormap(
+            "custom_colormap", segmentdata=cdict, N=256
+        )
+
+    for spec, color in spec_colour_dict["single"].items():
[…]
