### 2026-02-11 b64a85b
script for pipeline plots
