+++ src/patientflow/viz/pipeline_plots.py
@@ -19,9 +19,15 @@ build_pipeline_prediction_inputs : function
 in_ed_now_plot : function
     Scatter plot of patients currently in the ED, optionally coloured by
     predicted probability of admission.
+main : function
+    End-to-end pipeline that trains models, selects a random snapshot,
+    and generates all overview plots (a, b, c, e, f, h, i, j, k) to a
+    timestamped output folder.
 """
 
+import argparse
 import os
+from datetime import timedelta
 from pathlib import Path
 
 import matplotlib.colors as mcolors
@@ -334,7 +340,6 @@ def in_ed_now_plot(
     size=50,
     preds_col="preds",
     colour_map="Spectral_r",
-    title_suffix="admission",
     return_figure=False,
     ax=None,
 ):
@@ -381,9 +386,6 @@ def in_ed_now_plot(
     colour_map : str or Colormap, optional
         Matplotlib colour map (or name) used when *colour* is ``True``.
         Default ``"Spectral_r"``.
-    title_suffix : str, optional
-        Appended to *title* when *colour* is ``True``
-        (e.g. ``"admission"``).  Default ``"admission"``.
     return_figure : bool, optional
         If ``True`` return ``(fig, ax)`` instead of calling
         ``plt.show()``.  Default ``False``.
@@ -420,7 +422,6 @@ def in_ed_now_plot(
             fig = ax.figure
         else:
             fig, target_ax = plt.subplots(figsize=(figsize_x, figsize_y))
-        title = title + " with predicted probability of " + title_suffix
         for location, group in ex.groupby("loc_new", observed=True):
             jittered_y = loc_to_num[location] + np.random.uniform(
                 -jitter_amount, jitter_amount, size=len(group)
@@ -490,3 +491,493 @@ def in_ed_now_plot(
 
     plt.show()
     return None
+
+
+def _save_fig(fig, save_dir, filename, label):
+    """Save a figure to *save_dir* and close it."""
+    fig.savefig(save_dir / filename, dpi=300)
+    plt.close(fig)
+    print(f"  Saved figure ({label})")
+
+
+def main(
+    prediction_time=(9, 30),
+    output_dir="pipeline_plots_output",
+    data_folder_name="data-public",
+    random_state=0,
+    include_titles=False,
+):
+    """Run the full pipeline-overview plot generation.
+
+    Trains all prediction models, randomly selects a snapshot date from
+    the ED visits data, and generates a set of overview figures saved to
+    a timestamped sub-folder inside *output_dir*.
+
+    The generated figures are:
+
+    * **(a)** Patients currently in the ED (scatter, no colour).
+    * **(b)** Patients currently in the ED coloured by predicted
+      probability of admission.
+    * **(c)** Aggregate probability distribution for number of beds
+      needed.
+    * **(e)** Per-specialty scatter plots using specialty-specific
+      colour maps.
+    * **(f)** Per-specialty (2×2) probability distributions for beds
+      needed from current ED patients (without admission-in-window
+      probability).
+    * **(h)** Patients coloured by probability of admission within the
+      prediction window.
+    * **(i)** Per-specialty (2×2) probability distributions for beds
+      needed with admission-in-window probability.
+    * **(j)** Per-specialty (2×2) probability distributions for beds
+      needed from yet-to-arrive patients only.
+    * **(k)** Per-specialty (2×2) probability distributions for beds
+      needed combining current ED and yet-to-arrive patients.
+
+    Parameters
+    ----------
+    prediction_time : tuple of (int, int), optional
+        ``(hour, minute)`` of the prediction moment.
+        Default ``(9, 30)``.
+    output_dir : str or Path, optional
+        Root directory in which a dated sub-folder will be created for
[…]
