### 2026-02-16 53760fe
minor formatting changes to pipeline plots
