### 2026-02-16 e97a2c3
Replace preamble in 4c/4d/4e and move legacy code to appendix
