+++ src/patientflow/viz/pipeline_plots.py
@@ -47,12 +47,12 @@ DEFAULT_LOCATION_MAPPING = {
     "resus": "Majors/Resus",
     "taf": "Other",
 }
-"""dict : Default mapping from raw ``current_location_type`` values to
-simplified display categories used on the y-axis of pipeline plots."""
+#: Default mapping from raw ``current_location_type`` values to
+#: simplified display categories used on the y-axis of pipeline plots.
 
 DEFAULT_CATEGORY_ORDER = ["Majors/Resus", "Minors", "Other"]
-"""list of str : Default ordered list of display category names for the
-y-axis of pipeline plots."""
+#: Default ordered list of display category names for the
+#: y-axis of pipeline plots.
 
 
 def create_colour_dict():
@@ -62,7 +62,7 @@ def create_colour_dict():
 
     * ``"single"`` – maps each specialty name to a single hex colour.
     * ``"spectrum"`` – maps each specialty name to a
-      :class:`~matplotlib.colors.LinearSegmentedColormap` that runs from
+      `matplotlib.colors.LinearSegmentedColormap` that runs from
       a pale tint to the corresponding single colour at full intensity.
 
     Returns
@@ -174,20 +174,20 @@ def add_specialty_predictions(
     """Attach a per-row specialty probability dict to a snapshot DataFrame.
 
     For each row the trained *specialty_model* is called with the full row
-    (as a :class:`pandas.Series`) so that models which need additional
-    context (e.g. :class:`~patientflow.predictors.subgroup_predictor.MultiSubgroupPredictor`)
+    (as a `pandas.Series`) so that models which need additional
+    context (e.g. [MultiSubgroupPredictor][patientflow.predictors.subgroup_predictor.MultiSubgroupPredictor])
     can determine the correct sub-model.  The result is stored in a new
     ``specialty_prob`` column whose values are ``dict[str, float]``.
 
     Parameters
     ----------
     df : pandas.DataFrame
-        Snapshot data – typically the output of :func:`prepare_snapshot_data`.
+        Snapshot data – typically the output of [prepare_snapshot_data][patientflow.viz.pipeline_plots.prepare_snapshot_data].
         Must contain at least the column used by
         ``specialty_model.input_var`` (usually ``consultation_sequence``).
     specialty_model : object
         Trained specialty prediction model.  Must expose ``input_var``
-        (str) and ``predict(row)`` where *row* is a :class:`pandas.Series`.
+        (str) and ``predict(row)`` where *row* is a `pandas.Series`.
     specialties : list of str or None, optional
         Specialty names to ensure are present in every returned dict
         (missing keys are filled with ``0``).
@@ -239,9 +239,9 @@ def build_pipeline_prediction_inputs(
     """Build per-service prediction inputs for a snapshot.
 
     This is a convenience wrapper around
-    :func:`~patientflow.predict.service.build_service_data` that unpacks
+    [build_service_data][patientflow.predict.service.build_service_data] that unpacks
     the prediction inputs dictionary returned by
-    :func:`~patientflow.train.emergency_demand.prepare_prediction_inputs`
+    [prepare_prediction_inputs][patientflow.train.emergency_demand.prepare_prediction_inputs]
     and retrieves the appropriate admission classifier for the given
     *prediction_time*.
 
@@ -256,7 +256,7 @@ def build_pipeline_prediction_inputs(
     prediction_snapshots : pandas.DataFrame
         Snapshot of patients currently in the ED at the prediction moment.
         Must contain an ``elapsed_los`` column (in seconds – it will be
-        converted to :class:`~datetime.timedelta` internally).
+        converted to [timedelta][datetime.timedelta] internally).
[…]
