+++ docs/notebooks/README.md
@@ -32,8 +32,9 @@ A set of notebooks follow, that show how we have used the functions in `patientf
 
 - **[4a_Specify_emergency_demand_model](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4a_Specify_emergency_demand_model.md):** Explains design choices that were made to develop a practical model, and shows an example of the output that is sent five times a day at UCLH.
 - **[4b_Predict_emergency_demand](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4b_Predict_emergency_demand.md):** Shows a full implementation of the functions covered up to this point, to predict emergency demand at UCLH.
-- **[4c_Evaluate_emergency_demand_predictions](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4b_Predict_demand_from_patients_in_ED.md)** Shows an evaluation of the emergency demand predictions generated using the public data provided with this repo.
-- **[4d_Predict_emergency_demand_with_special_categories](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4d_Predict_emergency_demand_with_special_categories.md)** Develops the logic shown in the previous notebook to include the handling of particular sub-groups of patients differently.
+- **[4c_Evaluate_emergency_demand_predictions](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4c_Evaluate_emergency_demand_predictions.md):** Shows an evaluation of the emergency demand predictions generated using the public data provided with this repo.
+- **[4d_Predict_emergency_demand_for_sub-groups](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4d_Predict_emergency_demand_for_sub-groups.ipynb)** Develops the logic shown in the previous notebook to include the handling of particular sub-groups of patients differently.
+- **[4e_Generate_predictions_using_hierarchy](https://github.com/UCL-CORU/patientflow/blob/main/notebooks/4e_Generate_predictions_using_hierarchy.md)** Shows the use of a hierarchical approach to generate demand predictions at different levels of a hospital's reporting hierarchy.
 
 ## Preparing your notebook environment
 

[…]
