### 2026-02-17 bf9e0d3
add notebook on production pipeline
