### 2026-02-17 d9e8016
Improve notebooks - rename 3c to 3e and change title to refer to yet-to-arrive - rename 3d to 3c and change title to refer to hospital service
- create new 3d showing evaluation of bed demand by service
- create new 3f showing evaluation of yet-to-arrive
