+++ docs/notebooks/4c_Evaluate_emergency_demand_predictions.md
@@ -1,18 +1,19 @@
 # 4c. Evaluate predictions of emergency demand
 
-The previous notebook demonstrated the full implementation in code of training models for use at UCLH using the functions provided in `patientflow`. I demonstrated how we create predictions for a single snapshot.
+The previous notebook demonstrated the full implementation in code of training models for use at UCLH using the functions provided in `patientflow`. I demonstrated how we create predictions for a single snapshot. 
 
-This notebooks show evaluate the predicted distributions against observed numbers, for the whole test set. I will evaluate the predictions by specialty using the approaches covered in a [previous notebook](3b_Evaluate_group_snapshots.md).
+This notebooks show evaluate the predicted distributions against observed numbers, for the whole test set. I will evaluate the predictions by specialty using the approaches covered in notebook 3b.
 
-- Histograms of observed versus expected values
-- Adjusted QQ plots
+* Histograms of observed versus expected values
+* Adjusted QQ plots
 
 As the predictions for yet-to-arrive patients are aspirational, these cannot be directed evaluated against observed numbers of admissions in the prediction window. In reality, due to poor ED performance, few may have been admitted within the window. Similarly for the group of patients in the ED, we calculate the predicted number of beds needed within the prediction window, but the observed numbers will not reflect the targets.
 
-We can, however, evaluate the predictions in a slightly different way.
+We can, however, evaluate the predictions in a slightly different way. 
 
-- For the patients in ED, we can compare the predicted bed counts needed for each specialty against observed numbers admitted to each specialty from among patients comprising each group snapshot in the test set period, without taking into account how long it takes each patient to be admitted.
-- For the yet-to-arrive patients, we can compare the predicted with the observed arrival rates within the prediction window.
+* For the patients in ED, we can compare the predicted bed counts needed for each specialty against observed numbers admitted to each specialty from among patients comprising each group snapshot in the test set period, without taking into account how long it takes each patient to be admitted.
+* For the yet-to-arrive patients, we can compare the predicted with the observed arrival rates within the prediction window.
+
 
 ```python
 # Reload functions every time
@@ -22,9 +23,10 @@ We can, however, evaluate the predictions in a slightly different way.
 
 ## Load data and train models
 
-The data loading, configuration, and model training steps are identical to those demonstrated in detail in [notebook 4b](4b_Predict_emergency_demand.md). Here we use `prepare_prediction_inputs` to perform all of these steps in a single call.
+The data loading, configuration, and model training steps are identical to those demonstrated in detail in notebook 4b. Here we use `prepare_prediction_inputs` to perform all of these steps in a single call.
 
 You can request the UCLH datasets on [Zenodo](https://zenodo.org/records/14866057). If you don't have the public data, change `data_folder_name` from `'data-public'` to `'data-synthetic'`.
+
 
 ```python
 from patientflow.train.emergency_demand import prepare_prediction_inputs
@@ -48,18 +50,20 @@ params = prediction_inputs['config']
 
     Split sizes: [62071, 10415, 29134]
     Split sizes: [7716, 1285, 3898]
-
+    
     Processing: (6, 0)
-
+    
     Processing: (9, 30)
-
+    
     Processing: (12, 0)
-
+    
     Processing: (15, 30)
-
+    
     Processing: (22, 0)
 
-Below we use the training, validation and test sets dates set in config.yaml to retrieve the portion of the data that is the test set.
+
[…]
