+++ docs/notebooks/3a_Prepare_group_snapshots.md
@@ -3,9 +3,8 @@
 Collecting patient snapshots together into a group snapshot is useful when predicting a bed count distribution at a point in time. A group snapshot is a subset of patients who were in the ED on a single snapshot date, at a specific prediction time.
 
 In this notebook, I show how `patientflow` can be used to prepare group snapshots for current patients, by dividing patient snapshots into their groups, and storing the group snapshots as dictionary with:
-
-- `snapshot_date` as the key
-- `snapshot_ids` of each patient snapshot as the values
+* `snapshot_date` as the key
+* `snapshot_ids` of each patient snapshot as the values 
 
 This structure is a convenient way to organise the data when making bed count predictions for different snapshot dates and prediction times, and especially when evaluating those predictions (see next notebook).
 
@@ -15,34 +14,40 @@ In this notebook I use fake data that resemble visits to the Emergency Departmen
 
 To start with a very simple example, I apply a series of Bernoulli trials to one group snapshot and visualise the bed count distribution reflecting the sum of these trials. In that simple example, every patient has the same probability of the outcome.
 
-I then train a model that predicts a probability of admission for each patient within a group snapshot. This is a more realistic example, as each patient can have a different probability of admission, based on their data. I apply Bernoulli trials and again visualise the predicted distribution for the total number of beds needed for those patients.
+I then train a model that predicts a probability of admission for each patient within a group snapshot. This is a more realistic example, as each patient can have a different probability of admission, based on their data. I apply Bernoulli trials and again visualise the predicted distribution for the total number of beds needed for those patients. 
 
-I demonstrate functions in `patientflow` that handle the preparation of group snapshots for making predictions.
+I demonstrate functions in `patientflow` that handle the preparation of group snapshots for making predictions. 
+
 
 ```python
 # Reload functions every time
-%load_ext autoreload
+%load_ext autoreload 
 %autoreload 2
 ```
 
     The autoreload extension is already loaded. To reload it, use:
       %reload_ext autoreload
 
+
 ## Generate fake snapshots
 
-See [a previous notebook](2a_Create_patient_snapshots.md) for more background on this.
+See notebook 2a for more background on this. 
+
 
 ```python
 from patientflow.generate import create_fake_snapshots
 
-prediction_times = [(6, 0), (9, 30), (12, 0), (15, 30), (22, 0)]
-snapshots_df=create_fake_snapshots(prediction_times=prediction_times,
-                                   start_date='2023-01-01',
+prediction_times = [(6, 0), (9, 30), (12, 0), (15, 30), (22, 0)] 
+snapshots_df=create_fake_snapshots(prediction_times=prediction_times, 
+                                   start_date='2023-01-01', 
                                    end_date='2023-04-01',
                                    mean_patients_per_day=100)
 snapshots_df.head()
 
 ```
+
+
+
 
 <div>
 <style scoped>
@@ -57,7 +62,6 @@ snapshots_df.head()
     .dataframe thead th {
         text-align: right;
     }
-
 </style>
 <table border="1" class="dataframe">
   <thead>
@@ -172,7 +176,9 @@ snapshots_df.head()
 </table>
 </div>
 
-Note that each record in the snapshots dataframe is indexed by a unique snapshot_id.
+
+
+Note that each record in the snapshots dataframe is indexed by a unique snapshot_id. 
 
 ## Prepare group snapshots
 
[…]

[…]
