### 2026-02-19 c4d3959
move evaluation to 3x notebooks; focus 4x notebooks on production
