### 2026-02-20 e3be4d8
make colour in notebook charts meaningful
