### 2026-02-20 7983413
use refresh notebooks skill
