### 2026-02-20 c2bc670
loosen validation of specialties match on inference
