+++ src/patientflow/predict/service.py
@@ -23,6 +23,7 @@ from patientflow.predict.emergency_demand import (
     add_missing_columns,
     get_specialty_probs,
 )
+from patientflow.predict.validation import warn_specialty_mismatch
 from patientflow.predictors.incoming_admission_predictors import (
     ParametricIncomingAdmissionPredictor,
     EmpiricalIncomingAdmissionPredictor,
@@ -460,10 +461,11 @@ def _validate_models_and_data(
 
     # Validate specialties alignment
     if yet_to_arrive_model is not None and hasattr(yet_to_arrive_model, "filters"):
-        if not set(yet_to_arrive_model.filters.keys()) == set(specialties):
-            raise ValueError(
-                f"Requested specialties {set(specialties)} do not match the specialties of the trained yet-to-arrive model {set(yet_to_arrive_model.filters.keys())}"
-            )
+        warn_specialty_mismatch(
+            set(specialties),
+            set(yet_to_arrive_model.filters.keys()),
+            "yet-to-arrive model",
+        )
 
     special_params = spec_model.special_params if spec_model is not None else None
 
@@ -475,15 +477,16 @@ def _validate_models_and_data(
     if special_category_dict is not None and not set(specialties) == set(
         special_category_dict.keys()
     ):
-        # Only enforce the legacy check if there is no subgroup mapping available
         has_mapping = (
             hasattr(spec_model, "specialty_to_subgroups")
             and isinstance(getattr(spec_model, "specialty_to_subgroups"), dict)
             and len(getattr(spec_model, "specialty_to_subgroups")) > 0
         )
         if not has_mapping:
-            raise ValueError(
-                "Requested specialties do not match the specialty dictionary defined in special_params"
+            warn_specialty_mismatch(
+                set(specialties),
+                set(special_category_dict.keys()),
+                "special_category_dict",
             )
 
 
