+++ docs/notebooks/2c_Evaluate_patient_snapshot_models.md
@@ -16,7 +16,7 @@ We don't focus as much on typical classification metrics like Area under the ROC
 The ultimate goal is to predict bed count distributions for groups of patients. Bed count distributions will be calculated in two steps
 
 1. First, we predict the probability of the outcome we are interested in (admission or discharge) for each individual patient, as shown in previous notebooks.
-2. Then, we use these probabilities in Bernoulli trials to get bed count distributions. The Bernouill trials step will be shown in later notebooks.
+2. Then, we use these probabilities in Bernoulli trials to get bed count distributions. The Bernoulli trials step will be shown in later notebooks.
 
 Because of this approach, the accuracy of the probability values matters more than correct classification. That is why we use log loss to optimise our classifiers.
 
@@ -61,7 +61,7 @@ ed_visits = load_data(data_file_path,
 
     Inferred project root: /Users/zellaking/Repos/patientflow
 
-Inspecting the data that has been loaded, we can see that it is similar in structure to the fake data that was generated on the fly in the previous notebooks. The dates have been pushed into the future, to minimise the likelihood of re-identifcation of patients.
+Inspecting the data that has been loaded, we can see that it is similar in structure to the fake data that was generated on the fly in the previous notebooks. The dates have been pushed into the future, to minimise the likelihood of re-identification of patients.
 
 The dates for training, validation and test sets that match this dataset are defined in the config file in the root directory of `patientflow`.
 
@@ -114,7 +114,7 @@ train_visits, valid_visits, test_visits = create_temporal_splits(
 
     Split sizes: [62071, 10415, 29134]
 
-Next we specify the times of day at which are predictions are to be made. Here I'm deriving from the dataset. Note that there are many more snapshots in the later part of the day 12:00, 15:30 and 22:00
+Next we specify the times of day at which our predictions are to be made. Here I'm deriving from the dataset. Note that there are many more snapshots in the later part of the day 12:00, 15:30 and 22:00
 
 ```python
 prediction_times = ed_visits.prediction_time.unique()
@@ -142,7 +142,7 @@ Define ordinal mappings where appropriate. These include:
 - `age_group` - Age on arrival at the ED, defined in groups
 - `latest_obs_manchester_triage_acuity` - Manchester Triage Score (where blue is the lowest acuity and red the highest)
 - `latest_obs_objective_pain_score` - ranging from nil to very severe
-- `latest_obs_level_of_consciousness` the ACVPU measure of consciousness, where A (aware) and U (unconscious) at are the extremes.
+- `latest_obs_level_of_consciousness` — the ACVPU measure of consciousness, where A (aware) and U (unconscious) are the extremes.
 
 ```python
 ordinal_mappings = {
@@ -370,7 +370,7 @@ for prediction_time in prediction_times:
 
     Training model for (9, 30)
 
-From the plots below, we see improved discrimination. There are positive cases clustered at the right hand end of the distribution plot. However, this gain has come at the cost of much worse calibration when the models are applied to the whole test set, without undersampling the majority class, as shown in the calibation plot and MADCAP plots.
+From the plots below, we see improved discrimination. There are positive cases clustered at the right hand end of the distribution plot. However, this gain has come at the cost of much worse calibration when the models are applied to the whole test set, without undersampling the majority class, as shown in the calibration plot and MADCAP plots.
 
 ```python
 from patientflow.viz.estimated_probabilities import plot_estimated_probabilities
@@ -476,7 +476,7 @@ plot_madcap(
 
 It can be useful to look at sub-categories of patients, to understand whether models perform better for some groups. Here we show MADCAP plots by age group.
 
[…]
