### 2026-02-23 5079976
implement get_prob_dist_by_service to support evaluation
