### 2026-02-23 4b3fc8a
Revert "implement get_prob_dist_by_service to support evaluation"
This reverts commit 507997616e70b5254ff41b97a0fe3ca5b710db2b.
