+++ src/patientflow/aggregate.py
@@ -43,11 +43,6 @@ get_prob_dist : function
 get_prob_dist_using_survival_curve : function
     Calculate probability distributions for each snapshot date based on given model predictions, using a survival curve to predict the probability of each patient being admitted within a given prediction window.
 
-get_prob_dist_by_service : function
-    Evaluate composed service-level predictions across the test set for one or more
-    services, producing probability distributions and observed values in the standard
-    evaluation format.
-
 """
 
 import pandas as pd
@@ -453,14 +448,28 @@ def get_prob_dist(
         )
 
     prob_dist_dict = {}
+    if verbose:
+        print(
+            f"Calculating probability distributions for {len(snapshots_dict)} snapshot dates"
+        )
+
+        if len(snapshots_dict) > 10:
+            print(
+                "Using efficient generating function approach - much faster than before!"
+            )
+
+    # Initialize a counter for notifying the user every 10 snapshot dates processed
+    count = 0
 
     for dt, snapshots_to_include in snapshots_dict.items():
         if len(snapshots_to_include) == 0:
+            # Create an empty dictionary for the current snapshot date
             prob_dist_dict[dt] = {
                 "agg_predicted": pd.DataFrame({"agg_proba": [1]}, index=[0]),
                 "agg_observed": 0,
             }
         else:
+            # Ensure the lengths of test features and outcomes are equal
             assert len(X_test.loc[snapshots_to_include]) == len(
                 y_test.loc[snapshots_to_include]
             ), "Mismatch in lengths of X_test and y_test snapshots."
@@ -470,6 +479,7 @@ def get_prob_dist(
             else:
                 prediction_moment_weights = weights.loc[snapshots_to_include].values
 
+            # Apply category filter
             if category_filter is None:
                 prediction_moment_category_filter = None
             else:
@@ -477,6 +487,7 @@ def get_prob_dist(
                     snapshots_to_include
                 ]
 
+            # Use the refactored generating function approach
             prob_dist_dict[dt] = get_prob_dist_for_prediction_moment(
                 X_test=X_test.loc[snapshots_to_include],
                 y_test=y_test.loc[snapshots_to_include],
@@ -485,6 +496,11 @@ def get_prob_dist(
                 category_filter=prediction_moment_category_filter,
                 normal_approx_threshold=normal_approx_threshold,
             )
+
+        # Increment the counter and notify the user every 10 snapshot dates processed
+        count += 1
+        if verbose and count % 10 == 0 and count != len(snapshots_dict):
+            print(f"Processed {count} snapshot dates")
 
     if verbose:
         print(f"Processed {len(snapshots_dict)} snapshot dates")
@@ -562,7 +578,12 @@ def get_prob_dist_using_survival_curve(
         )
 
     prob_dist_dict = {}
+    if verbose:
+        print(
+            f"Calculating probability distributions for {len(snapshot_dates)} snapshot dates"
+        )
 
+    # Create prediction context that will be the same for all dates
     prediction_context = {category: {"prediction_time": prediction_time}}
 
     for dt in snapshot_dates:
@@ -600,276 +621,3 @@ def get_prob_dist_using_survival_curve(
         print(f"Processed {len(snapshot_dates)} snapshot dates")
 
     return prob_dist_dict
-
-
-def prediction_to_eval_dict(
-    probabilities: "np.ndarray",
-    observed: int,
-) -> Dict[str, Any]:
-    """Convert a production-format probability array and observed count to the
-    evaluation dictionary format used by the visualisation functions.
-
-    Parameters
-    ----------
-    probabilities : np.ndarray
[…]
