### 2026-02-23 bbb0e90
add aspirational flag to FlowInputs
