### 2026-02-23 1ae1eba
limit inference specialties to training specialties
