### 2026-03-02 a95acc5
ruff reforat
