### 2026-03-04 218a6ef
add script to run evaluation and update notebook 4d
