+++ src/patientflow/evaluate.py
@@ -1,46 +1,31 @@
-"""
-Patient Flow Evaluation Module
+"""Evaluation orchestration and scalar utilities for patient flow models.
 
-This module provides functions for evaluating and comparing different prediction models
-for non-clincal outcomes in a healthcare setting. It includes utilities for calculating
-metrics such as Mean Absolute Error (MAE) and Mean Percentage Error (MPE), as well as
-functions for predicting admissions based on historical data and combining different
-prediction models.
+This module provides:
 
-Functions
----------
-calculate_results : function
-    Calculate evaluation metrics based on expected and observed values
-calc_mae_mpe : function
-    Calculate MAE and MPE for probability distribution predictions
-calculate_admission_probs_relative_to_prediction : function
-    Calculate admission probabilities for arrivals relative to a prediction time window
-get_arrivals_with_admission_probs : function
-    Get arrivals before and after prediction time with their admission probabilities
-calculate_weighted_observed : function
-    Calculate actual admissions assuming ED targets are met
-create_time_mask : function
-    Create a mask for times before/after a specific hour:minute
-predict_using_previous_weeks : function
-    Predict admissions using average from previous weeks
-evaluate_six_week_average : function
-    Evaluate the six-week average prediction model
-combine_distributions : function
-    Combine two probability distributions using convolution
-evaluate_combined_model : function
-    Evaluate a combined prediction model
-
+1. Backwards-compatible scalar helpers used by existing notebooks:
+   ``calculate_results`` and ``calc_mae_mpe``.
+2. A target-driven evaluation runner (``run_evaluation``) that writes
+   plots and scalar summaries to a structured output directory.
 """
 
-from typing import Dict, List, Any, Union, Tuple
+from typing import Dict, List, Any, Union, Tuple, Optional
 import numpy as np
-import pandas as pd
-from datetime import datetime, timedelta
-from scipy import signal
-from patientflow.calculate.admission_in_prediction_window import (
-    get_y_from_aspirational_curve,
-)
+from datetime import datetime, timedelta, timezone
+from dataclasses import dataclass
+from pathlib import Path
+import json
+import shutil
+import re
+import matplotlib.pyplot as plt
+
 from patientflow.load import get_model_key
+from patientflow.viz.madcap import plot_madcap
+from patientflow.viz.estimated_probabilities import plot_estimated_probabilities
+from patientflow.viz.calibration import plot_calibration
+from patientflow.viz.epudd import plot_epudd
+from patientflow.viz.observed_against_expected import plot_deltas, plot_arrival_deltas
+from patientflow.viz.survival_curve import plot_admission_time_survival_curve
+from patientflow.predict.types import FlowSelection
 
 
 def calculate_results(
@@ -161,9 +146,11 @@ def calc_mae_mpe(
 
     # Sort results by prediction time
     def get_time_value(key: str) -> int:
-        # Extract time from key (e.g., 'admissions_1530' -> 1530)
-        time_str = key.split("_")[1]
-        return int(time_str)
+        # Extract trailing HHMM/HMM token if present (e.g. admissions_1530, flow_x_930).
+        match = re.search(r"(\d{3,4})$", key)
+        if match:
+            return int(match.group(1))
+        return 99999
 
     # Create sorted dictionary
     sorted_results = dict(
@@ -173,567 +160,937 @@ def calc_mae_mpe(
     return sorted_results
 
 
-def calculate_admission_probs_relative_to_prediction(
-    df, prediction_datetime, prediction_window, x1, y1, x2, y2, is_before=True
-):
-    """Calculate admission probabilities for arrivals relative to a prediction time window.
+@dataclass(frozen=True)
+class EvaluationTarget:
+    """Configuration for evaluating one named target.
 
     Parameters
[…]
