### 2026-03-04 0d67aa5
add script to run evaluation and update notebook 4d
