+++ tests/test_evaluate.py
@@ -1,0 +1,352 @@
+"""Tests for the patientflow.evaluate module.
+
+Tier 1: Unit tests for calculate_results and data-structure helpers.
+Tier 2: Orchestration tests verifying correct routing and scalar output.
+"""
+
+import json
+import tempfile
+import shutil
+import unittest
+from pathlib import Path
+from types import SimpleNamespace
+
+import numpy as np
+
+from patientflow.evaluate import (
+    EvaluationTarget,
+    _metrics_from_training_artifacts,
+    _upsert_scalar_metadata,
+    calculate_results,
+    evaluate_classifier,
+    evaluate_flow,
+    run_evaluation,
+)
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
+def _make_evaluation_target(**overrides):
+    defaults = dict(
+        name="test_target",
+        flow_type="pmf",
+        aspirational=False,
+        components=["arrivals"],
+        flow_selection=None,
+        evaluation_mode="distribution",
+    )
+    defaults.update(overrides)
+    return EvaluationTarget(**defaults)
+
+
+def _make_trained_model(prediction_time=(9, 30), metrics=None):
+    """Build a minimal mock trained model with the given metrics."""
+    selected = metrics or {
+        "log_loss": 0.35,
+        "auroc": 0.88,
+        "auprc": 0.72,
+        "n_samples": 200,
+        "n_positive_cases": 80,
+    }
+    training_results = SimpleNamespace(
+        prediction_time=prediction_time,
+        selected_eval_metrics=selected,
+    )
+    return SimpleNamespace(
+        training_results=training_results,
+        selected_eval_metrics=selected,
+    )
+
+
+# ---------------------------------------------------------------------------
+# Tier 1 – Core logic and data-structure helpers
+# ---------------------------------------------------------------------------
+
+
+class TestCalculateResults(unittest.TestCase):
+    """Tests for calculate_results."""
+
+    def test_basic_computation(self):
+        result = calculate_results([10, 20, 30], [12.0, 18.0, 33.0])
+        self.assertAlmostEqual(result["mae"], np.mean([2, 2, 3]))
+        self.assertGreater(result["mpe"], 0)
+
+    def test_perfect_prediction(self):
+        result = calculate_results([5, 10, 15], [5.0, 10.0, 15.0])
+        self.assertAlmostEqual(result["mae"], 0.0)
+        self.assertAlmostEqual(result["mpe"], 0.0)
+
+    def test_empty_inputs(self):
+        result = calculate_results([], [])
+        self.assertEqual(result["mae"], 0.0)
+        self.assertEqual(result["mpe"], 0.0)
+
+    def test_observed_zeros_excluded_from_mpe(self):
+        result = calculate_results([5, 0], [0.0, 0.0])
+        self.assertAlmostEqual(result["mpe"], 0.0)
+        self.assertAlmostEqual(result["mae"], 2.5)
+
+
+class TestUpsertScalarMetadata(unittest.TestCase):
+    """Tests for _upsert_scalar_metadata."""
+
+    def test_basic_insertion(self):
+        scalars = {}
+        target = _make_evaluation_target()
+        _upsert_scalar_metadata(
+            scalars=scalars,
+            flow_name="my_flow",
+            evaluation_target=target,
+            prediction_time_key="0930",
+            evaluated=True,
+            component_name="arrivals",
+        )
+        node = scalars["my_flow"]
+        self.assertEqual(node["flow_type"], "pmf")
+        time_node = node["components"]["arrivals"]["prediction_times"]["0930"]
+        self.assertTrue(time_node["evaluated"])
+
[…]

[…]
