### 2026-03-05 ecf6fc1
add prettier formatting to tox docs build to match pre-commit hook
