### 2026-03-05 d0e97f4
Improve specialty mismatch handling in transfer predictor
- Separate services and subgroups into distinct keys in the transfer
 model's internal dict to prevent subgroup names leaking into error
 messages.
- Return 0.0 with a warning for unknown services instead of
- raising.
- Shorten warn_specialty_mismatch to report counts only.
