+++ src/patientflow/predictors/transfer_predictor.py
@@ -29,6 +29,7 @@ When cohort_col is specified, separate probability models are trained for each
 cohort. When cohort_col is None, all data is used together to train a single model.
 """
 
+import warnings
 import pandas as pd
 from typing import Dict, Set, Optional, Callable, Union
 from sklearn.base import BaseEstimator, TransformerMixin
@@ -81,34 +82,17 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
     cohort_values : list or None
         List of expected cohort values for validation, or None
     transfer_probabilities : dict
-        Nested dictionary containing transfer statistics. Structure depends on
-        whether cohorts are used:
+        Nested dictionary of transfer statistics, keyed by cohort name.
+        Each cohort contains two sub-dicts:
 
-        If cohort_col is None:
-        {
-            'all': {
-                'source_service': {
-                    'prob_transfer': float,
-                    'destination_distribution': dict
-                }
-            }
-        }
+        - ``"services"`` maps each source service to a dict with
+          ``"prob_transfer"`` (float) and ``"destination_distribution"``
+          (dict of destination to probability).
+        - ``"subgroups"`` maps each subgroup name to a dict with the
+          same per-service structure as ``"services"``.
 
-        If cohort_col is specified:
-        {
-            'cohort_name': {
-                'source_service': {
-                    'prob_transfer': float,
-                    'destination_distribution': dict
-                },
-                'subgroup_name': {
-                    'source_service': {
-                        'prob_transfer': float,
-                        'destination_distribution': dict
-                    }
-                }
-            }
-        }
+        When ``cohort_col`` is None, a single cohort named ``"all"``
+        is used.
     services : set
         Set of all services in the system
     cohorts : set or None
@@ -170,9 +154,7 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         self.cohort_col = cohort_col
         self.cohort_values = cohort_values
         self.subgroup_functions = subgroup_functions or create_subgroup_functions()
-        self.transfer_probabilities: Optional[Dict[str, Dict[str, Dict[str, Dict]]]] = (
-            None
-        )
+        self.transfer_probabilities: Optional[Dict[str, Dict[str, Dict]]] = None
         self.services: Optional[Set[str]] = None
         self.cohorts: Optional[Set[str]] = None
         self.subgroups: Optional[Set[str]] = None
@@ -197,45 +179,32 @@ class TransferProbabilityEstimator(BaseEstimator, TransformerMixin):
         n_cohorts = len(self.cohorts) if self.cohorts is not None else 0
         n_subgroups = len(self.subgroups) if self.subgroups is not None else 0
 
-        n_with_transfers = 0
-        cohort_subspecialty_counts: Dict[str, Dict[str, int]] = {}
         services_with_transfers: Set[str] = set()
+        cohort_subspecialty_counts: Dict[str, Dict[str, int]] = {}
 
         if self.transfer_probabilities is not None:
-            # Process each cohort separately to get cohort-specific counts
             for cohort_name, cohort_data in self.transfer_probabilities.items():
-                cohort_subspecialty_counts[cohort_name] = {}
-                for key, value in cohort_data.items():
-                    # Check if this is subspecialty data (has prob_transfer at top level)
-                    if isinstance(value, dict) and "prob_transfer" in value:
-                        prob_transfer = value["prob_transfer"]
-                        if (
-                            isinstance(prob_transfer, (int, float))
-                            and prob_transfer > 0
-                        ):
-                            services_with_transfers.add(
-                                key
-                            )  # Track unique services with transfers
-                    # Check if this is subgroup data (contains nested subspecialty data)
-                    elif isinstance(value, dict) and any(
-                        isinstance(v, dict) and "prob_transfer" in v
-                        for v in value.values()
[…]
