+++ src/patientflow/evaluation/builder.py
@@ -1,0 +1,228 @@
+"""Builder utilities for assembling ``EvaluationInputs``.
+
+The builder in this module converts common upstream outputs into the nested
+typed input structure expected by the evaluation runner, avoiding manual
+dictionary reshaping in notebooks and scripts.
+
+Supported assembly paths include:
+
+- classifier payload registration
+- distribution payload ingestion from service-level dictionaries
+- arrival-delta payload registration
+- survival-curve payload registration
+"""
+
+from __future__ import annotations
+
+from datetime import date, timedelta
+from typing import Any, Dict, Iterable, List, Mapping, Optional, Tuple
+
+import pandas as pd
+
+from patientflow.evaluation.adapters import from_legacy_prob_dist_dict
+from patientflow.evaluation.targets import get_default_evaluation_targets
+from patientflow.evaluation.types import (
+    ArrivalDeltaPayload,
+    ClassifierInput,
+    EvaluationInputs,
+    EvaluationTarget,
+    SurvivalCurvePayload,
+)
+from patientflow.load import get_model_key
+
+
+class EvaluationInputsBuilder:
+    """Helper for assembling typed evaluation inputs from pipeline outputs.
+
+    Parameters
+    ----------
+    prediction_times
+        Prediction times that should be represented in the built inputs.
+    evaluation_targets
+        Optional explicit target registry. If omitted, defaults are used.
+    """
+
+    def __init__(
+        self,
+        prediction_times: List[Tuple[int, int]],
+        evaluation_targets: Optional[Dict[str, EvaluationTarget]] = None,
+    ):
+        self.prediction_times = prediction_times
+        self.evaluation_targets = evaluation_targets or get_default_evaluation_targets()
+        self.classifier_inputs: Dict[str, ClassifierInput] = {}
+        self.flow_inputs_by_service: Dict[str, Dict[str, Dict[Tuple[int, int], Any]]] = {}
+
+    def add_classifier(
+        self,
+        flow_name: str,
+        *,
+        trained_models: Any,
+        visits_df: Optional[pd.DataFrame] = None,
+        label_col: str = "is_admitted",
+    ) -> "EvaluationInputsBuilder":
+        """Register classifier payload for one target.
+
+        Parameters
+        ----------
+        flow_name
+            Classifier target name.
+        trained_models
+            Iterable or mapping of trained classifier artifacts.
+        visits_df
+            Evaluation dataframe used by classifier diagnostics.
+        label_col
+            Binary outcome label column in ``visits_df``.
+
+        Returns
+        -------
+        EvaluationInputsBuilder
+            Builder instance for fluent chaining.
+        """
+        models = list(trained_models.values()) if isinstance(trained_models, dict) else list(trained_models)
+        self.classifier_inputs[flow_name] = ClassifierInput(
+            trained_models=models,
+            visits_df=visits_df,
+            label_col=label_col,
+        )
+        return self
+
+    def add_distributions_from_service_dict(
+        self,
+        flow_name: str,
+        *,
+        prob_dist_by_service: Mapping[str, Mapping[str, Mapping[date, Mapping[str, Any]]]],
+        model_name: str,
+    ) -> "EvaluationInputsBuilder":
+        """Ingest service distribution outputs for one flow target.
+
+        Parameters
+        ----------
+        flow_name
+            Distribution flow target name.
+        prob_dist_by_service
+            Mapping returned by ``get_prob_dist_by_service`` keyed by model key.
+        model_name
+            Base model name used to derive model keys for prediction times.
+
+        Returns
+        -------
+        EvaluationInputsBuilder
+            Builder instance for fluent chaining.
+        """
+        for prediction_time in self.prediction_times:
+            model_key = get_model_key(model_name, prediction_time)
+            by_service = prob_dist_by_service.get(model_key, {})
[…]
