### 2026-03-16 839fad6
Replace evaluate module with typed evaluate package
- Delete evaluate.py, keeping calculate_results and calc_mae_mpe
- Add evaluate/ package with typed targets, builder, handlers, and runner
- Inline _metrics_from_training_artifacts into handlers
- Define native typed target registry (no legacy dependency)
- Update all source, test, and notebook imports
- Rewrite notebook 4d narrative to document the new workflow
