+++ src/patientflow/evaluate.py
@@ -1,1096 +1,0 @@
-"""Evaluation orchestration and scalar utilities for patient flow models.
-
-This module provides:
-
-1. Backwards-compatible scalar helpers used by existing notebooks:
-   ``calculate_results`` and ``calc_mae_mpe``.
-2. A target-driven evaluation runner (``run_evaluation``) that writes
-   plots and scalar summaries to a structured output directory.
-"""
-
-from typing import Dict, List, Any, Union, Tuple, Optional
-import numpy as np
-from datetime import datetime, timedelta, timezone
-from dataclasses import dataclass
-from pathlib import Path
-import json
-import shutil
-import re
-import matplotlib.pyplot as plt
-
-from patientflow.load import get_model_key
-from patientflow.viz.madcap import plot_madcap
-from patientflow.viz.estimated_probabilities import plot_estimated_probabilities
-from patientflow.viz.calibration import plot_calibration
-from patientflow.viz.epudd import plot_epudd
-from patientflow.viz.observed_against_expected import plot_deltas, plot_arrival_deltas
-from patientflow.viz.survival_curve import plot_admission_time_survival_curve
-from patientflow.predict.types import FlowSelection
-
-
-def calculate_results(
-    expected_values: List[Union[int, float]], observed_values: List[float]
-) -> Dict[str, Union[List[Union[int, float]], float]]:
-    """Calculate evaluation metrics based on expected and observed values.
-
-    Parameters
-    ----------
-    expected_values : List[Union[int, float]]
-        List of expected values.
-    observed_values : List[float]
-        List of observed values.
-
-    Returns
-    -------
-    Dict[str, Union[List[Union[int, float]], float]]
-        Dictionary containing:
-        - expected : List[Union[int, float]]
-            Original expected values
-        - observed : List[float]
-            Original observed values
-        - mae : float
-            Mean Absolute Error
-        - mpe : float
-            Mean Percentage Error
-    """
-    expected_array: np.ndarray = np.array(expected_values)
-    observed_array: np.ndarray = np.array(observed_values)
-
-    if len(expected_array) == 0 or len(observed_array) == 0:
-        return {
-            "expected": expected_values,
-            "observed": observed_values,
-            "mae": 0.0,
-            "mpe": 0.0,
-        }
-
-    absolute_errors: np.ndarray = np.abs(expected_array - observed_array)
-    mae: float = float(np.mean(absolute_errors)) if len(absolute_errors) > 0 else 0.0
-
-    non_zero_mask: np.ndarray = observed_array != 0
-    filtered_absolute_errors: np.ndarray = absolute_errors[non_zero_mask]
-    filtered_observed_array: np.ndarray = observed_array[non_zero_mask]
-
-    mpe: float = 0.0
-    if len(filtered_absolute_errors) > 0 and len(filtered_observed_array) > 0:
-        percentage_errors: np.ndarray = (
-            filtered_absolute_errors / filtered_observed_array * 100
-        )
-        mpe = float(np.mean(percentage_errors))
-
-    return {
-        "expected": expected_values,
-        "observed": observed_values,
-        "mae": mae,
-        "mpe": mpe,
-    }
-
-
-def calc_mae_mpe(
-    prob_dist_dict_all: Dict[Any, Dict[Any, Dict[str, Any]]],
-    use_most_probable: bool = False,
-) -> Dict[Any, Dict[str, Union[List[Union[int, float]], float]]]:
-    """Calculate MAE and MPE for all prediction times in the given probability distribution dictionary.
-
-    Parameters
-    ----------
-    prob_dist_dict_all : Dict[Any, Dict[Any, Dict[str, Any]]]
-        Nested dictionary containing probability distributions.
[…]
