+++ evaluate_script_and_notebook_plan.md
@@ -1,139 +1,0 @@
-# evaluate.py and notebook 4d plan (current)
-
-## Purpose
-
-`evaluate.py` is an orchestration layer that writes evaluation outputs (plots + scalar metadata)
-for configured targets. Notebook `4d` demonstrates the full evaluation workflow: building
-service-level probability distributions, preparing canonical inputs, calling `run_evaluation(...)`,
-and reviewing the outputs.
-
-## Agreed design decisions
-
-- **Single canonical non-classifier input shape:** `flow_inputs_by_service` only.
-- **No auto-building in `run_evaluation(...)`:** callers pass explicit prepared inputs.
-- **Classifier scalar metrics are read-only in evaluation:** read from model artifacts, do not
-  recompute from raw visits inside `evaluate.py`.
-- **One global scalar file per run:** `scalars.json` at run root, including skipped entries.
-- **Aspirational targets:** record metadata and skip invalid observed-vs-predicted diagnostics.
-- **Prediction time convention:** tuples `(hour, minute)` throughout.
-- **Multiple runs:** `run_evaluation(...)` writes to `output_root/<run_label_or_timestamp>/`.
-
-## Current `run_evaluation(...)` contract
-
-```python
-run_evaluation(
-    output_root,
-    prediction_times,
-    config_path=None,
-    run_label=None,
-    evaluation_targets=None,
-    classifier_inputs=None,
-    flow_inputs_by_service=None,
-    services=None,
-)
-```
-
-### Inputs
-
-- `prediction_times`: required list of `(hour, minute)` tuples.
-- `classifier_inputs`: `{target_name: {"trained_models": ..., "visits_df": ..., "label_col": ...}}`
-  (`visits_df` is used for classifier plots; classifier scalars come from artifacts).
-- `flow_inputs_by_service`: canonical map
-  `service -> flow_name -> prediction_time -> payload`.
-- `services`: optional explicit service list; otherwise inferred from `flow_inputs_by_service`.
-
-### Explicitly out of scope in `run_evaluation(...)`
-
-- Building flow inputs from model outputs internally.
-- Alternate non-canonical input surfaces (for example `flow_inputs_by_target`).
-
-## Evaluation target contract
-
-`EvaluationTarget` defines what/how to evaluate:
-
-- `name`
-- `flow_type` (`classifier`, `pmf`, `poisson`, `special`)
-- `evaluation_mode` (`classifier`, `distribution`, `arrival_deltas`, `survival_curve`, `aspirational_skip`)
-- `aspirational` (boolean)
-- optional `components`
-- optional `flow_selection`
-
-Graceful failure rule: missing input/model for a target/component/time records
-`evaluated: false` with a clear `reason` in `scalars.json`; run continues.
-
-## Classifier metrics source (read-only)
-
-For each classifier model/timepoint, evaluation reads metrics in this order:
-
-1. `TrainedClassifier.selected_eval_metrics` (canonical)
-2. `TrainingResults.selected_eval_metrics` (legacy compatibility)
-3. `TrainingResults.test_results` (legacy test metrics)
-4. best CV validation metrics from `TrainingResults.training_info["cv_trials"]` (legacy)
-
-If required metrics are still unavailable, mark that timepoint as skipped in `scalars.json`.
-
-## Model artifact notes
-
-- `selected_eval_metrics` now sits on **`TrainedClassifier`** (top-level), not nested in
-  `TrainingResults`.
-- Dataset metadata stores explicit positive-case counts:
-  `train_valid_test_positive_cases`.
-
-## Output structure
-
-```text
-eval-output/
-└── <run_label_or_timestamp>/
-    ├── config.yaml                  # copied when provided
-    ├── scalars.json                 # one global scalars file
-    ├── classifiers/
[…]

[…]
