### 2026-03-17 3803525
Add observation_mode to EvaluationTarget and extract observation counting
- Add observation_mode field to EvaluationTarget with validation against
  OBSERVATION_MODES; default preserves backward compatibility
- Assign correct observation_mode to each target in the default registry,
  with comments mapping targets to evaluation plan rows
- Create evaluate/observations.py with count_observed_at_some_point,
  count_observed_in_window, and a count_observed dispatcher
- Update get_prob_dist_by_service to accept observation_mode and delegate
  to the new dispatcher instead of the hardcoded _count_ob
[…]
