+++ src/patientflow/evaluate/targets.py
@@ -5,6 +5,10 @@ This module contains utilities that:
 - resolve a single component per target
 - convert target-like registry entries into typed ``EvaluationTarget`` objects
 - provide the default typed target registry used by builders/runners
+
+Row references in :func:`get_default_evaluation_targets` (e.g. "Row 3",
+"Row 5b") refer to the flow evaluation matrix in
+``docs/evaluation_plan.md``.
 """
 
 from typing import Any, Dict, Iterable, List, Mapping, Optional
@@ -81,13 +85,17 @@ def convert_legacy_target(name: str, legacy_target: Any) -> EvaluationTarget:
         flow_selection=getattr(legacy_target, "flow_selection", None),
         target_name=name,
     )
-    return EvaluationTarget(
+    kwargs: dict = dict(
         name=getattr(legacy_target, "name", name),
         flow_type=getattr(legacy_target, "flow_type"),
         evaluation_mode=getattr(legacy_target, "evaluation_mode"),
         component=component,
         flow_selection=getattr(legacy_target, "flow_selection", None),
     )
+    obs_mode = getattr(legacy_target, "observation_mode", None)
+    if obs_mode is not None:
+        kwargs["observation_mode"] = obs_mode
+    return EvaluationTarget(**kwargs)
 
 
 def convert_legacy_targets(
@@ -121,23 +129,29 @@ def get_default_evaluation_targets() -> Dict[str, EvaluationTarget]:
         supported flows and diagnostic modes.
     """
     return {
+        # Row 1: P(admission after ED) — patient-level binary classifier
         "ed_current_admission_classifier": EvaluationTarget(
             name="ed_current_admission_classifier",
             flow_type="classifier",
             evaluation_mode="classifier",
             component="arrivals",
+            observation_mode="classifier_binary",
         ),
+        # Row 10: P(discharge within window) — patient-level binary classifier
         "discharge_classifier": EvaluationTarget(
             name="discharge_classifier",
             flow_type="classifier",
             evaluation_mode="classifier",
             component="departures",
+            observation_mode="classifier_binary",
         ),
+        # Row 3: admitted at some point (no window constraint)
         "ed_current_beds": EvaluationTarget(
             name="ed_current_beds",
             flow_type="pmf",
             evaluation_mode="distribution",
             component="arrivals",
+            observation_mode="count_at_some_point",
             flow_selection=FlowSelection.custom(
                 include_ed_current=True,
                 include_ed_yta=False,
@@ -147,41 +161,53 @@ def get_default_evaluation_targets() -> Dict[str, EvaluationTarget]:
                 include_departures=False,
             ),
         ),
+        # Row 4b: survival curve comparison (train vs test)
         "ed_current_window_prob": EvaluationTarget(
             name="ed_current_window_prob",
             flow_type="special",
             evaluation_mode="survival_curve",
             component="arrivals",
+            observation_mode="survival_comparison",
         ),
+        # Row 4a: aspirational — no observation
         "ed_current_window_prob_aspirational": EvaluationTarget(
             name="ed_current_window_prob_aspirational",
             flow_type="special",
             evaluation_mode="aspirational_skip",
             component="arrivals",
+            observation_mode="not_applicable",
         ),
+        # Row 5b: beds needed in window (survival-curve-based)
         "ed_current_window_beds": EvaluationTarget(
             name="ed_current_window_beds",
             flow_type="pmf",
             evaluation_mode="distribution",
             component="arrivals",
+            observation_mode="count_in_window",
         ),
+        # Row 5a: aspirational — no observation
         "ed_current_window_beds_aspirational": EvaluationTarget(
             name="ed_current_window_beds_aspirational",
             flow_type="pmf",
             evaluation_mode="aspirational_skip",
             component="arrivals",
+            observation_mode="not_applicable",
         ),
[…]
