+++ src/patientflow/evaluate/runner.py
@@ -35,6 +35,7 @@ def run_evaluation(
     config_path: Optional[Union[str, Path]] = None,
     run_label: Optional[str] = None,
     services: Optional[List[str]] = None,
+    skip_inactive_services: bool = False,
     verbose: bool = True,
 ) -> Dict[str, Any]:
     """Run typed evaluation targets and write plots/scalars to disk.
@@ -52,6 +53,12 @@ def run_evaluation(
     services
         Optional service subset for non-classifier evaluations. If omitted,
         services are inferred from ``inputs.flow_inputs_by_service``.
+    skip_inactive_services
+        When True, distribution targets whose snapshots show negligible
+        activity (zero observed counts and near-zero predicted means) are
+        recorded in scalars with ``charts_generated: false`` but no chart
+        files are written.  This dramatically reduces output volume when
+        many services have zero activity.  Defaults to False.
     verbose
         Whether to print progress messages. Defaults to True.
 
@@ -133,6 +140,7 @@ def run_evaluation(
                     collector=collector,
                     output_root=root,
                     snapshots_by_time=payload_by_time,
+                    skip_inactive=skip_inactive_services,
                 )
             elif mode == "arrival_deltas":
                 evaluate_arrival_deltas(
@@ -177,11 +185,25 @@ def run_evaluation(
 
     scalars_path = root / "scalars.json"
     meta = default_scalars_meta(RELIABILITY_THRESHOLDS, schema_version=3)
+    include_summary = skip_inactive_services
     scalars_path.write_text(
-        json.dumps(collector.to_payload(meta=meta), indent=2, sort_keys=True)
+        json.dumps(
+            collector.to_payload(
+                meta=meta, include_service_summary=include_summary
+            ),
+            indent=2,
+            sort_keys=True,
+        )
     )
 
     if verbose:
+        if include_summary:
+            summary = collector.service_summary()
+            print(
+                f"  Services: {summary['active_services']} active, "
+                f"{summary['inactive_services']} inactive "
+                f"(of {summary['total_services']} total)"
+            )
         print(f"Evaluation complete. Results written to {root}")
 
     return {
