### 2026-03-18 25859ff
Fix classifier crash when categorical columns use StringDtype or CategoricalDtype instead of object dtype
