### 2026-03-18 d540cc5
Handle empty arrival-delta snapshots as inactive services - avoid UnboundLocalError when no arrivals exist in snapshot windows - skip arrival-delta chart generation only when all supplied snapshots are empty - record charts_generated: false / skip_reason: "inactive_service" in scalars and add regression tests
