### 2026-03-18 f46c3a3
Add incremental scalars, lazy directories, and classifier explainability plots
- scalars.json now accumulates across repeated run_evaluation calls with
  the same run_label; prior rows are preserved unless superseded by a new
  row with the same (flow, service, component, prediction_time) key
- skip_inactive_services defaults to True
- service and classifier directories are created only when charts will
  actually be written, avoiding empty folders in the output
- service_summary logic inverted: a service is active if any row
  generated charts, inactive only when all rows say otherwise
- eval
[…]
