### 2026-03-19 c335888
WIP count observed values
