### 2026-03-24 88b56a4
Move warn_specialty_mismatch from emergency_demand to service
