+++ docs/evaluation_plan.md
@@ -1,48 +1,76 @@
-# Evaluation plan for bed demand tool
-
-Grey-marked rows represent aspirational predictions — flows that assume operational targets (e.g. 4-hour ED targets) not matched by observed reality. Standard observed-vs-predicted evaluation is not valid for these. Combined predictions that include any aspirational component are themselves aspirational.
+# Evaluation plan v3
 
 ## Flow evaluation matrix
 
-| Row | Flow | Level | What is being predicted? | How modelled | Nature of predictions | Nature of observations | How to evaluate |
-|-----|------|-------|--------------------------|--------------|----------------------|----------------------|-----------------|
-| 1 | Admissions via ED/SDEC | Patient | P(admission after ED) | ML Classifier (XGBoost) applied to real-time data on patients in the ED without bed requests | Probability | Whether admitted or not (binary) | AUC and AUPRC, Discrimination plots, Calibration plots, MADCAP plots |
-| 2 | Admissions via ED/SDEC | Patient | P(admission to subspecialty if admitted) | Probabilistic mapping of consult sequence to subspecialty of admission | Vector of probabilities of admission to each subspecialty | Subspecialty of admission | Evaluated indirectly via row 3 |
-| 3 | Admissions via ED/SDEC | Service | Number of admissions of patients currently in the ED | Aggregation of predictions based on product of 1 and 2 | Probability distribution | Number admitted (at some point) | MAE and MPE, Obs-exp histogram, EPUDD |
-| **4a** | **Admissions via ED/SDEC** | **Patient** | **Probability of admission in prediction window (aspirational)** | **Aspirational curve** | **Probability** | **N/A** | ***Aspirational — no evaluation*** |
-| 4b | Admissions via ED/SDEC | Patient | Probability of admission in prediction window (based on prior data) | Survival curve | Probability | Whether admitted in window | Compare survival curves of training and test sets |
-| **5a** | **Admissions via ED/SDEC** | **Service** | **Number of beds needed in prediction window, if 4-hour targets are met, for patients currently in ED** | **Aggregation of predictions based on product of 1, 2 and 4a** | **Probability distribution** | **N/A** | ***Aspirational — no evaluation*** |
-| 5b | Admissions via ED/SDEC | Service | Number of beds needed in prediction window, assuming past ED performance | Aggregation of predictions based on product of 1, 2 and 4b | Probability distribution | Number admitted in window | MAE and MPE, Obs-exp histogram, EPUDD |
-| 6 | Admissions via ED/SDEC of patients yet-to-arrive | Service | Arrivals in prediction window of patients later admitted | Rate of arrivals via ED/SDEC front door of patients later admitted | Arrival rate in each 15-min segment of the prediction window | Observed arrival rates | Histogram of difference between cumulative arrival rates and observed numbers |
-| **7a** | **Admissions via ED/SDEC of patients yet-to-arrive** | **Service** | **Arrivals in prediction window of patients later admitted, if 4-hour targets are met** | **Time-varying Poisson using arrival rates from 6, weighted by probability of admission in window from 4a** | **Probability distribution** | **N/A** | ***Aspirational — no evaluation*** |
-| 7b | Admissions via ED/SDEC of patients yet-to-arrive | Service | Number of beds needed in prediction window, assuming past ED performance | Time-varying Poisson using arrival rates from 6, weighted by probability of admission in window from 4b | Probability distribution | Number admitted in prediction window | MAE and MPE, Obs-exp histogram, EPUDD |
[…]
