### 2026-04-07 452b3d1
use FlowSelection to determine validation in get_prob_dist_by_service - Validate snapshots, curve params, and models from FlowSelection (flow_selection_checks) - Require flow_selection on build_service_data, predict_service, and hierarchy helpers - get_prob_dist_by_service: require flow_selection; optional ED / x1–y2 when unused; fix empty-ED date skip - Rename flow_requirements module to flow_selection_checks - Add targeted tests for inpatient and parametric-curve validation plus component checks - Update notebooks 4c/4d/4e and docs/notebooks/4d_Evaluate_demand_predictions.md
d#
