+++ tests/test_service.py
@@ -10,6 +10,7 @@ from patientflow.predict.service import (
     FlowInputs,
     compute_transfer_arrivals,
 )
+from patientflow.predict.types import FlowSelection
 from patientflow.predictors.transfer_predictor import TransferProbabilityEstimator
 from patientflow.model_artifacts import TrainedClassifier, TrainingResults
 from patientflow.load import get_model_key
@@ -331,10 +332,11 @@ class TestBuildServiceData(unittest.TestCase):
         result = build_service_data(
             models=self.models,
             prediction_time=self.prediction_time,
-            ed_snapshots=ed_snapshots,
-            inpatient_snapshots=inpatient_snapshots,
+            flow_selection=FlowSelection.default(),
             specialties=self.specialties,
             prediction_window=self.prediction_window,
+            ed_snapshots=ed_snapshots,
+            inpatient_snapshots=inpatient_snapshots,
             x1=self.x1,
             y1=self.y1,
             x2=self.x2,
@@ -403,10 +405,11 @@ class TestBuildServiceData(unittest.TestCase):
         result = build_service_data(
             models=models,
             prediction_time=self.prediction_time,
-            ed_snapshots=ed_snapshots,
-            inpatient_snapshots=inpatient_snapshots,
+            flow_selection=FlowSelection.default(),
             specialties=self.specialties,
             prediction_window=self.prediction_window,
+            ed_snapshots=ed_snapshots,
+            inpatient_snapshots=inpatient_snapshots,
             x1=self.x1,
             y1=self.y1,
             x2=self.x2,
@@ -425,10 +428,11 @@ class TestBuildServiceData(unittest.TestCase):
             build_service_data(
                 models=self.models,
                 prediction_time=(8, 0),
-                ed_snapshots=ed_snapshots,
-                inpatient_snapshots=inpatient_snapshots,
+                flow_selection=FlowSelection.default(),
                 specialties=self.specialties,
                 prediction_window=self.prediction_window,
+                ed_snapshots=ed_snapshots,
+                inpatient_snapshots=inpatient_snapshots,
                 x1=self.x1,
                 y1=self.y1,
                 x2=self.x2,
@@ -453,10 +457,11 @@ class TestBuildServiceData(unittest.TestCase):
             build_service_data(
                 models=models,
                 prediction_time=self.prediction_time,
-                ed_snapshots=ed_snapshots,
-                inpatient_snapshots=inpatient_snapshots,
+                flow_selection=FlowSelection.default(),
                 specialties=self.specialties,
                 prediction_window=self.prediction_window,
+                ed_snapshots=ed_snapshots,
+                inpatient_snapshots=inpatient_snapshots,
                 x1=self.x1,
                 y1=self.y1,
                 x2=self.x2,
@@ -471,10 +476,11 @@ class TestBuildServiceData(unittest.TestCase):
             build_service_data(
                 models=self.models,
                 prediction_time=self.prediction_time,
-                ed_snapshots=ed_snapshots.drop(columns=["elapsed_los"]),
-                inpatient_snapshots=inpatient_snapshots,
+                flow_selection=FlowSelection.default(),
                 specialties=self.specialties,
                 prediction_window=self.prediction_window,
+                ed_snapshots=ed_snapshots.drop(columns=["elapsed_los"]),
+                inpatient_snapshots=inpatient_snapshots,
                 x1=self.x1,
                 y1=self.y1,
                 x2=self.x2,
@@ -485,10 +491,11 @@ class TestBuildServiceData(unittest.TestCase):
             build_service_data(
                 models=self.models,
                 prediction_time=self.prediction_time,
-                ed_snapshots=ed_snapshots,
-                inpatient_snapshots=inpatient_snapshots.drop(columns=["elapsed_los"]),
[…]
