+++ src/patientflow/predict/service.py
@@ -51,6 +51,19 @@ from patientflow.calculate.admission_in_prediction_window import (
     calculate_admission_probability_from_survival_curve,
 )
 from patientflow.model_artifacts import TrainedClassifier
+from patientflow.predict.flow_selection_checks import (
+    requires_admission_curve_params,
+    requires_ed_snapshots,
+    requires_inpatient_snapshots,
+    validate_ed_classifier,
+    validate_elective_yta,
+    validate_inpatient_classifier,
+    validate_non_ed_yta,
+    validate_spec_model_for_ed,
+    validate_transfer_model,
+    validate_yta_ed,
+)
+from patientflow.predict.types import FlowSelection
 
 
 @dataclass(frozen=True)
@@ -318,17 +331,16 @@ def _validate_models_and_data(
     inpatient_snapshots: Optional[pd.DataFrame],
     prediction_window,
     specialties: List[str],
+    flow_selection: FlowSelection,
+    use_admission_in_window_prob: bool,
+    x1: Optional[float],
+    y1: Optional[float],
+    x2: Optional[float],
+    y2: Optional[float],
 ) -> None:
-    """Validate all models and input data.
+    """Validate models and snapshots required by *flow_selection*."""
+    flow_selection.validate()
 
-    Raises
-    ------
-    TypeError
-        If any model is not of the expected type
-    ValueError
-        If required columns are missing, models are not fitted, or parameters
-        don't match between models and requested parameters
-    """
     (
         ed_classifier,
         inpatient_classifier,
@@ -339,7 +351,6 @@ def _validate_models_and_data(
         transfer_model,
     ) = models
 
-    # Validate model types
     if ed_classifier is not None and not isinstance(ed_classifier, TrainedClassifier):
         raise TypeError("First model must be of type TrainedClassifier (ED classifier)")
     if inpatient_classifier is not None and not isinstance(
@@ -371,7 +382,6 @@ def _validate_models_and_data(
                 "If you're using Jupyter, try restarting the kernel."
             )
 
-    # Validate that non-ED and elective models are DirectAdmissionPredictor
     if non_ed_yta_model is not None and not isinstance(
         non_ed_yta_model, DirectAdmissionPredictor
     ):
@@ -391,8 +401,19 @@ def _validate_models_and_data(
             "Seventh model must be of type TransferProbabilityEstimator (transfer)"
         )
 
-    # Validate elapsed_los column presence and dtype for ED snapshots
-    if ed_snapshots is not None:
+    if requires_ed_snapshots(flow_selection) and ed_snapshots is None:
+        raise ValueError(
+            "ed_snapshots is required when flow_selection.include_ed_current is True"
+        )
+    if requires_inpatient_snapshots(flow_selection, transfer_model) and (
+        inpatient_snapshots is None
+    ):
+        raise ValueError(
+            "inpatient_snapshots is required when include_departures is True, or when "
+            "include_transfers_in is True and a transfer model is provided"
+        )
+
+    if ed_snapshots is not None and len(ed_snapshots) > 0:
         if "elapsed_los" not in ed_snapshots.columns:
             raise ValueError("Column 'elapsed_los' not found in ed_snapshots")
         if not pd.api.types.is_timedelta64_dtype(ed_snapshots["elapsed_los"]):
@@ -402,8 +423,7 @@ def _validate_models_and_data(
                 f"{actual_type}"
             )
 
-    # Validate elapsed_los column presence and dtype for inpatient snapshots
-    if inpatient_snapshots is not None:
+    if inpatient_snapshots is not None and len(inpatient_snapshots) > 0:
         if "elapsed_los" not in inpatient_snapshots.columns:
[…]
