### 2026-04-08 8136287
route timedelta features via FeatureKind, not one-hot
- Add infer_feature_kind / FeatureKind; map kinds to sklearn steps (timedelta → seconds + StandardScaler).
- Pass ordinal_mappings into FeatureColumnTransformer for consistent defaults.
