+++ src/patientflow/train/classifiers.py
@@ -16,6 +16,8 @@ initialise_model
     Initialize a model with given hyperparameters
 create_column_transformer
     Create a column transformer for a dataframe with dynamic column handling
+infer_feature_kind
+    Classify each column for preprocessing (ordinal, numeric, timedelta, etc.)
 calculate_class_balance
     Calculate class balance ratios for target labels
 get_feature_metadata
@@ -32,6 +34,7 @@ train_multiple_classifiers
     Train admission prediction models for multiple prediction times
 """
 
+from enum import Enum, auto
 from typing import Dict, List, Any, Tuple, Optional, Union, TypedDict, Type
 import numpy as np
 import numpy.typing as npt
@@ -43,7 +46,12 @@ import pandas as pd
 from sklearn.compose import ColumnTransformer
 from sklearn.metrics import roc_auc_score, log_loss, average_precision_score
 from sklearn.model_selection import TimeSeriesSplit, ParameterGrid
-from sklearn.preprocessing import OneHotEncoder, OrdinalEncoder, StandardScaler
+from sklearn.preprocessing import (
+    FunctionTransformer,
+    OneHotEncoder,
+    OrdinalEncoder,
+    StandardScaler,
+)
 from sklearn.pipeline import Pipeline
 from sklearn.calibration import CalibratedClassifierCV
 from sklearn import __version__ as sk_version
@@ -60,21 +68,111 @@ from patientflow.model_artifacts import (
 )
 
 
-def _is_string_like_column(series: Series) -> bool:
-    """Check whether a Series holds string-like values that need encoding.
+class FeatureKind(Enum):
+    """How a column is preprocessed in ``create_column_transformer``."""
 
-    Uses an exclusion-based approach so that any current or future
-    string-like dtype (``object``, ``StringDtype``, ``ArrowDtype("string")``,
-    ``CategoricalDtype`` with string categories, etc.) is detected without
-    needing to enumerate each one explicitly.
+    ORDINAL = auto()
+    BOOLEAN = auto()
+    TIMEDELTA = auto()
+    DATETIME = auto()
+    NUMERIC_BINARY = auto()
+    NUMERIC_SCALED = auto()
+    CATEGORICAL = auto()
+
+
+def _timedelta_to_float_seconds(X: npt.NDArray[Any]) -> npt.NDArray[np.float64]:
+    arr = np.asarray(X)
+    if not np.issubdtype(arr.dtype, np.timedelta64):
+        return arr.astype(np.float64, copy=False)
+    flat = arr.ravel()
+    sec = pd.Series(flat).dt.total_seconds().to_numpy(dtype=np.float64)
+    return sec.reshape(arr.shape[0], -1)
+
+
+def infer_feature_kind(
+    series: Series,
+    col: str,
+    ordinal_mappings: Dict[str, List[Any]],
+) -> FeatureKind:
+    """Assign a single preprocessing role to a column.
+
+    This is the single source of truth for both ``create_column_transformer``
+    and ``FeatureColumnTransformer`` defaults. Checks are ordered from explicit
+    contracts (ordinal mapping, concrete dtypes) to broad fallbacks.
     """
+    if col in ordinal_mappings:
+        return FeatureKind.ORDINAL
+    if series.dtype == "bool":
+        return FeatureKind.BOOLEAN
+    if pd.api.types.is_timedelta64_dtype(series):
+        return FeatureKind.TIMEDELTA
+    if pd.api.types.is_datetime64_any_dtype(series):
+        return FeatureKind.DATETIME
+
     if isinstance(series.dtype, pd.CategoricalDtype):
-        series = series.cat.categories.to_series()
-    return not (
-        pd.api.types.is_numeric_dtype(series)
[…]
